% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{add_comma}
\alias{add_comma}
\title{Add thousands commas to numbers}
\usage{
add_comma(x, big_mark = ",", include_decimal = FALSE)
}
\arguments{
\item{x}{Numeric or character vector}

\item{big_mark}{Mark for thousands. Default is \code{","}}

\item{include_decimal}{If \code{FALSE} (default), number is rounded and decimal removed}
}
\description{
Add thousands commas to numbers
}
