% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{add_id_var}
\alias{add_id_var}
\title{Add column for row number}
\usage{
add_id_var(df, new_colname = "id")
}
\arguments{
\item{df}{Data frame or grouped_df}

\item{new_colname}{Column name for id variable. Default is \code{"id"}}
}
\value{
Data frame with new column for id variable. If \code{df} is a grouped data frame, rows will be numbered separately for each group
}
\description{
Add column for row number
}
