% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{add_km_labels}
\alias{add_km_labels}
\title{Label for Kaplan-Meier plots}
\usage{
add_km_labels(x, labels = NULL, font_size = NULL)
}
\arguments{
\item{x}{ggplot2 object (Kaplan-Meier plot)}

\item{labels}{Character vector containing labels for curves}

\item{font_size}{Size of text used to generated labels for KM curves. Unit in pts. Default uses \code{base_size}}
}
\description{
Label for Kaplan-Meier plots
}
