% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{add_line_error}
\alias{add_line_error}
\title{Add error region to line plot or Kaplan-Meier curve}
\usage{
add_line_error(
  x,
  mapping = aes(ymin = ymin, ymax = ymax),
  colors = "grey70",
  alpha = 0.4,
  ...
)
}
\arguments{
\item{x}{ggplot object}

\item{mapping}{Enter using \code{aes()}. \code{ymin} and \code{ymax} must be entered. Default is \code{ymin = ymin} and \code{ymax = ymax}}

\item{colors}{Color for error region. Default is \code{"grey70"}. If > length is 1, \code{x$data} must contain a column names \code{group}}

\item{alpha}{Transparency setting. Enter as numeric 0-1. Default is \code{0.4}}

\item{...}{Arguments passed to \code{geom_ribbon}}
}
\value{
Add to plot using \code{plot + add_line_error()}
}
\description{
Add error region to line plot or Kaplan-Meier curve
}
