% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sig_bar.R
\name{add_p_vals}
\alias{add_p_vals}
\title{Add significance annotation to plot}
\usage{
add_p_vals(
  plot,
  bar_nudge = 0.08,
  star_nudge = 0.02,
  text_nudge = 0.05,
  step_increase = 0.11,
  size = 0.65,
  color = "black",
  colour = color,
  bar_thickness = size,
  vjust = 0.5,
  hjust = 0.5,
  ...
)
}
\arguments{
\item{plot}{ggplot object}

\item{bar_nudge}{Nudge factor along y axis for significance bar as proportion of viewport. Default is \code{0.08}}

\item{star_nudge}{udge factor along y axis for stars as proportion of viewport. Default is \code{0.02}}

\item{text_nudge}{Nudge factor along y axis for text as proportion of viewport. Default is \code{0.05}}

\item{step_increase}{Step increase as proportion of viewport. Default is \code{0.11}}

\item{size}{Alias for \code{bar_thickness}}

\item{color}{Color of significance bar and annotation. Default is \code{"black"}}

\item{colour}{Alias for \code{color}}

\item{bar_thickness}{Thickness of significance bars. Default is \code{0.65}}

\item{vjust}{Vertical alignment. Default is \code{0.5}}

\item{hjust}{Horizontal alignment. Default is \code{0.5}}

\item{...}{Arguments passed to \code{.create_df_sig}}
}
\description{
Add significance annotation to plot
}
