% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi.R
\name{align_plot_dim}
\alias{align_plot_dim}
\title{Apply uniform plot dimensions to a list of plots}
\usage{
align_plot_dim(plot_list)
}
\arguments{
\item{plot_list}{List of plots}
}
\value{
List of plots with identical dimensions which can be passed to plot_grid function
}
\description{
Apply uniform plot dimensions to a list of plots
}
