% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{all_identical}
\alias{all_identical}
\alias{all_equal}
\title{Recursively determine whether all inputs are identical}
\usage{
all_identical(...)

all_equal(...)
}
\arguments{
\item{...}{Either a comma separated list of vectors not wrapped in \code{c()} or a list containing vectors}
}
\description{
Recursively determine whether all inputs are identical

Recursively determine whether all inputs are equal
}
