% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{all_named}
\alias{all_named}
\title{Determine whether all elements are named}
\usage{
all_named(x)
}
\arguments{
\item{x}{List or vector}
}
\value{
Logical of length 1
}
\description{
Determine whether all elements are named
}
