% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{any_dupes}
\alias{any_dupes}
\alias{all_unique}
\title{Determine whether vector contains any duplicate values}
\usage{
any_dupes(x)

all_unique(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Length 1 logical
}
\description{
Determine whether vector contains any duplicate values

Determine whether all values in a vector are unique
}
