% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{any_named}
\alias{any_named}
\alias{has_names}
\title{Determine whether any elements are named}
\usage{
any_named(x)

has_names(x)
}
\arguments{
\item{x}{List or vector}
}
\value{
Logical of length 1
}
\description{
Determine whether any elements are named

Alias for \code{any_named}
}
