% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{as_binary_01}
\alias{as_binary_01}
\title{Convert binary variables to integer coded as 0 or 1}
\usage{
as_binary_01(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Integer vector coded as 0 or 1 with length equal to input
}
\description{
Convert binary variables to integer coded as 0 or 1
}
