% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{as_date}
\alias{as_date}
\title{Transform input to YMD format}
\usage{
as_date(x)
}
\arguments{
\item{x}{Vector containing year, month, day}
}
\value{
Date in YYYY-MM-DD format
}
\description{
Transform input to YMD format
}
