% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{as_formula}
\alias{as_formula}
\title{Coerce string to formula}
\usage{
as_formula(x, env = parent.frame())
}
\arguments{
\item{x}{Formula as a string}

\item{env}{Environment of formula. Default is \code{parent.frame()}}
}
\value{
Formula
}
\description{
Coerce string to formula
}
