% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{as_numeric_factor}
\alias{as_numeric_factor}
\alias{fct_to_num}
\title{Convert categorical variable to factor encoded as integer}
\usage{
as_numeric_factor(x, levels = NULL, reverse = FALSE)

fct_to_num(x, levels = NULL, reverse = FALSE)
}
\arguments{
\item{x}{Factor, character, numeric, or logical vector}

\item{levels}{Order of levels in \code{x}}

\item{reverse}{If \code{TRUE} order of levels is reversed. \code{FALSE} by default}
}
\value{
Integer version of \code{x} (lowest level is 1, highest value is equal to the number of levels)
}
\description{
Similar to as.integer(factor(x)) but faster
}
