% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area under the curve}
\usage{
auc(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  interval = NULL,
  method = c("trapezoidal", "step", "spline", "all")
)
}
\arguments{
\item{df}{Data frame containing dependent variable (\code{y}) and independent variable (\code{x})}

\item{formula}{Formula as y ~ x}

\item{y, x}{Variables for y and x axis respectively. Enter as quoted or unquoted variable name}

\item{interval}{Minimum and maximum x values. Range is inclusive (i.e. [])}

\item{method}{Method used to calculate area under the curve. Options: \code{"trapezoidal"} (default), \code{"step"}, \code{"spline"}, \code{"all"}}
}
\value{
If method is \code{"trapezoidal"}, \code{"step"}, or \code{"spline"}, result is a length 1 numeric vector. If method is \code{"all"}, result is a data frame with 1 column for method and 1 column for auc
}
\description{
Area under the curve
}
