% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc_ci}
\alias{auc_ci}
\alias{auc_ci.default}
\alias{auc_ci.data.frame}
\alias{auc_ci.lm}
\alias{auc_ci.glm}
\alias{auc_ci.coxph}
\title{AUC and confidence interval}
\usage{
auc_ci(
  x,
  continuous_var = NULL,
  outcome_var = NULL,
  wilcox = TRUE,
  ci = 0.95,
  ...
)

\method{auc_ci}{default}(
  x,
  continuous_var = NULL,
  outcome_var = NULL,
  wilcox = TRUE,
  ci = 0.95,
  ...
)

\method{auc_ci}{data.frame}(
  x,
  continuous_var = NULL,
  outcome_var = NULL,
  wilcox = TRUE,
  ci = 0.95,
  ...
)

\method{auc_ci}{lm}(
  x,
  continuous_var = NULL,
  outcome_var = NULL,
  wilcox = TRUE,
  ci = 0.95,
  ...
)

\method{auc_ci}{glm}(
  x,
  continuous_var = NULL,
  outcome_var = NULL,
  wilcox = TRUE,
  ci = 0.95,
  ...
)

\method{auc_ci}{coxph}(
  x,
  continuous_var = NULL,
  outcome_var = NULL,
  wilcox = TRUE,
  ci = 0.95,
  ...
)
}
\arguments{
\item{x}{Data frame or regression output}

\item{continuous_var}{Continuous variable (predictor). Enter as quoted or unquoted variable name. Only required if \code{x} is a data frame}

\item{outcome_var}{Binary outcome variable. Enter as quoted or unquoted variable name. Only required if \code{x} is a data frame}

\item{wilcox}{If \code{TRUE} (default), \code{wilcox.test} is used to calculate confidence interval. If \code{FALSE}, method created by Hanley and McNeil is used to calculate confidence interval}

\item{ci}{Confidence interval. Default is \code{0.95}}

\item{...}{Arguments passed to class-specific \code{auc_ci} function}
}
\value{
List containing columns for auc, auc_lower, auc_upper. If \code{x} is a data frame,  "positive" and "negative" outcomes are also included
}
\description{
AUC and confidence interval

auc_ci - default method

auc_ci - data frame

auc_ci - lm

auc_ci - glm

auc_ci - coxph
}
