% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{axis_breaks_continuous}
\alias{axis_breaks_continuous}
\title{Create axis breaks for continuous variable}
\usage{
axis_breaks_continuous(
  n_breaks = 4,
  breaks_fn = pretty,
  min_breaks = 4,
  max_breaks = 6,
  expand = c(0, 0),
  include_negative = FALSE
)
}
\arguments{
\item{n_breaks}{Desired number of axis breaks. Default is \code{4}}

\item{breaks_fn}{Function used to generate breaks. Enter as call using (). Default is \code{pretty.} Alternative is \code{scales::breaks_extended(n = n_breaks, only.loose = TRUE)}}

\item{min_breaks}{Minimum number of possible axis breaks. Default is \code{4}}

\item{max_breaks}{Maximum number of breaks allowed. Default is \code{6}}

\item{expand}{Length 2 numeric vector containing expansion arguments for lower and upper ends of axis respectively. Default is \code{0} for both}

\item{include_negative}{If \code{FALSE} (default), negative numbers are excluded from breaks}
}
\value{
Enter as \code{scale_y_continuous(breaks = axis_breaks_continuous())}
}
\description{
Create axis breaks for continuous variable
}
