% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{axis_label_log}
\alias{axis_label_log}
\title{Create base^x axis labels}
\usage{
axis_label_log(base = 10)
}
\arguments{
\item{base}{Logarithm base. Default is \code{10}}
}
\value{
Function that returns an expression. Enter as \verb{scale_(x|y)_continuous(labels = axis_label_log())}
}
\description{
Create base^x axis labels
}
