% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{axis_label_numeric}
\alias{axis_label_numeric}
\title{Labeling function for continuous axis}
\usage{
axis_label_numeric(
  x,
  max_leading_zeroes = 3,
  max_trailing_zeroes = 4,
  unicode = TRUE
)
}
\arguments{
\item{x}{Numeric vector}

\item{max_leading_zeroes}{Maximum number of leading zeroes when max < 1. Default is \code{3}}

\item{max_trailing_zeroes}{Maximum number of leading zeroes when max > 1. Default is \code{4}}

\item{unicode}{If \code{TRUE} (default), multiplication sign is displayed as unicode. If \code{FALSE}, unicode not used}
}
\value{
Enter as \code{scale_y_continuous(labels = axis_label_numeric)}
}
\description{
Labeling function for continuous axis
}
