% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{axis_label_x10}
\alias{axis_label_x10}
\title{Labeling function to generate scientific "multiplier x 10^exp" labels}
\usage{
axis_label_x10(x, unicode = TRUE)
}
\arguments{
\item{x}{Numeric vector containing position of axis breaks}

\item{unicode}{If \code{TRUE} (default), multiplication sign is displayed as unicode. If \code{FALSE}, unicode not used}
}
\value{
Enter as \code{scale_y_continuous(labels = axis_label_x10)}. Output includes multiplier prior to x sign
}
\description{
Labeling function to generate scientific "multiplier x 10^exp" labels
}
