% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{axis_limits}
\alias{axis_limits}
\title{Create axis limits for continuous variable (including log scale)}
\usage{
axis_limits(breaks_fn = axis_breaks_continuous())
}
\arguments{
\item{breaks_fn}{Function used to generate breaks. Default is \code{axis_breaks_continuous()}}
}
\value{
Enter as \code{scale_y_continuous(limits = axis_limits())}
}
\description{
Create axis limits for continuous variable (including log scale)
}
