% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{calc_surv_time}
\alias{calc_surv_time}
\title{Calculate follow up time for survival data using dates}
\usage{
calc_surv_time(
  df,
  date_t0,
  date_censor,
  date_outcome,
  outcome_var = "death",
  unit = "weeks",
  colname_time = "time"
)
}
\arguments{
\item{df}{Data frames}

\item{date_t0}{Column in \code{df} containing start date. Enter as quoted variable name}

\item{date_censor}{Column in \code{df} containing date subject last known to be well (without outcome). Enter as quoted variable name}

\item{date_outcome}{Column in \code{df} containing date of outcome. Enter as quoted variable name}

\item{outcome_var}{Variable containing outcome variable coded as \code{1} (event) or \code{0} (no event/censored). Enter as quoted variable name. Default is \code{"death"}}

\item{unit}{Unit of time. Options: \code{"weeks"} (default), \code{"days"}, \code{"hours"}, \code{"mins"}, \code{"secs"}}

\item{colname_time}{Name for calculated time variable. Default is \code{"time"}}
}
\value{
Data frame with new column for follow up time
}
\description{
Calculate follow up time for survival data using dates
}
