% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{CI}
\alias{CI}
\title{Confidence interval difference from mean}
\usage{
CI(x, ci = 0.95)
}
\arguments{
\item{x}{Numeric vector}

\item{ci}{Confidence level. Default is \code{0.95}}
}
\value{
Absolute difference between mean and either lower or upper bound of CI
}
\description{
Confidence interval difference from mean
}
