% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{CI_prop}
\alias{CI_prop}
\title{Confidence interval for proportions}
\usage{
CI_prop(
  n,
  total,
  ci = 0.95,
  method = c("wilson", "exact", "wald", "wald_corrected", "agresti_coull", "exact",
    "asymptotic", "all")
)
}
\arguments{
\item{n}{Numerator of proportion}

\item{total}{Denominator of proportion}

\item{ci}{Type of confidence interval. Default is \code{0.95}}

\item{method}{Method for calculating confidence interval. Options include: \code{"wilson"} (default, also used by prism), \code{"wald"}, \code{"wald_corrected"}, \code{"agresti_coull"}, \code{"exact"}, \code{"asymptotic"}, \code{"all"}}
}
\value{
Length 2 numeric vector containing lower and upper bounds of CI. If \code{method = "all"}, output is a matrix with 1 row for each method
}
\description{
Confidence interval for proportions
}
