% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{classify_binary}
\alias{classify_binary}
\title{Generate data required for 2 x 2 table}
\usage{
classify_binary(df, outcome_var, predictor_var, cutpoint = NULL, type = ">=")
}
\arguments{
\item{df}{Data frame}

\item{outcome_var}{Outcome variable. Binary variable with the lower value representing controls and the higher value cases (i.e. outcome occurred). Enter as quoted variable name}

\item{predictor_var}{Predictor variable. Either a binary variable with the lower value representing a negative result and the higher value a positive result or a continuous variable with \code{cutpoint} and type input specifying a positive result. Enter as quoted variable name}

\item{cutpoint}{Cutpoint that defines a positive or negative result for \code{predictor_var}}

\item{type}{If \code{">="} (default), positivity is determined by values greater than or equal to \code{cutpoint.}  If \code{">"}, positivity is determined by values of \code{predictor_var} greater than \code{cutpoint}}
}
\value{
List containing the following counts: total (non-missing), cases, controls, positive, negative, tp, fn, fp, tn
}
\description{
Generate data required for 2 x 2 table
}
