% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_as_greyscale}
\alias{clr_as_greyscale}
\alias{greyscale}
\title{Transform colors to greyscale}
\usage{
clr_as_greyscale(..., method = "luma")

greyscale(..., method = "luma")
}
\arguments{
\item{...}{List, character vector, or comma separated list of colors as quoted color names or hexadecimal codes}

\item{method}{Method for transformation. Options: \code{"luma"} (default), \code{"average"}, \code{"min_decomp"}, \code{"max_decomp"}, \code{"red_channel"}, \code{"green_channel"}, \code{"blue_channel"}}
}
\value{
Hexadecimal codes as character vector of greyscale colors
}
\description{
Functionality from Emil Hvitfeldt's excellent package prismatic
}
