% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_blind_deutanan}
\alias{clr_blind_deutanan}
\alias{clr_blind_protan}
\alias{clr_blind_tritan}
\title{Convert colors to those that would be viewed by an individual with deuteranopia colorblindness}
\usage{
clr_blind_deutanan(...)

clr_blind_protan(...)

clr_blind_tritan(...)
}
\arguments{
\item{...}{Colors. Enter as comma separated list of quoted color names or hexadecimal codes with or without \code{c()}}
}
\value{
Hexadecimal code as character vector containing color as viewed by an individual with deuteranopia colorblindness

Hexadecimal code as character vector containing color as viewed by an individual with protanopia colorblindness

Hexadecimal code as character vector containing color as viewed by an individual with tritanopia colorblindness
}
\description{
Functionality from Emil Hvitfeldt's excellent package prismatic

Functionality from Emil Hvitfeldt's excellent package prismatic

Functionality from Emil Hvitfeldt's excellent package prismatic
}
\examples{

# clr_blind_deutanan(c("#0072B5", "#BC3C29"))

}
