% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_continuous}
\alias{clr_continuous}
\title{Generate continuous color scale from 2-3 colors}
\usage{
clr_continuous(..., n = 20)
}
\arguments{
\item{...}{Colors. If 3 colors are entered, output is a continuous scale using 2nd color as midpoint}

\item{n}{Number of colors in output color scheme. Default is \code{20}}
}
\value{
Hexadecimal codes for continuous color scale
}
\description{
Functionality from gplots package
}
\examples{

# clr_continuous(c("#2166AC", "#F7F7F7", "#B2182B"), n = 20)
# clr_continuous("blue", "forestgreen", "gold", "orange", "red", n = 20)
# clr_continuous(c("#2166AC", "#B2182B"), n = 20)

}
