% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_continuous_grey}
\alias{clr_continuous_grey}
\title{Create shades of grey}
\usage{
clr_continuous_grey(n = 20, start = 0.3, end = 0.8, reverse = FALSE)
}
\arguments{
\item{n}{Number of colors. Default is \code{20}}

\item{start, end}{Hues used for first and last colors, respectively. Enter as numeric 0-1}

\item{reverse}{Whether to reverse default color order. If \code{FALSE} (default), colors are ordered from darkest to lightest. If \code{TRUE}, colors are ordered from lightest to darkest}
}
\value{
Hexadecimal codes
}
\description{
Create shades of grey
}
\examples{
# clr_continuous_grey(20)

}
