% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_continuous_spectral}
\alias{clr_continuous_spectral}
\title{Spectral palette}
\usage{
clr_continuous_spectral(n = 20, type = "dark")
}
\arguments{
\item{n}{Number of colors. Default is \code{20}}

\item{type}{Options: \code{"dark"} (default), \code{"medium"}, \code{"light"}, \code{"turbo"}}
}
\description{
Spectral palette
}
\examples{

# clr_continuous_spectral(20)

}
