% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_continuous_viridis}
\alias{clr_continuous_viridis}
\title{Viridis palette}
\usage{
clr_continuous_viridis(
  n = 20,
  start = 0,
  end = 1,
  palette = "viridis",
  reverse = FALSE,
  alpha = 1
)
}
\arguments{
\item{n}{Number of colors. Default is \code{20}}

\item{start, end}{Hues used for first and last colors, respectively}

\item{palette}{Options include \code{"viridis"} (default), \code{"plasma"}, \code{"magma"}, \code{"inferno"}, \code{"cividis"}, \code{"rocket"}, \code{"turbo"}, \code{"mako"}}

\item{reverse}{Whether to reverse default color order. If \code{FALSE} (default), colors are ordered from darkest to lightest. If \code{TRUE}, colors are ordered from lightest to darkest}

\item{alpha}{Transparency. Numeric in between 0 and 1. Default is \code{1}}
}
\value{
Hexadecimal codes for viridis color scale
}
\description{
Viridis palette
}
\examples{
# clr_continuous_viridis(20)

}
