% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_luminance}
\alias{clr_luminance}
\title{Calculate luminance of a color}
\usage{
clr_luminance(color)
}
\arguments{
\item{color}{Color entered as quoted color or hexadecimal code}
}
\value{
Numeric with same length as input
}
\description{
Calculate luminance of a color
}
