% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_mix}
\alias{clr_mix}
\alias{clr_opposite}
\alias{clr_lighter}
\alias{clr_alpha_filter}
\alias{clr_darker}
\alias{clr_desaturate}
\title{Mix 2 colors}
\usage{
clr_mix(..., ratio = 1, prop_color1 = NULL, return_input = FALSE)

clr_opposite(color, return_input = FALSE)

clr_lighter(color, lightness = 0.5, return_input = FALSE)

clr_alpha_filter(color, alpha = 0.8, return_input = FALSE)

clr_darker(color, darkness = 0.2, return_input = FALSE)

clr_desaturate(color, saturation = 0.5, return_input = FALSE)
}
\arguments{
\item{...}{Character vector or comma separated list of colors. Enter as quoted color names or hexadecimal codes}

\item{ratio}{Ratio of 1st color to 2nd color. Enter as numeric (i.e. 3 for 3:1 mixing). Default is \code{1} (i.e. 1:1 mixing)}

\item{prop_color1}{Proportion of 1st color relative to second color in mixing. Enter as numeric 0-1. If value entered for \code{prop_color1}, then \code{ratio} is ignored}

\item{return_input}{If \code{FALSE} (default), only mixed color is returned. If \code{TRUE}, input color included in output}

\item{color}{Enter as character vector of color names or hexadecimal codes}

\item{lightness}{Amount of transparency to add. Enter as numeric 0-1. Default is \code{0.5}. To create color eqivalent to output of ggplot with alpha applied, enter 1 - a}

\item{alpha}{Transparency. Default is \code{0.8}}

\item{darkness}{Fraction (numeric 0-1) of black to add to starting color. Default is \code{0.2}}

\item{saturation}{Level of saturation. Enter as numeric 0-1. Default is \code{0.5}}
}
\value{
Hexadecimal code of color generated by mixing input colors +/- input colors

Hexadecimal codes as character vector containing opposite color +/- input color

Hexadecimal code of lightened color +/- input color

Hexadecimal code of new color

Hexadecimal code of darkened color +/- input color

Hexadecimal code of desaturated color +/- input color
}
\description{
Functionality from Gaston Sanchez's excellent package colortools
}
