% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_sort}
\alias{clr_sort}
\title{Sort hsl colors}
\usage{
clr_sort(x, order = c("h", "s", "l"), rev = FALSE)
}
\arguments{
\item{x}{Either a character vector (color names or hexadecimal codes) or matrix (rows for rgb or hsl)}

\item{order}{Options: \code{"hex"} (alphabetical), \code{"name"} (alphabetical based on color names, or a vec containing 1 or more rows in rgb, hsl, or hcl to sort by in order of importance (i.e. \code{c("r", "g", "b")}, \code{c("h", "s", "l")}, \code{"r"}, \code{"l"}, etc.). Enter as character vector. Default is \code{c("h", "s", "l")}}

\item{rev}{Logical vector with length of 1 or same length as \code{order} indicating whether to reverse the order for each component in \code{order}. To sort from darkest to lightest, set \code{order = c("h", "s", "l")} and \code{rev = c(FALSE, FALSE, TRUE)}}
}
\value{
Character vector of hexadecimal codes
}
\description{
Sort hsl colors
}
