% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_contrast_ratio}
\alias{clr_contrast_ratio}
\alias{clr_text}
\title{Calculate contrast ratio between colors}
\usage{
clr_contrast_ratio(col1, col2 = "white", as_df = FALSE)

clr_text(...)
}
\arguments{
\item{col1, col2}{Colors 1 and 2. col1 can be vector of colors. Default for col2 is \code{"white"}}

\item{as_df}{If \code{TRUE}, output will be data frame with columns for col1, col2, ratio, contrast}

\item{...}{List, character vector, or comma separated list of colors as quoted color names or hexadecimal codes}
}
\value{
Ratio of luminance for 2 colors. > 3 decent, > 4.5 preferred. If \code{as_df = TRUE}, output will be data frame with columns col1, col2, ratio, contrast
}
\description{
Functionality from colorspace package
}
\examples{
# clr_text(c("#0072B5", "#BC3C29"))

}
