% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{cmyk2hex}
\alias{cmyk2hex}
\alias{col2hex}
\alias{rgb2hsl}
\alias{col2hsl}
\alias{hsl2rgb}
\alias{rgb2hex}
\alias{hsl2hex}
\title{Convert CMYK to hexadecimal code}
\usage{
cmyk2hex(c, m, y, k, alpha = NULL)

col2hex(x)

rgb2hsl(x)

col2hsl(x)

hsl2rgb(x)

rgb2hex(r, g = NULL, b = NULL, alpha = NULL)

hsl2hex(x)
}
\arguments{
\item{c}{Cyan percentage. Enter as integer 0-100}

\item{m}{Magenta percentage. Enter as integer 0-100}

\item{y}{Yellow percentage. Enter as integer 0-100}

\item{k}{Black percentage. Enter as integer 0-100}

\item{alpha}{Transparency. Enter as numeric 0-1. If \code{r} is a matrix with alpha in 4th row, \code{alpha} will override 4th row of \code{r}. If \code{alpha = NULL}, transparency is ignored}

\item{x}{Matrix with 3 rows for hue, saturation, luminance}

\item{r}{Either red color as integer of values 0-255 or a matrix with rows for red, green, blue (can also contain alpha) and 1 column per color}

\item{g, b}{Green and blue as integer vector of values 0-255}
}
\value{
Hexadecimal code(s). Length equal to \code{length(c)}

Hexadecimal code as character vector with length equal to input

Matrix with 3 rows for hue, saturation, luminance

Matrix with 3 rows for hue, saturation, luminance

Matrix with 3 rows for red, green, blue. 1 column per color

Hexadecimal code as character vector with length equal to \code{length(r)} (or \code{ncol(r)} if \code{r} is a matrix

Hexadecimal code as character vector with length equal to \code{ncol(x)}
}
\description{
Functionality from plotwidgets package

Functionality from plotwidgets package
}
