% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{col_perc}
\alias{col_perc}
\title{Convert counts in contingency table to column percentages}
\usage{
col_perc(x, ..., incl_na = FALSE, digits = 1)
}
\arguments{
\item{x}{Contingency table or data frame}

\item{...}{If \code{x} is a data frame, enter column names using tidyselect syntax}

\item{incl_na}{If \code{x} is a data frame, \code{incl_na} determines whether missing values form distinct groups}

\item{digits}{Number of digits to include after decimal. Default is \code{1}}
}
\value{
Table containing column percentages (i.e., column percentages sum to 100)
}
\description{
Convert counts in contingency table to column percentages
}
