% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{col_prop}
\alias{col_prop}
\title{Convert counts in contingency table to column proportions}
\usage{
col_prop(x, ..., incl_na = FALSE)
}
\arguments{
\item{x}{Contingency table or data frame}

\item{...}{If \code{x} is a data frame, enter column names using tidyselect syntax}

\item{incl_na}{If \code{x} is a data frame, \code{incl_na} determines whether missing values form distinct groups}
}
\value{
Table containing column proportions (i.e., column proportions sum to 1)
}
\description{
Convert counts in contingency table to column proportions
}
