% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combos}
\alias{combos}
\title{Generate all combinations of any number of input vectors}
\usage{
combos(..., n = 2, col_names = paste0("V", seq_len(n)), na.rm = TRUE)
}
\arguments{
\item{...}{Enter 1 or more vectors}

\item{n}{Number of elements from \code{x} per combination. Default is \code{2}. Only relevant input is a single vector}

\item{col_names}{Character vector containing names of columns to be generated}

\item{na.rm}{If \code{TRUE} (default), missing values are removed from vectors prior to generating combinations. Only relevant if 1 vector is supplied}
}
\value{
Data frame of combinations. If 1 vector, output is a data frame with 1 row for each unique combination of elements taken n at a time. Pairs of elements not repeated (i.e. if a row of output is a data frame contains a in V1 and B in V2 there will be no row with b in V1 and a in V2) entered. If > 1 vector entered, output consists of a data frame containing 1 column for each input vector and 1 row for each possible combination of values across vectors
}
\description{
Generate all combinations of any number of input vectors
}
