% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combos_list}
\alias{combos_list}
\title{List of combinations}
\usage{
combos_list(x, n = 2, na.rm = TRUE)
}
\arguments{
\item{x}{Vector. Number of elements must be > \code{n}}

\item{n}{Number of elements per group. Must be > 0. Default is \code{2} (pairs of elements in \code{x})}

\item{na.rm}{If TRUE (default), missing values are removed from x prior to generating combinations}
}
\value{
List of vectors (each of length \code{n}) containing unique \code{n} element combos of values in \code{x}
}
\description{
Functionality from \code{utils::combn}
}
