% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_means.R
\name{compare_means}
\alias{compare_means}
\title{Compare continuous variable between 2 groups}
\usage{
compare_means(
  df,
  formula = NULL,
  ...,
  x = NULL,
  y = NULL,
  ns_symbol = "ns",
  normality_test = p_shapiro,
  variance_test = p_F_test,
  summary_fns = list(median = Median, mean = Mean, q1 = Q1, q3 = Q3, sd = SD, se = SE, ci
    = CI, min = min, max = max),
  vars_remove = c("p_normality_1", "p_normality_2", "p_var_test"),
  paired = FALSE
)
}
\arguments{
\item{df}{Data frame in long format}

\item{formula}{enter as continuous variable ~ grouping variable}

\item{...}{Subgrouping variables (statistical test will be performed for each unique combination of subgrouping variables). Enter as comma separated list of quoted or unquoted variable names. If quoted, can wrap in \code{c()}. Can also enter character vector containing variable names}

\item{x, y}{Variables for grouping variable (\code{x}) and continuous variable (\code{y}). Enter as quoted or unquoted variable names}

\item{ns_symbol}{Symbol to use when P > 0.05. Default is \code{"ns"}}

\item{normality_test}{Normality test used to determine whether t-test or Mann Whitney U test will be used. Enter as unquoted function name. Function must take a numeric vector as input and returns p value. Default is \code{p_shapiro}}

\item{variance_test}{Variance test used to determine whether Welch's correction should be used for t-test (if P value for variance test is < 0.05). Enter as unquoted function name. Function must take a numeric vector as input and returns p value. Default is \code{p_F_test}}

\item{summary_fns}{Named list of summary functions that will be applied to each group. Enter as \code{list(col_name = fn_name)}}

\item{vars_remove}{Variables to remove from output. Enter as comma separated list of quoted variable names wrapped in \code{c()}}

\item{paired}{If \code{FALSE} (default), unpaired test performed. If \code{TRUE}, paired test performed}
}
\description{
Compare continuous variable between 2 groups
}
