% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{convert}
\alias{convert}
\title{Convert units}
\usage{
convert(x, from, to, pt_per_in = 72.27)
}
\arguments{
\item{x}{Quantity to convert. Units of \code{x} specified in \code{from} argument}

\item{from}{Units for \code{x}. Options: \code{"in"}, \code{"cm"}, \code{"mm"}, \code{"pt"}, \code{"lines"}, \code{"npc"}, \code{"C"}, \code{"F"}. Enter as quoted units}

\item{to}{Desired output units. Same options as \code{from}. Enter as quoted unis}

\item{pt_per_in}{Points per inch. Default is \code{72.27} (used by ggplot2). Powerpoint uses 72.009}
}
\value{
Value expressed in units determined by \code{to} argument
}
\description{
Convert units
}
