% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{convert_time}
\alias{convert_time}
\title{Convert between units of time}
\usage{
convert_time(x, from, to)
}
\arguments{
\item{x}{Numeric vector}

\item{from, to}{Initial and desired units of time, respectively. Options for both: \code{"msec"}, \code{"sec"}, \code{"min"}, \code{"hr"}, \code{"day"}, \code{"week"}, \code{"month"} (30 days), \code{"year"}, \code{"decade"}, \code{"century"}, \code{"millennium"}}
}
\value{
Numeric vector with length equal to input
}
\description{
Convert between units of time
}
