% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{copy_to_clipboard}
\alias{copy_to_clipboard}
\title{Copy entered text to clipboard in mac}
\usage{
copy_to_clipboard(...)
}
\arguments{
\item{...}{Contents to copy to clipboard}
}
\description{
Copy entered text to clipboard in mac
}
