% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor_pairs}
\alias{cor_pairs}
\title{Pairwise combination of p-values and correlation coefficients}
\usage{
cor_pairs(df, ..., method = c("spearman", "pearson", "kendall"), ci = 0.95)
}
\arguments{
\item{df}{Data frame in wide format or grouped data frame}

\item{...}{Variables in \code{df} to include in correlation analysis. Enter using tidyselect syntax}

\item{method}{Method of comparison. Options: \code{"spearman"} (default), \code{"pearson"}, \code{"kendall"}}

\item{ci}{Confidence level. Default is \code{0.95}. Enter as numeric 0-1}
}
\value{
Data frame with columns col_1, col_2 (variables compared), cor, cor_lower, cor_upper, p, label, method, n, n_nna, perc_na
}
\description{
Pairwise combination of p-values and correlation coefficients
}
