% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{count_any_of}
\alias{count_any_of}
\alias{count_none_of}
\title{Count the number of times a value appears in a vector}
\usage{
count_any_of(x, ...)

count_none_of(x, ...)
}
\arguments{
\item{x}{Vector}

\item{...}{Value(s) to exclude from \code{x}}
}
\value{
Number of times value appears in \code{x}

Number of times values other than those included in values argument appear in \code{x}
}
\description{
Count the number of times a value appears in a vector

Count the number of values in a vector other than a predefined subset
}
