% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{Coxph}
\alias{Coxph}
\alias{Survfit}
\alias{Survdiff}
\title{coxph wrapper function}
\usage{
Coxph(df, predictor_var, time_var = "time", outcome_var = "death", ...)

Survfit(df, predictor_var, time_var = "time", outcome_var = "death", ...)

Survdiff(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  method = "MH",
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{predictor_var}{Variable(s) to use as covariates in model. Enter as comma separated list of quoted variables names wrapped in \code{c()}}

\item{time_var}{Variable containing follow up time. Enter as quoted variable name. Default is "time"}

\item{outcome_var}{Variable containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as quoted variable name. Default is \code{death}}

\item{...}{Arguments passed to \code{survival::coxph}}

\item{method}{Method for determining P value. Options: \code{MH} (default, Mantel-Haenszel test), \code{PP} (Peto and Peto modification of Gehan-Wilcoxon test)}
}
\value{
Output from \code{survival::coxph}
}
\description{
coxph wrapper function

survfit wrapper function

survdiff wrapper function
}
