% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{create_formula}
\alias{create_formula}
\title{Create formula}
\usage{
create_formula(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{Quoted variable name or comma separated list of quoted variable names}
}
\description{
Create formula
}
