% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{create_km_data}
\alias{create_km_data}
\title{Generate Kaplan-Meier estimate data for each group}
\usage{
create_km_data(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  times = seq.int(from = 0, to = ceiling(max(df[[time_var]])), by = 1L)
)
}
\arguments{
\item{df}{Data frame. May contain missing values}

\item{predictor_var}{Variable(s) to use as covariates in model. Enter as comma separated list of quoted variables names wrapped in \code{c()}}

\item{time_var}{Variable containing follow up time. Enter as quoted variable name. Default is \code{"time"}}

\item{outcome_var}{Variable containing outcome variable coded as \code{1} (event) or \code{0} (no event/censored). Enter as quoted variable name. Default is \code{"death"}}

\item{times}{Time points to be included in output data. Default uses 0 to max time by 1}
}
\value{
Data frame used for plotting with columns for group, time, n_risk, n_event, n_censor, surv, surv_lower, surv_upper, surv_se, cumulative_hazard, cumulative_hazard_sd
}
\description{
Functionality from print and summary methods for \code{survfit} object
}
