% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{create_levels}
\alias{create_levels}
\alias{as_fct}
\title{Create reasonable levels for factor regardless of input variable class}
\usage{
create_levels(x, reverse = FALSE, droplevels = TRUE)

as_fct(x, reverse = FALSE, droplevels = FALSE)
}
\arguments{
\item{x}{Variable to converted to factor}

\item{reverse}{If \code{TRUE} order of levels is reversed. \code{FALSE} by default}

\item{droplevels}{If \code{TRUE} (default) and \code{x} is a factor, unused levels are dropped}
}
\value{
Character vector containing factor levels. Enter as \code{factor(x, levels = create_levels(x))}
}
\description{
Create reasonable levels for factor regardless of input variable class

Convert vector to factor
}
