% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{create_surv_formula}
\alias{create_surv_formula}
\title{Create survival formula}
\usage{
create_surv_formula(
  predictor_var = ".",
  time_var = "time",
  outcome_var = "death"
)
}
\arguments{
\item{predictor_var}{Variable(s) to use as covariates in model. Enter as character vector. Default includes all variables}

\item{time_var}{Variable containing follow up time. Enter as quoted variable name. Default is \code{time}}

\item{outcome_var}{Variable containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as quoted variable name. Default is \code{death}}
}
\value{
Survival formula
}
\description{
Create survival formula
}
