% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{crossings}
\alias{crossings}
\title{Crossing for 2 or more vectors}
\usage{
crossings(..., .prefix_colnames = "V")
}
\arguments{
\item{...}{Either a comma separated list of vectors not wrapped in \code{c()} or a list containing vectors}

\item{.prefix_colnames}{Prefix for columns names. Default is \code{"V"}}
}
\value{
Data frame with 1 column per vector entered into \code{...} and each row containing unique combinations of 1 element per vector
}
\description{
Functionality from \code{base::expand.grid}
}
