% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{cutpoint_survival}
\alias{cutpoint_survival}
\title{Cutpoint for survival outcome}
\usage{
cutpoint_survival(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  min_prop = 0.1
)
}
\arguments{
\item{df}{Data frame}

\item{predictor_var}{Continuous predictor variable(s). Enter as comma separated list of quoted variable names wrapped in \code{c()}}

\item{time_var}{Variable containing follow up time. Enter as quoted variable name. Default is \code{time}}

\item{outcome_var}{Variable containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as quoted variable name. Default is \code{death}}

\item{min_prop}{Minimum proportion of all subjects in smaller of 2 groups when stratifying by cutpoint. Enter as numeric proportion. Default is \code{0.1}}
}
\value{
Data frame containing columns for outcome_var, time_var, predictor_var, cutpoint
}
\description{
Cutpoint for survival outcome
}
