% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{day}
\alias{day}
\alias{month}
\alias{year}
\title{Extract day of month from date}
\usage{
day(x)

month(x)

year(x)
}
\arguments{
\item{x}{Vector of dates}
}
\value{
Integer vector containing day, month, or year (4 digit year). Length equal to input
}
\description{
Extract day of month from date

Extract month from date

Extract year from date
}
