% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{df_as_numeric}
\alias{df_as_numeric}
\title{If a column is coercible to a numeric, convert to numeric}
\usage{
df_as_numeric(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with same number of columns as input, but all columns that are coercible to a numeric will be converted to numeric
}
\description{
If a column is coercible to a numeric, convert to numeric
}
