% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_coerce.R
\name{df_to_list_rowwise}
\alias{df_to_list_rowwise}
\title{Convert data frame to list of rows}
\usage{
df_to_list_rowwise(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
List of length \code{nrow(df)} with each element containing a data frame with 1 row
}
\description{
Convert data frame to list of rows
}
