% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_coerce.R
\name{df_to_named_vector}
\alias{df_to_named_vector}
\title{Convert 2 columns of data frame to a named vector}
\usage{
df_to_named_vector(df, values, names)
}
\arguments{
\item{df}{Data frame}

\item{values}{Column in \code{df} to create values in output vector. Enter as quoted or unquoted column name}

\item{names}{Column in \code{df} to create names in output vector. Enter as quoted or unquoted column name}
}
\value{
Named vector
}
\description{
Convert 2 columns of data frame to a named vector
}
