% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{Diff}
\alias{Diff}
\title{Difference between values in a vector}
\usage{
Diff(x, lag = 1L)
}
\arguments{
\item{x}{Numeric vector (not dates, time series, or matrices)}

\item{lag}{Difference in position between numbers to be compared. Enter as integer. Default is \code{1}}
}
\value{
Numeric vector of length equal to \code{length(x) - lag}
}
\description{
Functionality from \code{base::diff}
}
