% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_misc.R
\name{diff_mean}
\alias{diff_mean}
\title{Mean difference with confidence intervals}
\usage{
diff_mean(
  x,
  y,
  test_fn = NULL,
  normality_test = p_shapiro,
  ci = 0.95,
  paired = FALSE,
  hypothesis_type = "two.sided",
  ...
)
}
\arguments{
\item{x, y}{Values in group 1 and 2 respectively. If paired, order of \code{x} and \code{y} must be the same and their lengths must be equal}

\item{test_fn}{Function used for testing. Options: \code{NULL} (default, uses normality test to determine whether to use t-test or Wilcoxon ranked sum test), \code{diff_mean_ttest}, \code{diff_mean_wilcox}}

\item{normality_test}{Normality test used to determine method for calculating mean difference. Default is \code{p_shapiro.} Only relevant when \code{test_fn} is \code{NULL}}

\item{ci}{Type of confidence interval. Default is \code{0.95}}

\item{paired}{If \code{FALSE} (default), unpaired test is performed}

\item{hypothesis_type}{Type of hypothesis test to perform. Options: \code{"two.sided"} (default), \code{"less"}, \code{"greater"}}

\item{...}{Arguments passed to either \code{diff_mean_ttest} or \code{diff_mean_wilcox}}
}
\value{
Length 3 numeric vector containing mean difference (diff_mean), lower ci (ci_lower), and upper ci (ci_upper)
}
\description{
Mean difference with confidence intervals
}
