% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_misc.R
\name{diff_mean_ttest}
\alias{diff_mean_ttest}
\title{t-test estimation of mean difference}
\usage{
diff_mean_ttest(
  x,
  y = NULL,
  ci = 0.95,
  welch = NULL,
  paired = FALSE,
  variance_test = .p_F_test,
  hypothesis_type = "two.sided",
  ...
)
}
\arguments{
\item{x, y}{Values in group 1 and 2 respectively. If \code{paired}, \code{y} is ignored}

\item{ci}{Confidence level. Default is \code{0.95}}

\item{welch}{If \code{NULL} (default), use of Welch's correction is determined by variance test (If P < 0.05, variance is not equal between groups and Welch's correction is applied). If \code{TRUE}, Welch's correction is applied, otherwise student's t-test is performed}

\item{paired}{If \code{FALSE} (default), unpaired test is performed}

\item{variance_test}{Function used to perform variance test. Must take numeric vector and returns P value. Default is \code{p_F_test}}

\item{hypothesis_type}{Type of hypothesis test to perform. If \code{"two.sided"} (default), 2 sided test is performed. Other options: \code{"less"}, \code{"greater"}}

\item{...}{Not used}
}
\value{
Length 3 numeric vector containing mean difference (diff_mean), lower ci (ci_lower), and upper ci (ci_upper)
}
\description{
t-test estimation of mean difference
}
