% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_misc.R
\name{diff_mean_wilcox}
\alias{diff_mean_wilcox}
\alias{diff_mean_mann_whitney}
\title{Wilcoxon estimation of mean difference}
\usage{
diff_mean_wilcox(
  x,
  y = NULL,
  hypothesis_type = c("two.sided", "less", "greater"),
  paired = FALSE,
  correct = TRUE,
  ci = 0.95,
  ...
)

diff_mean_mann_whitney(
  x,
  y = NULL,
  hypothesis_type = c("two.sided", "less", "greater"),
  paired = FALSE,
  correct = TRUE,
  ci = 0.95,
  ...
)
}
\arguments{
\item{x, y}{Values in group 1 and 2 respectively. If paired, order of \code{x} and \code{y} must be the same and their lengths must be equal. Must have missing values removed}

\item{hypothesis_type}{Type of hypothesis test to perform. Options: \code{"two.sided"} (default), \code{"less"}, \code{"greater"}. Enter as quoted hypothesis test type}

\item{paired}{If \code{FALSE} (default), unpaired test is performed}

\item{correct}{If \code{TRUE} (default), continuity correction is applied}

\item{ci}{Confidence level. Default is \code{0.95}}

\item{...}{Not used}
}
\value{
Length 3 numeric vector containing mean difference (diff_mean), lower ci (ci_lower), and upper ci (ci_upper)
}
\description{
Wilcoxon estimation of mean difference

Alias for diff_mean_wilcox
}
