% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_misc.R
\name{diff_prop}
\alias{diff_prop}
\title{Difference and CI between proportions}
\usage{
diff_prop(
  x,
  ci = 0.95,
  yates = FALSE,
  hypothesis_type = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x}{Table (2 x 2) in which columns represent controls (1st column) and cases (2nd column) and rows represent negative (1st row) and positive (2nd row) individuals}

\item{ci}{Confidence interval. Default is \code{0.95}}

\item{yates}{If \code{FALSE} (default), Yates continuity correction is not used If \code{TRUE} and x is 2 x 2, Yates' continuity correction is applied}

\item{hypothesis_type}{Type of hypothesis test to perform. Options: \code{"two.sided"} (default), \code{"less"}, \code{"greater"}. Enter as quoted hypothesis test type}
}
\value{
Length 3 numeric vector containing diff_prop, lower_ci, upper_ci
}
\description{
Difference and CI between proportions
}
