% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{dir_exists}
\alias{dir_exists}
\alias{is_dir}
\title{Determine whether input is a directory (and not a file path)}
\usage{
dir_exists(x)

is_dir(x)
}
\arguments{
\item{x}{Character vector of directory paths}
}
\value{
Logical with length equal to input length
}
\description{
Determine whether input is a directory (and not a file path)

Alias for \code{dir_exists}
}
