% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{.ggplot_extract_scale}
\alias{.ggplot_extract_scale}
\title{Extract scale from ggplot object}
\usage{
.ggplot_extract_scale(x, scale)
}
\arguments{
\item{x}{ggplot object}

\item{scale}{Scale to extract. Enter as string}
}
\value{
Scale object
}
\description{
Extract scale from ggplot object
}
