% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_variance.R
\name{.p_bartlett}
\alias{.p_bartlett}
\title{P value for Bartlett test}
\usage{
.p_bartlett(..., otherwise = 0)
}
\arguments{
\item{...}{Vectors containing values for each group or a list of vectors}

\item{otherwise}{Output if Bartlett test fails. Default is \code{0}}
}
\description{
P value for Bartlett test
}
