% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_variance.R
\name{.p_F_test}
\alias{.p_F_test}
\title{P value for F test (aka variance test)}
\usage{
.p_F_test(x, y, hypothesis_type = "two.sided", otherwise = 0, ...)
}
\arguments{
\item{x, y}{Values for each group. Must have missing values removed}

\item{hypothesis_type}{Type of hypothesis test to perform. Options: \code{"two.sided"} (default), \code{"less"}, \code{"greater"}. Enter as quoted hypothesis test type}

\item{otherwise}{Output if F test fails. Default is \code{0}}

\item{...}{Not used}
}
\description{
P value for F test (aka variance test)
}
