% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_variance.R
\name{.p_levene}
\alias{.p_levene}
\title{P value for Levene test}
\usage{
.p_levene(x, y, brown_forsythe = TRUE, otherwise = 0, ...)
}
\arguments{
\item{x, y}{Values for each group}

\item{brown_forsythe}{If \code{TRUE} (default), Brown Forsythe version of Levene test performed (using median). If \code{FALSE}, original Levene test (using mean) performed}

\item{otherwise}{Output if Levene test fails. Default is \code{0}}

\item{...}{Not used}
}
\description{
P value for Levene test
}
