% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{dots_as_quoted}
\alias{dots_as_quoted}
\title{Collect dots as quoted terms}
\usage{
dots_as_quoted(..., .use_names = FALSE)
}
\arguments{
\item{...}{Comma separated list of quoted or unquoted terms}

\item{.use_names}{Whether to use names. Default is \code{FALSE}}
}
\value{
Character vector (terms entered in dots are not evaluated/interpreted)
}
\description{
Collect dots as quoted terms
}
