% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{dots_as_unquoted}
\alias{dots_as_unquoted}
\title{Collect dots as unquoted terms}
\usage{
dots_as_unquoted(..., .use_names = FALSE)
}
\arguments{
\item{...}{Comma separated list of quoted or unquoted terms}

\item{.use_names}{Whether to use names. Default is \code{FALSE}}
}
\value{
List of unquoted terms (no interpretation/evaluation)
}
\description{
Collect dots as unquoted terms
}
