% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{dots_names}
\alias{dots_names}
\title{Determine names of dots}
\usage{
dots_names(...)
}
\arguments{
\item{...}{Dots}
}
\value{
Character vector of named elements in dots. If unnamed, result is \code{NA}
}
\description{
Determine names of dots
}
