% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{draw_axis_clean}
\alias{draw_axis_clean}
\title{Grob to draw clean axis}
\usage{
draw_axis_clean(
  break_positions,
  break_labels,
  axis_position,
  theme,
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1
)
}
\arguments{
\item{break_positions}{position of ticks}

\item{break_labels}{labels at ticks}

\item{axis_position}{Axis position. Options: \code{"left"} (default for y axis), \code{"bottom"} (default for x axis), \code{"right"}, \code{"top"}}

\item{theme}{Complete theme}

\item{check.overlap}{If \code{FALSE} (default), overlapping labels not removed. If \code{TRUE}, overlapping labels are removed (priority is for first, last, and middle labels)}

\item{angle}{Angle for axis tick labels}

\item{n.dodge}{Number of rows (for y axis) or columns (for x axis) that labels should span when overlapping. Relevant when \code{check.overlap = TRUE}}
}
\value{
Called by \code{guide_gengrob.clean}
}
\description{
Functionality from Charlotte Dawson's excellent package ggprism
}
