% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowplot.R
\name{draw_grob}
\alias{draw_grob}
\title{Draw grob}
\usage{
draw_grob(
  grob,
  x = 0,
  y = 0,
  width = 1,
  height = 1,
  scale = 1,
  clip = "inherit",
  hjust = 0,
  vjust = 0,
  halign = 0.5,
  valign = 0.5
)
}
\arguments{
\item{grob}{grob object}

\item{x, y}{x and y starting positions Enter as numeric 0-1. Default is \code{1} for both}

\item{height, width}{Height and width of grob Enter as numeric 0-1. Default is \code{0} for both}

\item{scale}{Scaling factor for plot. Default is \code{1}}

\item{clip}{Clipping for grob. Default is \code{"inherit"} (inherits from input grob)}

\item{hjust}{Adjust horizontal position of each label. More negative values move the label further to the right on the plot canvas. Can be a single value (applied to all labels) or a vector of values (one for each label). Default is \code{-0.5}}

\item{vjust}{Adjust vertical position of each label. More positive values move the label further down on the plot canvas. Can be a single value (applied to all labels) or a vector of values (one for each label). Default is \code{1.5}}

\item{halign, valign}{Horizontal and vertical alignment, respectively. Enter as numeric 0-1. Default is \code{0.5} for both}
}
\description{
Functionality from Claus Wilke's excellent package cowplot
}
