% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{draw_plot_axes}
\alias{draw_plot_axes}
\title{Draw plot axes only}
\usage{
draw_plot_axes(x, axis = "both")
}
\arguments{
\item{x}{ggplot object}

\item{axis}{Options: \code{"both"} (default), \code{"x"}, \code{"y"}}
}
\description{
Draw plot axes only
}
