% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{dupe_cols}
\alias{dupe_cols}
\title{Identify duplicate columns in data frame}
\usage{
dupe_cols(df, ignore_class = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{ignore_class}{If \code{TRUE}, all columns are converted to characters prior to analysis}
}
\value{
Named character vector of duplicated columns in \code{df} in which values represent column numbers and names represent column names
}
\description{
Identify duplicate columns in data frame
}
