% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{duplicated_incl_first}
\alias{duplicated_incl_first}
\alias{dupe_vals}
\title{Include the first occurrence of the subsequently duplicated value in duplicated output}
\usage{
duplicated_incl_first(x)

dupe_vals(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Logical vector in which the index of any value that is duplicated is \code{TRUE.} Unique values are \code{FALSE}

Values in \code{x} that are duplicated
}
\description{
Include the first occurrence of the subsequently duplicated value in duplicated output

Extract duplicated entries in a vector
}
