% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{earliest_sig_surv}
\alias{earliest_sig_surv}
\title{Evaluate all time points to identify time point when curves first separate}
\usage{
earliest_sig_surv(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  time_by = 1L
)
}
\arguments{
\item{df}{Data frame (can be grouped)}

\item{predictor_var}{Variable(s) to use as covariates in model. Enter as comma separated list of quoted variables names wrapped in \code{c()}}

\item{time_var}{Variable containing follow up time. Enter as quoted variable name. Default is \code{time}}

\item{outcome_var}{Variable containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as quoted variable name. Default is \code{death}}

\item{time_by}{Interval between time points. Default is \code{1}}
}
\value{
Data frame containing columns for time, variable, predictor, hr, hr_lower, hr_upper, p_lrt, p_wald, n_events, n, n_events_ever, cfd_events (proportion of events that occurred by a given time as as a proportion of total events that ever occurred)
}
\description{
Evaluate all time points to identify time point when curves first separate
}
