% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{elementwise}
\alias{elementwise}
\title{Element-wise function}
\usage{
elementwise(.list, .f)
}
\arguments{
\item{.list}{List of vectors passed to \code{.f} to be evaluated element-wise}

\item{.f}{Function to be applied to input vectors}
}
\description{
Element-wise function
}
