% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errorbar.R
\name{errorbar}
\alias{errorbar}
\title{Error bars}
\usage{
errorbar(
  summary_fn = Mean,
  error_fn = SE,
  error_limits = "upper",
  line_thickness = 1,
  width = 0.1,
  dodge = NULL,
  ...
)
}
\arguments{
\item{summary_fn}{Function used to generated summary estimate for y. Default is \code{Mean}}

\item{error_fn}{Function used to determine error limits for y. Default is \code{SE}}

\item{error_limits}{Options: \code{"upper"} (default), \code{"lower"}, \code{"both"}}

\item{line_thickness}{Thickness of error bar lines in pts. Default is \code{1}}

\item{width}{Width of error bar. Default is \code{0.1}}

\item{dodge}{Width used to dodge along x axis. Default is \code{0.9}}

\item{...}{Arguments passed to \code{stat_summary()}}
}
\value{
Add errorbars to plot using plot + errorbar()
}
\description{
Error bars
}
