% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{eval_all_cutpoints_survival}
\alias{eval_all_cutpoints_survival}
\title{Evaluate all possible cutpoints}
\usage{
eval_all_cutpoints_survival(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  type = ">=",
  min_prop = 0.1
)
}
\arguments{
\item{df}{Data frame}

\item{predictor_var}{Continuous predictor variable. Enter as quoted or unquoted variable name}

\item{time_var}{Variable containing follow up time. Enter as quoted variable name. Default is \code{time}}

\item{outcome_var}{Outcome variable coded as \code{1} (event) or \code{0} (no event/censored). Enter as quoted variable name. Default is \code{death}}

\item{type}{If \code{>=} (default), positivity is determined by values greater than or equal to cutpoint.  If \code{>}, positivity is determined by values of predictor_var greater than cutpoint}

\item{min_prop}{Minimum proportion of all subjects in smaller of 2 groups created when stratifying by cutpoint to positive and negative groups. Enter as numeric proportion. Default is \code{0.1}}
}
\value{
Data frame with 1 row for each unique value of predictor variable and columns for outcome_var, time_var, predictor_var, type, cutpoint, hr, hr_lower, hr_upper, p_lrt, label, p_wald, n, n_events, pos, perc_pos, neg, perc_neg
}
\description{
Positive refers to patients with predictor_var > cp (not >= cp)
}
