% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{eval_cutpoint_binary}
\alias{eval_cutpoint_binary}
\alias{eval_all_cutpoints_binary}
\title{Evaluate a specific cutpoint for classification of a binary variable}
\usage{
eval_cutpoint_binary(
  df,
  outcome_var,
  predictor_var,
  cp,
  type = ">=",
  cutpoint = cp
)

eval_all_cutpoints_binary(
  df,
  outcome_var,
  predictor_var,
  type = ">=",
  min_prop = 0.1
)
}
\arguments{
\item{df}{Data frame}

\item{outcome_var}{Binary outcome variable. Enter as quoted or unquoted variable name}

\item{predictor_var}{Continuous predictor variable. Enter as quoted or unquoted variable name}

\item{cp}{Cutpoints to be evaluated. Enter as numeric vector}

\item{type}{If \code{>=} (default), positivity is determined by values greater than or equal to cutpoint.  If \code{>}, positivity is determined by values of predictor_var greater than cutpoint}

\item{cutpoint}{Alias for \code{cp}}

\item{min_prop}{Minimum proportion of all subjects in smaller of 2 groups when stratifying by cutpoint. Enter as numeric proportion. Default is \code{0.1}}
}
\value{
Data frame with 1 row for each cutpoint and columns for outcome_var, predictor_var, type, cutpoint, sens, sens_lower, sens_upper, spec, spec_lower, spec_upper, p, label, method, p_fisher, p_chi, youden, cohen_kappa, cohen_kappa_lower, cohen_kappa_upper, tpr, fpr, tnr, fnr, ppv, npv, fdr, lr_pos, lr_neg, dor, dor_lower, dor_upper, accuracy, auc, gini, jaccard_index, n, cases, controls, pos, neg, tp, fn, fp, tn, perc_cases, perc_controls, perc_pos, perc_neg

Data frame with 1 row for each unique value of predictor variable and columns for outcome_var, predictor_var, type, cutpoint, sens, sens_lower, sens_upper, spec, spec_lower, spec_upper, p, label, method, p_fisher, p_chi, youden, cohen_kappa, cohen_kappa_lower, cohen_kappa_upper, tpr, fpr, tnr, fnr, ppv, npv, fdr, lr_pos, lr_neg, dor, dor_lower, dor_upper, accuracy, auc, gini, jaccard_index, n, cases, controls, pos, neg, tp, fn, fp, tn, perc_cases, perc_controls, perc_pos, perc_neg
}
\description{
Evaluate a specific cutpoint for classification of a binary variable

Evaluate all cutpoints for classification of a binary variable
}
