% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{extract}
\alias{extract}
\title{Extract a subset of a vector, list, or data frame, either by position or names}
\usage{
extract(x, idx = 1, depth = 1)
}
\arguments{
\item{x}{Vector, list, or data frame}

\item{idx}{Either a numeric vector of positions or character vector of names to extract}

\item{depth}{Number of brackets. Options: \code{1}, \code{2}. Default is \code{1}}
}
\value{
Subset of \code{x}
}
\description{
Extract a subset of a vector, list, or data frame, either by position or names
}
