% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{fct_cross}
\alias{fct_cross}
\title{Add new factor variable to data frame by pasting values from existing columns together}
\usage{
fct_cross(
  df,
  ...,
  add_colname_prefix = TRUE,
  sep_colname_value = " = ",
  sep_cols = ", ",
  new_colname = "group",
  reverse = FALSE
)
}
\arguments{
\item{df}{Data frame}

\item{...}{Columns to paste together. Enter using tidyselect syntax}

\item{add_colname_prefix}{If \code{TRUE} (default), values in each column are prefixed by column names}

\item{sep_colname_value}{Symbol used to separate column name and corresponding values. Default is \code{" = "}. Only relevant when \code{add_colname_prefix = TRUE}}

\item{sep_cols}{Symbol used to separate columns. Default is \code{", "}}

\item{new_colname}{Column name for new variable. Default is \code{"group"}}

\item{reverse}{If \code{TRUE}, levels of output variable are reversed in \code{create_levels()}}
}
\value{
\code{df} with new variable for combination of values in columns defined in \code{...}
}
\description{
Add new factor variable to data frame by pasting values from existing columns together
}
