% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{fct_reorder}
\alias{fct_reorder}
\alias{fct_by_input_order}
\alias{fct_by_count}
\alias{fct_increasing}
\alias{fct_decreasing}
\title{Reorder levels of a factor}
\usage{
fct_reorder(.x, .reorder_by, .fn = Mean, ..., .increasing = TRUE)

fct_by_input_order(.x)

fct_by_count(.x, .increasing = TRUE)

fct_increasing(.x)

fct_decreasing(.x)
}
\arguments{
\item{.x}{Factor vector to be reordered}

\item{.reorder_by}{Vector used to determine new order of \code{.x} levels. Levels of \code{.x} will be determined according to \code{.fn(.reorder_by)} calculation}

\item{.fn}{Function applied to reorder_by to determine order of \code{.x} levels. Default is \code{Mean}}

\item{...}{Arguments passed to \code{.fn}}

\item{.increasing}{If \code{TRUE} (default), levels of \code{.x} will be in ascending order of \code{.fn(.reorder_by)} calculation. If \code{FALSE}, levels of \code{.x} will be in descending order of \code{.fn(.reorder_by)} calculation}
}
\value{
Factor with updated levels according to \code{.fn(.reorder_by)} calculation and increasing argument
}
\description{
Functionality from forcats package
}
