% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{fct_severity}
\alias{fct_severity}
\title{Convert severity to a factor variable}
\usage{
fct_severity(x, levels = NULL)
}
\arguments{
\item{x}{Character vector}

\item{levels}{Severity levels from least to most severe. Enter as character vector}
}
\description{
Convert severity to a factor variable
}
