% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_exists}
\alias{file_exists}
\alias{is_file}
\title{Determine whether input is a file path (and not a directory)}
\usage{
file_exists(x)

is_file(x)
}
\arguments{
\item{x}{Character vector of file paths}
}
\value{
Logical with length equal to input length
}
\description{
Determine whether input is a file path (and not a directory)

Alias for \code{file_exists}
}
