% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_ext<-}
\alias{file_ext<-}
\title{Set file extension}
\usage{
file_ext(x) <- value
}
\arguments{
\item{x}{File name or file path. Enter as character vector. Length can be > 1}

\item{value}{File extension. Enter as quoted file extension without ".". Must have length of 1 or length equal to \code{length(x)}}
}
\value{
Input with new file extension specified by \code{value}
}
\description{
Set file extension
}
