% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_misc.R
\name{find_elbow}
\alias{find_elbow}
\title{Find elbow point in curve}
\usage{
find_elbow(df, formula)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Enter in y ~ x format}
}
\value{
Coordinates of elbow point as length 2 numeric vector
}
\description{
Find elbow point in curve
}
