% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_n_perc}
\alias{format_n_perc}
\title{Format n (or n/total) and \%}
\usage{
format_n_perc(
  n,
  total,
  digits = 1,
  incl_denominator = TRUE,
  perc_first = TRUE,
  bracket_type = "("
)
}
\arguments{
\item{n}{Numerator for percentage. Enter as numeric or character}

\item{total}{Denominator for percentage. Enter as numeric or character}

\item{digits}{Number of digits after decimal. Only used for percentage. Default is \code{1}}

\item{incl_denominator}{If \code{TRUE} (default), format for count is n/total. If \code{FALSE}, format is n}

\item{perc_first}{If \code{TRUE} (default), format is \% (n) or \% (n/n_total). If \code{FALSE}, format is n (\%) or n/total (\%)}

\item{bracket_type}{Type of bracket to surround range. Default is \code{"("}}
}
\description{
Format n (or n/total) and \%
}
