% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_num_range}
\alias{format_num_range}
\title{Format number and range}
\usage{
format_num_range(
  x,
  x_lower,
  x_upper,
  sep = "-",
  bracket_type = "(",
  digits = 2
)
}
\arguments{
\item{x}{Risk ratio. Enter as numeric}

\item{x_lower, x_upper}{Lower and upper confidence intervals of risk ratio. Enter as numeric}

\item{sep}{Separator character between \code{x_lower} and \code{x_upper.} Default is \code{"-"}}

\item{bracket_type}{Type of bracket to surround range. Default is \code{"("}}

\item{digits}{Number of digits to display after decimal point. Default is \code{2}}
}
\description{
Format number and range
}
