% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{format_p_value}
\alias{format_p_value}
\title{Format p values}
\usage{
format_p_value(
  p,
  p_prefix = "P",
  trim_ws = FALSE,
  show_leading_0 = TRUE,
  min_digits = 2
)
}
\arguments{
\item{p}{Numeric vector containing p values}

\item{p_prefix}{Prefix used for P value. Default is \code{"P"}}

\item{trim_ws}{If \code{FALSE} (default), white space is allowed in output. If \code{TRUE}, white space is removed from output}

\item{show_leading_0}{If \code{TRUE} (default), 0s before decimal point are included in output}

\item{min_digits}{Number of digits included after decimal for P values > 0.06}
}
\value{
p value formatted as character
}
\description{
Format p values
}
