% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_percent}
\alias{format_percent}
\title{Format percentages}
\usage{
format_percent(x, digits = 1, big_mark = "")
}
\arguments{
\item{x}{Numeric vector of proportions (i.e. 0.95, not 95)}

\item{digits}{Number of digits to include after decimal. Default is \code{1}}

\item{big_mark}{Character to separate thousands. Default is \code{""}}
}
\description{
Format percentages
}
