% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_scientific}
\alias{format_scientific}
\alias{format_scientific2}
\title{Convert number from scientific notation to custom style scientific notation}
\usage{
format_scientific(x, digits = 1, sep = "×", as_exp = FALSE, trim_ws = FALSE)

format_scientific2(x)
}
\arguments{
\item{x}{Numeric or character containing number}

\item{digits}{Number of digits to include after decimal. Default is \code{1}}

\item{sep}{Separator between base and 10^exponent. Default is multiplication symbol}

\item{as_exp}{If \code{FALSE} (default), output is character. If \code{TRUE}, character is expression}

\item{trim_ws}{If \code{FALSE} (default), white space is allowed in output. If \code{TRUE}, white space is removed from output}
}
\description{
Convert number from scientific notation to custom style scientific notation

Alternative version of format_scientific
}
