% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula2vars}
\alias{formula2vars}
\alias{vars2formula}
\alias{get_vars_formula}
\title{#' Extract x and y variables from formula or x/y inputs}
\usage{
formula2vars(formula = NULL, x = NULL, y = NULL, parent_fn = "")

vars2formula(formula = NULL, x = NULL, y = NULL, parent_fn = "")

get_vars_formula(formula = NULL, x = NULL, y = NULL, parent_fn = "")
}
\arguments{
\item{formula}{Formula in y ~ x format. Left-hand side cannot contain > 1 variable}

\item{x, y}{Character vectors of x and y variables}

\item{parent_fn}{Name of function called. Enter as quoted function name}
}
\value{
Formula or list containing "y" and "x", each a character vector of variables
}
\description{
#' Extract x and y variables from formula or x/y inputs

#' Extract formula from formula or x/y inputs

#' Extract formula and x and y variables from formula or x/y inputs
}
