% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowplot.R
\docType{data}
\name{GeomDrawGrob}
\alias{GeomDrawGrob}
\title{Geom for draw_grob}
\format{
An object of class \code{GeomDrawGrob} (inherits from \code{GeomCustomAnn}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomDrawGrob
}
\description{
Functionality from Claus Wilke's excellent package cowplot
}
\keyword{datasets}
