% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errorbar.R
\docType{data}
\name{GeomErrorUpper}
\alias{GeomErrorUpper}
\title{Error bar - upper half only}
\format{
An object of class \code{GeomErrorUpper} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
GeomErrorUpper
}
\description{
Error bar - upper half only
}
\keyword{datasets}
