% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{geometric_ci}
\alias{geometric_ci}
\title{Geometric confidence interval}
\usage{
geometric_ci(x, log_fn = log1p, exp_fn = expm1, ci = 0.95)
}
\arguments{
\item{x}{Numeric vector}

\item{log_fn}{Function used to log transform data. Default is \code{log1p}}

\item{exp_fn}{Function used to back transform log data. Default is \code{expm1}}

\item{ci}{Type of confidence interval. Default is \code{0.95}}
}
\description{
Geometric confidence interval
}
