% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{geometric_mean}
\alias{geometric_mean}
\alias{geometric_sd}
\alias{geometric_se}
\title{Geometric mean}
\usage{
geometric_mean(x, log_fn = log1p, exp_fn = expm1)

geometric_sd(x, log_fn = log1p, exp_fn = expm1)

geometric_se(x, log_fn = log1p, exp_fn = expm1)
}
\arguments{
\item{x}{Numeric vector}

\item{log_fn}{Function used to log transform data. Default is \code{log1p}}

\item{exp_fn}{Function used to back transform log data. Default is \code{expm1}}
}
\description{
Geometric mean

Geometric standard deviation

Geometric standard error
}
