% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sig_bar.R
\docType{data}
\name{GeomSigBar}
\alias{GeomSigBar}
\title{ggproto for significance bar}
\format{
An object of class \code{GeomSigBar} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomSigBar
}
\description{
ggproto for significance bar
}
\keyword{datasets}
