% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{get_dupes.data.frame}
\alias{get_dupes.data.frame}
\title{get_dupes - data frame}
\usage{
\method{get_dupes}{data.frame}(x, ...)
}
\arguments{
\item{x}{Data frame}

\item{...}{Columns to search for duplicates. Default uses all columns. Enter using tidyselect syntax}
}
\value{
Original data frame with new column (dupe_count) containing the number of duplicate rows
}
\description{
Searches for duplicate rows
}
