% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{get_from_workspace}
\alias{get_from_workspace}
\title{Get objects stored in another R work space}
\usage{
get_from_workspace(..., obj = NULL)
}
\arguments{
\item{...}{Path to directory. Arguments passed to \code{paste_path()}}

\item{obj}{Name of object in work space. Enter as vector or unquoted list of object names (input does not need to be wrapped in \code{c()}). If \code{NULL} (default), output includes all objects in the work space}
}
\value{
If a single object is entered, class of output is determined by object entered. If  multiple objects entered or \code{obj = NULL}, output is a list of objects
}
\description{
Get objects stored in another R work space
}
