% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{get_input}
\alias{get_input}
\title{Get input to function if evaluation is possible, otherwise get input as character}
\usage{
get_input(.input)
}
\arguments{
\item{.input}{Input. Can be piped}
}
\value{
If .input is \code{NULL} or can be evaluated to a vector, output is same as input. Otherwise, output is a character representing the typed input. Objects in global environment are ignored
}
\description{
Get input to function if evaluation is possible, otherwise get input as character
}
