% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{get_piped_input}
\alias{get_piped_input}
\title{Get name of first term in a pipe argument}
\usage{
get_piped_input(x)
}
\arguments{
\item{x}{Input to function (piped or unpiped. Can't be a number or a call)}
}
\value{
\code{x} as a character
}
\description{
Get name of first term in a pipe argument
}
