% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{get_plot_aes}
\alias{get_plot_aes}
\title{Extract aes arguments for each plot layer}
\usage{
get_plot_aes(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
List of aes/variables for each plot layer. Length of output equal to number of plots layers
}
\description{
Extract aes arguments for each plot layer
}
