% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{get_plot_aes_code}
\alias{get_plot_aes_code}
\title{Get R code for plot aesthetics}
\usage{
get_plot_aes_code(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
Named list of aes input for each plot layer. List from 1 layer can be passed to \code{aes()}
}
\description{
Get R code for plot aesthetics
}
