% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_aspect_ratio}
\alias{get_plot_aspect_ratio}
\title{Get plot aspect ratio}
\usage{
get_plot_aspect_ratio(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
List containing entry into theme function for aspect ratio and actual aspect ratio in data units (if set)
}
\description{
Get plot aspect ratio
}
