% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_axis_breaks}
\alias{get_plot_axis_breaks}
\title{Get axis breaks}
\usage{
get_plot_axis_breaks(x, axis = "both", units = "data")
}
\arguments{
\item{x}{ggplot object}

\item{axis}{Options: \code{"both"} (default), \code{"x"}, \code{"y"}}

\item{units}{Options: \code{"data"} (default, raw data units), \code{"npc"}}
}
\description{
Get axis breaks
}
