% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_axis_labels}
\alias{get_plot_axis_labels}
\alias{get_plot_axis_title}
\title{Get axis tick labels}
\usage{
get_plot_axis_labels(x, axis = "both")

get_plot_axis_title(x, axis = "both")
}
\arguments{
\item{x}{ggplot object}

\item{axis}{Options: \code{"both"} (default), \code{"x"}, \code{"y"}}
}
\value{
Character vector containing axis tick labels
}
\description{
Get axis tick labels

Get axis title
}
