% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_axis_vars}
\alias{get_plot_axis_vars}
\title{Get variables used to generate x and y axes}
\usage{
get_plot_axis_vars(x, axis = "both")
}
\arguments{
\item{x}{ggplot object}

\item{axis}{Options: \code{"both"} (default), \code{"x"}, \code{"y"}}
}
\value{
Raw data input variables used to generate x and/or y axis
}
\description{
Get variables used to generate x and y axes
}
