% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Extract plot data}
\usage{
get_plot_data(x, source = c("plot", "build", "user"))
}
\arguments{
\item{x}{ggplot object}

\item{source}{Source of data returned. Options: \code{"plot"} (\code{x$data}, default. Data entered to \code{ggplot} function), \code{"build"} (\code{ggplot_build(x)$data}. Data used internally by ggplot2 to construct each layer), \code{user} (user input data to each layer)}
}
\value{
Data frame or list of data frames
}
\description{
Extract plot data
}
