% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{get_plot_geom_data}
\alias{get_plot_geom_data}
\title{Extract data used in a specified geom}
\usage{
get_plot_geom_data(x, geom)
}
\arguments{
\item{x}{ggplot object}

\item{geom}{Name of geom (output from \code{get_plot_geom_names}). First letter only must be capitalized with lowercase for other letters}
}
\value{
Data frame or list of data frames
}
\description{
Extract data used in a specified geom
}
