% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_geom_names}
\alias{get_plot_geom_names}
\alias{get_plot_stat_names}
\title{Get names of geoms used by plot}
\usage{
get_plot_geom_names(x)

get_plot_stat_names(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
Geom names as character vector in lower case

Stat names as character vector in lower case
}
\description{
Get names of geoms used by plot

List stat functions used by a plot
}
