% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{get_plot_legend}
\alias{get_plot_legend}
\title{Get plot legend}
\usage{
get_plot_legend(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
Plot legend as ggplot object
}
\description{
Get plot legend
}
