% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_limits}
\alias{get_plot_limits}
\title{Overall limits for plot (data + axis)}
\usage{
get_plot_limits(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
List containing \code{x} and \code{y} (each length 2 numeric vectors representing min and max value). If plot has multiple facets, output will be a list (1 for each facet), each containing a list with \code{x} and \code{y} (each length 2 numeric vectors representing min and max value)
}
\description{
Overall limits for plot (data + axis)
}
