% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{get_plot_scale}
\alias{get_plot_scale}
\title{Extract axis scale}
\usage{
get_plot_scale(x, axis)
}
\arguments{
\item{x}{ggplot object}

\item{axis}{Axis to extract. Options: \code{"x"} or \code{"y"}}
}
\value{
Same class as same as scale_(x|y)_(discrete|continuous) output
}
\description{
Extract axis scale
}
