% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{get_plot_theme}
\alias{get_plot_theme}
\title{Extract theme from ggplot object}
\usage{
get_plot_theme(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
theme object
}
\description{
Extract theme from ggplot object
}
