% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_title}
\alias{get_plot_title}
\alias{get_plot_vars}
\title{Get plot title}
\usage{
get_plot_title(x)

get_plot_vars(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
Plot title as string

List of variables in data frames used to create plots. Also plot title
}
\description{
Functionality from gginnards
}
