% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{ggplot_add.continuous_axis}
\alias{ggplot_add.continuous_axis}
\title{Add continuous axis scale for ggplot}
\usage{
\method{ggplot_add}{continuous_axis}(object, plot, object_name)
}
\arguments{
\item{object}{Output from \code{scale_continuous}}

\item{plot}{ggplot object}

\item{object_name}{Name of object}
}
\value{
Called indirectly via by \code{scale_continuous()}
}
\description{
Must be exported to work correctly. Not called directly by user
}
