% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.new_color}
\alias{ggplot_add.new_color}
\title{Add new color scale}
\usage{
\method{ggplot_add}{new_color}(object, plot, object_name)
}
\arguments{
\item{object}{New scale to add}

\item{plot}{ggplot object}

\item{object_name}{Name of object}
}
\value{
Not called directly by user
}
\description{
Must be exported to work correctly
}
