% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{ggplot_add.x10_axis}
\alias{ggplot_add.x10_axis}
\title{Add simplified x10^n axis scale for ggplot}
\usage{
\method{ggplot_add}{x10_axis}(object, plot, object_name)
}
\arguments{
\item{object}{New scale to add}

\item{plot}{ggplot object}

\item{object_name}{Name of object}
}
\value{
Called when \code{axis_10x()} used
}
\description{
Must be exported to work correctly. Not called directly by user
}
