% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{Glm}
\alias{Glm}
\title{Modified version of glm}
\usage{
Glm(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  family = stats::binomial,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Enter in y ~ x format}

\item{y, x}{Variables. Enter as quoted variable names. \code{x} can be multiple variable names}

\item{family}{Function passed to family argument of glm. Default is binomial}

\item{na.rm}{If \code{TRUE} (default), \code{NA} will be removed}

\item{...}{Arguments passed to \code{glm}}
}
\value{
Output of glm
}
\description{
Functionality from stats package
}
