% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{glm_tidy}
\alias{glm_tidy}
\title{Tidy glm output}
\usage{
glm_tidy(x, ci = 0.95)
}
\arguments{
\item{x}{A \code{glm} object}

\item{ci}{Confidence interval. Default is \code{0.95}}
}
\value{
Data frame with columns for outcome, variable, predictor, or, or_lower, or_upper, p, type (Univariate or Multivariate), method (glm), covariates
}
\description{
Tidy glm output
}
