% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%W\%}
\alias{\%W\%}
\title{Waiver operator}
\usage{
x \%W\% y
}
\arguments{
\item{x}{Object possibly of class "waiver"}

\item{y}{Value to return if \code{x} is a waiver}
}
\value{
Output is \code{x} unless it is a "waiver" object in which case \code{y} is returned
}
\description{
Functionality from Hadley Wickham's excellent package ggplot2
}
