% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{has_ext}
\alias{has_ext}
\alias{str_has_ext}
\title{Determine whether a file path contains an extension}
\usage{
has_ext(x, ext = NULL, ignore_case = TRUE)

str_has_ext(x, ext = NULL, ignore_case = TRUE)
}
\arguments{
\item{x}{Character vector of file paths}

\item{ext}{Character vector of length 1 containing file extension with or without ".". If \code{NULL} (default), function will test whether \code{x} contains any file extension}

\item{ignore_case}{If \code{TRUE} (default), search for \code{ext} will be case insensitive}
}
\value{
Logical vector of length equal to input
}
\description{
Determine whether a file path contains an extension

Alias for has_ext
}
