% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_ci.R
\name{hr_ci}
\alias{hr_ci}
\alias{hr_ci.default}
\alias{hr_ci.coxph}
\alias{hr_ci.data.frame}
\alias{hr_ci.grouped_df}
\title{Hazard ratio, confidence interval, P values}
\usage{
hr_ci(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)

\method{hr_ci}{default}(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)

\method{hr_ci}{coxph}(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)

\method{hr_ci}{data.frame}(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)

\method{hr_ci}{grouped_df}(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)
}
\arguments{
\item{x}{Data frame, coxph object, survdiff object, or ggplot object}

\item{predictor_var}{Variable(s) to use as covariates in model. Enter as character vector}

\item{time_var}{Variable containing follow up time. Enter as quoted variable name. Default is \code{time}}

\item{outcome_var}{Variable containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as quoted variable name. Default is \code{death}}

\item{ci}{Confidence interval. Default is \code{0.95}}

\item{...}{Arguments passed to class-specific \code{hr_ci} function}
}
\value{
Data frame containing columns for predictor_var, predictor, hr, hr_lower, hr_upper, p_lrt (if possible), p_wald, n, n_events
}
\description{
Hazard ratio, confidence interval, P values

hr_ci - default method

hr_ci - coxph

hr_ci - data frame

hr_ci - grouped_dr
}
