% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{Intersect}
\alias{Intersect}
\title{Elements in x and y}
\usage{
Intersect(x, y)
}
\arguments{
\item{x, y}{Vectors}
}
\description{
Similar to \code{base::intersect} but doesn't convert \code{x} and \code{y} to vectors
}
