% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{intersect_all}
\alias{intersect_all}
\title{Elements common to all input vectors}
\usage{
intersect_all(...)
}
\arguments{
\item{...}{Vectors or list of vectors}
}
\description{
Equivalent to sequentially running \code{Intersect} on > 2 vectors
}
