% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{is_between}
\alias{is_between}
\title{Determine whether values are included in a specified range}
\usage{
is_between(x, min, max, include_min = TRUE, include_max = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{min, max}{Lower and upper bound of values respectively}

\item{include_min, include_max}{If \code{TRUE}, include values equal to min or max, respectively. Both \code{TRUE} by default}
}
\value{
Logical with length and order same as input. \code{NA} values in \code{x} will return \code{NA} in output
}
\description{
Determine whether values are included in a specified range
}
