% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{is_binary}
\alias{is_binary}
\alias{is_not_binary}
\alias{is_binary_01}
\title{Determine whether a vector is a binary variable}
\usage{
is_binary(x, na.rm = TRUE)

is_not_binary(x, na.rm = TRUE)

is_binary_01(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector}

\item{na.rm}{If \code{TRUE} (default), missing values are removed}
}
\value{
Logical vector of length 1
}
\description{
Determine whether a vector is a binary variable

Determine whether a vector is not a binary variable

Determine whether a vector consists exclusively of 0 and 1
}
