% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{is_categorical}
\alias{is_categorical}
\title{Determine whether a variable is likely a categorical variable}
\usage{
is_categorical(x, max_n_unique = 5L)
}
\arguments{
\item{x}{Vector}

\item{max_n_unique}{If number of \code{n_unique(x) <= max_n_unique}, variable is considered a factor}
}
\value{
If \code{x} is likely a categorical variable, \code{TRUE.} Otherwise, \code{FALSE}
}
\description{
Determine whether a variable is likely a categorical variable
}
