% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{is_constant}
\alias{is_constant}
\title{Determine if a variable is constant (all values identical)}
\usage{
is_constant(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector}

\item{na.rm}{If \code{TRUE} (default), missing values are removed}
}
\value{
Length 1 logical vector
}
\description{
Determine if a variable is constant (all values identical)
}
