% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{is_even}
\alias{is_even}
\alias{is_odd}
\title{Determine whether integer is even}
\usage{
is_even(x)

is_odd(x)
}
\arguments{
\item{x}{Integer vector}
}
\value{
Logical vector with length equal to input. If integer is even, output is \code{TRUE.} If integer is odd, output is \code{FALSE}

Logical vector with length equal to input. If integer is odd, output is \code{TRUE.} If integer is even, output is \code{FALSE}
}
\description{
Determine whether integer is even

Determine whether integer is odd
}
