% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{is_input_piped}
\alias{is_input_piped}
\title{Determine whether input was piped}
\usage{
is_input_piped(n = 1)
}
\arguments{
\item{n}{Number of layers deep in call stack. Default is \code{1}. Enter as numeric}
}
\value{
Call \code{is_input_piped()} inside function to determine whether input to function was piped
}
\description{
Determine whether input was piped
}
