% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{is_normal}
\alias{is_normal}
\title{Determine whether values in 2 groups are both normally distributed}
\usage{
is_normal(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  normality_test = p_shapiro,
  ...
)
}
\arguments{
\item{df}{Data frame in long format}

\item{formula}{enter as continuous variable ~ grouping variable}

\item{x, y}{Variables for grouping variable (\code{x}) and continuous variable (\code{y}). Enter as quoted or unquoted variable names}

\item{normality_test}{Normality test used to determine whether t-test or Mann Whitney U test will be used. Enter as unquoted function name. Function must take a numeric vector as input and returns p value. Default is \code{p_shapiro}}

\item{...}{Subgrouping variables (statistical test will be performed for each unique combination of subgrouping variables). Enter as comma separated list of quoted or unquoted variable names. If quoted, can wrap in \code{c()}. Can also enter character vector containing variable names}
}
\value{
\code{TRUE} if normally distributed (all P >= 0.05), \code{FALSE} if not normally distributed (any P < 0.05)
}
\description{
Determine whether values in 2 groups are both normally distributed
}
