% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{is_POSIXct}
\alias{is_POSIXct}
\alias{is_date}
\title{Determine whether the class of a vector is POSIXct}
\usage{
is_POSIXct(x)

is_date(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Logical
}
\description{
Determine whether the class of a vector is POSIXct

Check whether vector contains dates
}
