% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_join.R
\name{join_many}
\alias{join_many}
\title{Join > 2 data frames}
\usage{
join_many(..., join_fn = dplyr::left_join, by = NULL)
}
\arguments{
\item{...}{Enter data frames as comma separated list}

\item{join_fn}{Joing function. Unless \code{full_join} is used, order of data frames in \code{...} influences the resulting data frame. Default uses \code{left_join.} Enter as quoted or unquoted function name}

\item{by}{Variable shared by data frames used to join all data frames. Must be the same in all data frames. Enter as quoted variable name}
}
\value{
Single data frame a result of joining all data frames
}
\description{
Join > 2 data frames
}
