% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{label_genotype}
\alias{label_genotype}
\title{Create genotype labels using plotmath notation}
\usage{
label_genotype(
  genes,
  superscripts,
  font_face_genes = "italic",
  font_face_superscripts = "italic"
)
}
\arguments{
\item{genes, superscripts}{Genes and superscripts respectively. Enter as character vector}

\item{font_face_genes, font_face_superscripts}{Fontface for genes and superscripts respectively. Options: \code{"italic" }(default), \code{"plain"}, \code{"bold"}, \code{"bolditalic"}. Enter as quoted type}
}
\value{
Expression that can be used for ggplot2
}
\description{
Create genotype labels using plotmath notation
}
