% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_info.R
\name{lib}
\alias{lib}
\title{Silently load 1 or more packages}
\usage{
lib(...)
}
\arguments{
\item{...}{Enter comma separated list of unquoted or quoted package names without \code{c()}}
}
\description{
Silently load 1 or more packages
}
