% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{list_dirs}
\alias{list_dirs}
\title{Search for folders}
\usage{
list_dirs(
  ...,
  pattern = NULL,
  pattern_exclude = NULL,
  all_terms = TRUE,
  ignore_case = TRUE,
  recursive = TRUE,
  exact_match = FALSE,
  start_path = desktop_path(include_slash = FALSE),
  incl_hidden = TRUE
)
}
\arguments{
\item{...}{Arguments passed to \code{paste_path()}. Enter comma separated list of quoted or unquoted folder names following a starting path not wrapped in \code{c()}. Starting path is specified using start argument. Default uses desktop path as start argument}

\item{pattern}{Pattern in folder name to search}

\item{pattern_exclude}{Pattern in folder name to exclude from search}

\item{all_terms}{If \code{TRUE} (default) and multiple patterns entered, result is folders containing all patterns}

\item{ignore_case}{If \code{FALSE} (default), case of \code{pattern} is ignored in search. If \code{TRUE}, search for \code{pattern} is case sensitive}

\item{recursive}{If \code{TRUE} (default), all subfolders are searched}

\item{exact_match}{If \code{FALSE} (default), entry for \code{...} doesn't need to be exact match of folder name}

\item{start_path}{Directory used as prefix to the path built using dots}

\item{incl_hidden}{If \code{TRUE} (default), hidden folders included in output}
}
\value{
Character vector of folder paths
}
\description{
Search for folders
}
