% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{list_files}
\alias{list_files}
\title{List files in folders}
\usage{
list_files(
  ...,
  pattern = NULL,
  ext = NULL,
  pattern_exclude = NULL,
  all_terms = TRUE,
  ignore_case = TRUE,
  exact_match = FALSE,
  recursive = TRUE,
  start_path = desktop_path(include_slash = FALSE)
)
}
\arguments{
\item{...}{Arguments passed to \code{paste_path()}. Enter comma separated list of quoted or unquoted folder names following a starting path not wrapped in \code{c()}. Starting path is specified using start argument. Default uses desktop path as start argument. Do not include file name in \code{...}}

\item{pattern}{Pattern in file name to search}

\item{ext}{File extension. Enter as quoted file extension with or without "."}

\item{pattern_exclude}{Pattern in file name to exclude from search}

\item{all_terms}{If \code{TRUE} (default) and multiple patterns entered, result is files containing all patterns}

\item{ignore_case}{If \code{FALSE} (default), case of \code{pattern} is ignored in search. If \code{TRUE}, search for \code{pattern} is case sensitive}

\item{exact_match}{If \code{FALSE} (default), entry for \code{...} doesn't need to be exact match of file name}

\item{recursive}{If \code{FALSE} (default), only files in specified path returned}

\item{start_path}{Directory used as prefix to the path built using dots}
}
\value{
Character vector of file paths
}
\description{
List files in folders
}
