% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_coerce.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert a list to a data frame}
\usage{
list_to_df(x)
}
\arguments{
\item{x}{List of vectors. All elements of list (each vector) must have same length or must share a common multiple (see examples for details). To convert a list of vectors to rows of a data frame, use \code{list_to_df_rowwise()}}
}
\value{
Data frame with elements of input list (vectors) forming columns of output data frame
}
\description{
Opposite of \code{as.list} applied to a data frame
}
