% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_coerce.R
\name{list_to_df_rowwise}
\alias{list_to_df_rowwise}
\title{Convert a list of vectors to rows of a data frame}
\usage{
list_to_df_rowwise(
  ...,
  rownames_to_col = FALSE,
  colname_rownames = "rowname",
  prefix = "V"
)
}
\arguments{
\item{...}{Vectors or list of vectors used to create rows in output data frame. If the lengths of list components are not equal, shorter elements will be recycled to the length of the longest element (i.e. for a single row, values will be recycled across the leftover columns)}

\item{rownames_to_col}{If \code{TRUE}, rownames of \code{x} are used to create a new column in output data frame. Default is \code{FALSE}}

\item{colname_rownames}{Name of column in output data frame containing rownames of \code{x}. Only relevant when \code{rownames_to_col = TRUE}. Default is \code{"rowname"}}

\item{prefix}{Prefix for newly created column names. Default is \code{"V"}}
}
\value{
Data frame with elements of input combined to form rows of data frame
}
\description{
Similar to \code{Reduce(rbind, list)}, but faster
}
