% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{Lm}
\alias{Lm}
\title{Modified version of lm}
\usage{
Lm(df, formula = NULL, y = NULL, x = NULL, na.rm = TRUE)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Formula}

\item{y, x}{Variables. Enter as quoted variable names. \code{x} can be multiple variable names}

\item{na.rm}{If \code{TRUE} (default), \code{NA} will be removed}
}
\description{
Functionality from stats package
}
