% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm_reg_line_equation}
\alias{lm_reg_line_equation}
\title{Create regression formula}
\usage{
lm_reg_line_equation(x, digits = 2, bracket_type = "{")
}
\arguments{
\item{x}{An \code{lm} object}

\item{digits}{Number of digits after decimal to include in formula. Default is \code{2}}

\item{bracket_type}{Type of bracket to surround variables in model}
}
\value{
Equation as length 1 character vector
}
\description{
Create regression formula
}
