% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{log10_1}
\alias{log10_1}
\title{Calculate log10 + 1}
\usage{
log10_1(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector with length equal to input
}
\description{
Calculate log10 + 1
}
