% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_info.R
\name{match_fun}
\alias{match_fun}
\title{Extract function including}
\usage{
match_fun(fn)
}
\arguments{
\item{fn}{Function. Enter function definition, quoted or unquoted function name}
}
\value{
Function
}
\description{
Slightly faster version of \code{match.fun}
}
