% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_pdf_png_jpg.R
\name{merge_pdf}
\alias{merge_pdf}
\title{Merge multiple pdf files}
\usage{
merge_pdf(..., new_file_name = "combined", directory = NULL)
}
\arguments{
\item{...}{Complete paths of pdf files to merge. Enter as comma separated list of quoted file paths}

\item{new_file_name}{File name for new pdf. Enter as length 1 character vector}

\item{directory}{Directory where new file should be saved. Default uses directory of 1st file entered into \code{...}}
}
\value{
New pdf file silently generated
}
\description{
Merge multiple pdf files
}
