% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{mgsub}
\alias{mgsub}
\alias{msub}
\title{Substitute multiple characters over multiple patterns}
\usage{
mgsub(x, pattern, replacement, ...)

msub(x, pattern, replacement, ...)
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Character vector containing patterns to be replaced}

\item{replacement}{Character vector containing replacement patterns. Must have length of 1 or \code{length(pattern)}}

\item{...}{Arguments passed to \code{gsub}}
}
\value{
Character vector with length same as input
}
\description{
Substitute multiple characters over multiple patterns

Substitute multiple characters over multiple patterns (1st matching pattern only)
}
