% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{Mode}
\alias{Mode}
\title{Mode of a vector}
\usage{
Mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector}

\item{na.rm}{If \code{TRUE} (default), missing values are removed prior to determining mode}
}
\value{
Value(s) in vector that occur most frequently. Class matches input class
}
\description{
Mode of a vector
}
