% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{move_cols}
\alias{move_cols}
\title{Change column order}
\usage{
move_cols(df, ..., before = NULL, after = NULL)
}
\arguments{
\item{df}{Data frame}

\item{...}{Columns to move. Enter using tidyselect syntax}

\item{before, after}{Where to place columns selected in \code{...} Enter as quoted or unquoted column name. Use before to place \code{...} columns before columns selected by before. Use after to place \code{...} columns after columns selected by after. If both are \code{NULL}, \code{...} columns are placed before all other columns}
}
\description{
Change column order
}
