% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi.R
\name{multislide_plot_list}
\alias{multislide_plot_list}
\title{List of multipanel plots for printing}
\usage{
multislide_plot_list(
  plot_list,
  plots_per_slide = 4,
  rows = 2,
  cols = plots_per_slide/rows
)
}
\arguments{
\item{plot_list}{List of ggplot objects}

\item{plots_per_slide}{Number of plots per slide. Default is \code{4}}

\item{rows}{Number of rows of plots per slide. Default is \code{2}}

\item{cols}{Number of columns of plots per slide. Default is \code{2}}
}
\value{
List of multipanel plots that can be plotted using ppt_plot function for printing
}
\description{
List of multipanel plots for printing
}
