% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{N}
\alias{N}
\alias{N_na}
\alias{Perc}
\alias{Perc_na}
\title{Number of non-missing values}
\usage{
N(x)

N_na(x)

Perc(x)

Perc_na(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Number of non-missing values

Number of missing values

Percent of values that are true or 1

Percent of values that are missing
}
