% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{n_combos}
\alias{n_combos}
\title{Number of combinations}
\usage{
n_combos(n, k, replacement = TRUE)
}
\arguments{
\item{n}{Number of options}

\item{k}{Number of items selected (note: \code{n} must be >= \code{k})}

\item{replacement}{If \code{TRUE} (default), combinations with replacement is calculated. If \code{FALSE}, combinations without replacement is calculated. Example of replacement: if creating all combinations of 2 letters, "AA" is allowed (i.e. "A" can be selected twice)}
}
\value{
Number of possible combinations
}
\description{
Order does not matter (i.e. a, b is the same as b, a)
}
