% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{n_cores}
\alias{n_cores}
\title{Determine number of CPU cores}
\usage{
n_cores()
}
\value{
Length 1 integer containing number of cores available (including R)
}
\description{
Functionality from detectCores function in parallel package
}
