% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{n_per_group}
\alias{n_per_group}
\title{Counts for each unique value}
\usage{
n_per_group(x, ...)
}
\arguments{
\item{x}{Vector or data frame}

\item{...}{Columns in \code{x}. Enter using tidyselect syntax. Ignored if \code{x} is a vector}
}
\value{
If input is a vector, out is a named integer vector containing number of observations per level for a categorical variable. Name of values refers to the level. Listed in decreasing order of group size with NAs listed last. If input is a data frame, output is a data frame containing columns for variable, group (unique value for variable), Freq (number of rows in \code{x} in which variable = group). Similar to output from \code{dplyr::count}
}
\description{
Counts for each unique value
}
