% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{n_permutations}
\alias{n_permutations}
\title{Number of permutations}
\usage{
n_permutations(n, k = n, replacement = FALSE)
}
\arguments{
\item{n}{Number of options}

\item{k}{Number of items selected (note: \code{n} >= \code{k})}

\item{replacement}{If \code{TRUE} (default), combinations with replacement is calculated. If \code{FALSE}, combinations without replacement is calculated. Example of replacement: if creating all combinations of 2 letters, "AA" is allowed (i.e. "A" can be selected twice)}
}
\value{
Number of permutations
}
\description{
Order matters
}
