% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{n_unique}
\alias{n_unique}
\title{Determine number of unique values for vector}
\usage{
n_unique(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector}

\item{na.rm}{If \code{TRUE} (default), missing values removed}
}
\value{
Length 1 integer vector
}
\description{
Determine number of unique values for vector
}
