% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{n_unique_per_var}
\alias{n_unique_per_var}
\title{Number of unique values for variables of data frame}
\usage{
n_unique_per_var(df, na.rm = TRUE)
}
\arguments{
\item{df}{Data frame}

\item{na.rm}{If \code{TRUE} (default), missing values are not considered a unique value}
}
\value{
Named integer vector of unique values for each variable
}
\description{
Number of unique values for variables of data frame
}
