% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{name_contains}
\alias{name_contains}
\title{Determine whether names contain a specified pattern}
\usage{
name_contains(x, pattern, invert = FALSE, ignore_case = FALSE, ...)
}
\arguments{
\item{x}{Data frame, matrix, list, or named vector}

\item{pattern}{Pattern in \code{names(x)} to search}

\item{invert}{If \code{TRUE}, search will be performed for names of \code{x} that do not contain \code{pattern}. Equivalent of \code{!name_contains(x)}}

\item{ignore_case}{If \code{FALSE} (default), case of both \code{names(x)} and \code{pattern} are ignored in search. If \code{TRUE}, case used in \code{names(x)} and \code{pattern} must match}

\item{...}{Arguments passed to \code{str_contains}}
}
\value{
Logical vector with length equal to \code{length(x)} (or \code{ncol(x)} if \code{x} is a matrix)
}
\description{
Equivalent to \code{grepl(pattern = pattern, x = names(x))}
}
