% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{normalize}
\alias{normalize}
\title{Normalize numeric vector}
\usage{
normalize(x, desired_range = NULL)
}
\arguments{
\item{x}{Numeric vector}

\item{desired_range}{Desired range for output values. If \code{NULL}, normalize to z-score. Enter as length 2 numeric vector containing desired minimum and maximum values}
}
\value{
Numeric vector with values transformed to make the range equal to \code{desired_range}
}
\description{
Normalize numeric vector
}
