% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{normalize_by_row}
\alias{normalize_by_row}
\title{Normalize subset of columns by row to specified range}
\usage{
normalize_by_row(df, ..., desired_range = NULL)
}
\arguments{
\item{df}{Data frame}

\item{...}{Columns to include. Enter using tidyselect syntax}

\item{desired_range}{Desired range for normalized values. If \code{NULL}, normalize to z-score. Enter as length 2 numeric vector containing desired minimum and maximum values}
}
\value{
Data frame with row-wise normalized values
}
\description{
Normalize subset of columns by row to specified range
}
