% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{normalize_cols}
\alias{normalize_cols}
\title{Normalize a subset of variables in data frame}
\usage{
normalize_cols(df, ..., suffix = "_norm", norm_fn = normalize)
}
\arguments{
\item{df}{Data frame}

\item{...}{Variables to normalize. Enter using tidyselect syntax}

\item{suffix}{Suffix for normalized variable names. Default is \code{"_norm"}}

\item{norm_fn}{Function used to normalize data. Default is \code{normalize()}}
}
\value{
Data frame with additional columns for normalized version of selected variables
}
\description{
Normalize a subset of variables in data frame
}
