% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{nullop}
\alias{nullop}
\alias{\%||\%}
\title{Null operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Object for which \code{NULL} status will be determined}

\item{y}{Result if x is \code{NULL}}
}
\description{
Null operator
}
