% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{odds_ratio}
\alias{odds_ratio}
\alias{odds_ratio.default}
\alias{odds_ratio.matrix}
\alias{odds_ratio.table}
\alias{odds_ratio.data.frame}
\alias{odds_ratio.lm}
\alias{odds_ratio.glm}
\title{Calculate OR, 95\% CI, and P value}
\usage{
odds_ratio(
  x,
  ...,
  outcome_var = NULL,
  predictor_var = NULL,
  formula = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5
)

\method{odds_ratio}{default}(
  x,
  ...,
  outcome_var = NULL,
  predictor_var = NULL,
  formula = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5
)

\method{odds_ratio}{matrix}(
  x,
  ...,
  outcome_var = NULL,
  predictor_var = NULL,
  formula = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5
)

\method{odds_ratio}{table}(
  x,
  ...,
  outcome_var = NULL,
  predictor_var = NULL,
  formula = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5
)

\method{odds_ratio}{data.frame}(
  x,
  ...,
  outcome_var = NULL,
  predictor_var = NULL,
  formula = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5
)

\method{odds_ratio}{lm}(
  x,
  ...,
  outcome_var = NULL,
  predictor_var = NULL,
  formula = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5
)

\method{odds_ratio}{glm}(
  x,
  ...,
  outcome_var = NULL,
  predictor_var = NULL,
  formula = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5
)
}
\arguments{
\item{x}{matrix, table, integer vector of counts (only used for 2 x 2 matrix. Must enter in following order: tn, fn, fp, tp), data frame (with predictor/outcome variables specified individually or in a formula), glm or lm object. If x is a vector of counts, enter as test-/control, test-/case, test+/control, test+/case. Also predictor-/outcome-, predictor-/outcome+, predictor+/outcome-, predictor+/outcome+. Entered into matrix by column. For matrix input, columns = predictor/test result/risk factor/case vs. control (- or control in col 1, + or case in col 2), rows = control truth/outcome (outcome- in row 1, outcome+ in row 2)}

\item{...}{Arguments passed to class-specific \code{odds_ratio} function}

\item{outcome_var}{Outcome variable. Enter as quoted or unquoted variable name}

\item{predictor_var}{Predictor variable. Enter as quoted or unquoted variable name}

\item{formula}{Formula as y ~ x}

\item{cp, cutpoint}{Cutpoint. Will be ignored if entered but \code{predictor_var} is categorical}

\item{type}{If \code{>=} (default), positivity is determined by values greater than or equal to cutpoint.  If \code{>}, positivity is determined by values of \code{predictor_var} greater than cutpoint}

\item{ci}{Confidence interval. Default is \code{0.95}}

\item{yates}{If \code{TRUE}, Yates' correction is applied to estimation of chi-squared statistic}

\item{n_min_chisq}{Minimum group size that will be allowed for chi-squared test. Default is \code{5}}
}
\value{
Data frame with columns for or, or_lower, or_upper, p, label, method, p_fisher, p_chi

Data frame containing the following columns: outcome_var, predictor_var, type, cutpoint, or, or_lower, or_upper, p, label, method, p_fisher, p_chi, n, outcome, perc_outcome, no_outcome, perc_no_outcome, pos, perc_pos, neg, perc_neg, tn, fn, fp, tp, def_outcome, def_no_outcome, def_predictor_pos, def_predictor_neg
}
\description{
Calculate OR, 95\% CI, and P value

odds_ratio - default method

odds_ratio - matrix

odds_ratio - table

odds_ratio - data frame

odds_ratio - lm

odds_ratio - glm
}
