% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{only_decreasing}
\alias{only_decreasing}
\alias{any_decreasing}
\title{Determine whether a numeric vector is only decreasing (decreases monotonically)}
\usage{
only_decreasing(x, incl_plateau = TRUE, na.rm = TRUE)

any_decreasing(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{incl_plateau}{If \code{TRUE} (default), output can still be \code{TRUE} even if \code{x} contains instances in which adjacent values. If \code{FALSE}, values must be strictly monotonically decreasing}

\item{na.rm}{If \code{TRUE} (default), missing values are removed}
}
\value{
Logical. \code{TRUE} if each successive value is less than the previous value. If \code{incl_plateau} is \code{TRUE}, output can still be \code{TRUE} even if \code{x} contains instances in which adjacent values

Length 1 logical. \code{TRUE} if any occurrences of a value less than the previous value
}
\description{
Determine whether a numeric vector is only decreasing (decreases monotonically)

Determine whether a numeric vector contains any decreases
}
