% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{only_increasing}
\alias{only_increasing}
\alias{any_increasing}
\title{Determine whether a numeric vector is only increasing (increases monotonically)}
\usage{
only_increasing(x, incl_plateau = TRUE, na.rm = TRUE)

any_increasing(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{incl_plateau}{If \code{TRUE} (default), output can still be \code{TRUE} even if \code{x} contains instances in which adjacent values. If \code{FALSE}, values must be strictly monotonically increasing}

\item{na.rm}{If \code{TRUE} (default), missing values are removed}
}
\value{
Length 1 logical. \code{TRUE} if each successive value is greater than the previous value. If \code{incl_plateau} is \code{TRUE}, output can still be \code{TRUE} even if \code{x} contains instances in which adjacent values

Length 1 logical. \code{TRUE} if any occurrences of a value greater than the previous value
}
\description{
Determine whether a numeric vector is only increasing (increases monotonically)

Determine whether a numeric vector contains any increases
}
