% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{open_file}
\alias{open_file}
\title{Open file}
\usage{
open_file(
  file_name,
  ...,
  ext = "R",
  desktop = TRUE,
  recursive = TRUE,
  ignore_case = TRUE
)
}
\arguments{
\item{file_name}{Either file path or name of file with or without file extension. Input can be quoted or unquoted}

\item{...}{Comma separated list without of folders (in order) to open in order to arrive at file location not wrapped in \code{c()}. Folder names can be entered as quoted or unquoted if there are no spaces. If there are spaces, must quote file name}

\item{ext}{File extension with or without ".". Default is \code{"R"}}

\item{desktop}{If \code{TRUE} (default), files can be located on desktop}

\item{recursive}{If \code{FALSE} (default), only files in specified path returned}

\item{ignore_case}{If \code{TRUE} (default), case of file name is ignored}
}
\description{
Open file
}
