% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{open_folder}
\alias{open_folder}
\title{Open finder window for folder}
\usage{
open_folder(..., start = desktop_path(), silent = FALSE)
}
\arguments{
\item{...}{Either a vector (or comma separated list) of full folder paths or path created by passing arguments to \code{paste_path}. Enter comma separated list of quoted or unquoted folder names following a starting path not wrapped in \code{c()}. Starting path is specified using start argument. Default uses desktop path as start argument}

\item{start}{Beginning of file path up to desktop}

\item{silent}{If \code{TRUE}, messages will be suppressed}
}
\description{
Open finder window for folder
}
