% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{open_wd}
\alias{open_wd}
\title{Open working directory folder or R file located in working directory}
\usage{
open_wd(file_name = NULL, ignore_case = TRUE, recursive = TRUE)
}
\arguments{
\item{file_name}{File name. Enter as quoted or unquoted file name, with or without extension. If empty, working directory folder will be opened}

\item{ignore_case}{If \code{TRUE} (default), case of \code{file_name} is ignored}

\item{recursive}{If \code{TRUE} (default), subfolders in working directory are searched}
}
\description{
Open working directory folder or R file located in working directory
}
