% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{p_adjust}
\alias{p_adjust}
\alias{p_adjust.default}
\alias{p_adjust.data.frame}
\title{Generic adjust p value function}
\usage{
p_adjust(x, p_column = "p", p_adj_name = "p_adj", method = "holm", ...)

\method{p_adjust}{default}(x, p_column = "p", p_adj_name = "p_adj", method = "holm", ...)

\method{p_adjust}{data.frame}(x, p_column = "p", p_adj_name = "p_adj", method = "holm", ...)
}
\arguments{
\item{x}{Data frame or numeric vector containing p values}

\item{p_column}{Column containing p values. Enter as quoted column name}

\item{p_adj_name}{Column name for adjusted p values. Default is \code{"p_adj"}}

\item{method}{Method for p value correction. Options: \code{"holm"} (default), \code{"BH"}, \code{"BY"}, \code{"hochberg"}, \code{"hommel"}, \code{"bonferroni"}, \code{"fdr"} (same as \verb{"BH")}, \code{"none"}. Enter as quoted method. Can be any case}

\item{...}{Arguments passed to class-specific \code{p_adjust} function}
}
\description{
Simplified version of \code{stats::p.adjust} for broader range of input types
}
