% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_means.R
\name{p_anova}
\alias{p_anova}
\alias{p_kruskal}
\alias{p_tukey}
\title{One-way ANOVA with or without Welch's correction}
\usage{
p_anova(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  na.rm = TRUE,
  welch = NULL,
  variance_test = p_levene,
  otherwise = NA_real_,
  ...
)

p_kruskal(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  na.rm = TRUE,
  otherwise = NA_real_,
  ...
)

p_tukey(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  na.rm = TRUE,
  variance_test = p_F_test,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Entered as continuous variable ~ grouping variable}

\item{y}{Continuous variable. Enter as quoted variable name}

\item{x}{Continuous variable. Enter as quoted variable name}

\item{na.rm}{If \code{TRUE} (default), missing values for grouping variable are not considered a distinct group}

\item{welch}{If \code{NULL} (default), variance testing is used to determine whether Welch's correction should be applied. If \code{TRUE}, Welch's correction performed. If \code{FALSE}, ordinary ANOVA performed}

\item{variance_test}{Test used to determine whether variance differs by group. Default is Levene test}

\item{otherwise}{Output if ANOVA fails. Default is \code{NA}}

\item{...}{Not used}
}
\value{
P value as length 1 numeric vector

Data frame with comparisons similar to compare_means and p_dunn output
}
\description{
Functionality from oneway.test function in stats package
}
