% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_prop.R
\name{p_cochrane_armitage}
\alias{p_cochrane_armitage}
\title{Cochrane Armitage test for trend}
\usage{
p_cochrane_armitage(
  x,
  hypothesis_type = c("two.sided", "increasing", "decreasing")
)
}
\arguments{
\item{x}{Table containing 2 columns (1 for each group). Rows contain values at each time point}

\item{hypothesis_type}{Type of hypothesis test. Options: \code{"two.sided"} (default), \code{"increasing"}, \code{"decreasing"}}
}
\value{
P value as length 1 numeric vector
}
\description{
Functionality from Andri Signorell's excellent package DescTools
}
