% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_correlation.R
\name{p_cor}
\alias{p_cor}
\title{P value for correlation test}
\usage{
p_cor(
  df,
  formula = NULL,
  x = NULL,
  y = NULL,
  method = c("spearman", "pearson", "kendall"),
  otherwise = NA_real_
)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Formula in format y ~ x (same as x ~ y)}

\item{x, y}{Column names containing data to correlate. Enter as quoted variable names}

\item{method}{Method for correlation. Options: \code{"spearman"} (default), \code{"pearson"}, \code{"kendall"}}

\item{otherwise}{Value to return if unable to perform correlation test}
}
\description{
P value for correlation test
}
