% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_means.R
\name{p_dunn}
\alias{p_dunn}
\title{Dunn test}
\usage{
p_dunn(
  df,
  formula,
  ...,
  x = NULL,
  y = NULL,
  p_adj_method = "BH",
  ns_symbol = "ns",
  summary_fns = list(median = Median, mean = Mean, q1 = Q1, q3 = Q3, sd = SD, se = SE, ci
    = CI, min = min, max = max)
)
}
\arguments{
\item{df}{Data frame in long format}

\item{formula}{enter as continuous variable ~ grouping variable}

\item{...}{Subgrouping variables (statistical test will be performed for each unique combination of subgrouping variables). Enter as comma separated list of quoted or unquoted variable names. If quoted, can wrap in \code{c()}. Can also enter character vector containing variable names}

\item{x, y}{Variables for grouping variable (\code{x}) and continuous variable (\code{y}). Enter as quoted or unquoted variable names}

\item{p_adj_method}{Method for p value adjustment. Options: \code{BH} (default), \code{Holm}, \code{hochberg}, \code{hommell}, \code{bonferroni}, \code{BY}, \code{fdr}, \code{none.} Enter as quoted method}

\item{ns_symbol}{Symbol to use when P > 0.05. Default is \code{"ns"}}

\item{summary_fns}{Named list of summary functions that will be applied to each group. Enter as \code{list(col_name = fn_name)}}
}
\value{
Data frame with comparisons similar to compare_means output
}
\description{
Dunn test
}
