% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_variance.R
\name{p_F_test}
\alias{p_F_test}
\title{F test (aka var.test)}
\usage{
p_F_test(df, formula = NULL, y = NULL, x = NULL, otherwise = 0)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Formula entered as continuous variable ~ grouping variable}

\item{y}{Continuous variable. Enter as quoted variable name}

\item{x}{Categorical variable. Enter as quoted variable name}

\item{otherwise}{Output if F test fails. Default is \code{0}}
}
\value{
P value as length 1 numeric vector
}
\description{
F test (aka var.test)
}
