% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_outlier.R
\name{p_grubbs}
\alias{p_grubbs}
\title{Grubbs test}
\usage{
p_grubbs(
  x,
  n_outliers = 2,
  hypothesis_type = "one.sided",
  test_min = NULL,
  ...
)
}
\arguments{
\item{x}{Numeric vector}

\item{n_outliers}{Number of outliers to search for. Options: \code{1} (tests whether min or max value in x is an outlier), \code{2} (tests whether min or max values are outliers)}

\item{hypothesis_type}{Type of hypothesis test to perform. If \code{"one.sided"} (default), 1 sided test is performed. If \code{"two.sided"}, 2 sided test is performed. Enter as quoted hypothesis test type}

\item{test_min}{If \code{NULL} (default), test determines whether min or max value (whichever is further from mean value) is an outlier. If \code{TRUE}, outlier test performed on min value. If \code{FALSE}, outlier test performed on max value}

\item{...}{Not used}
}
\description{
Determine whether the minimum or maximum value is an outlier
}
