% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_variance.R
\name{p_levene}
\alias{p_levene}
\alias{p_bartlett}
\title{P value for Levene test}
\usage{
p_levene(
  df,
  formula = NULL,
  x = NULL,
  y = NULL,
  na.rm = TRUE,
  brown_forsythe = TRUE,
  otherwise = 0,
  ...
)

p_bartlett(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  na.rm = TRUE,
  otherwise = 0,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Formula entered as continuous variable ~ grouping variable}

\item{x}{Continuous variable. Enter as quoted variable name}

\item{y}{Categorical variable. Enter as quoted variable name}

\item{na.rm}{If \code{TRUE} (default), observations with missing values for the grouping variable are removed prior to analysis}

\item{brown_forsythe}{If \code{TRUE} (default), Brown Forsythe version of Levene test performed (using median). If \code{FALSE}, original Levene test (using mean) performed}

\item{otherwise}{Output if F test fails. Default is \code{0}}

\item{...}{Not used}
}
\value{
P value as length 1 numeric vector. If P < 0.05, variance is not same in each group
}
\description{
P value for Levene test

Bartlett test
}
