% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_prop.R
\name{p_prop_test_trend}
\alias{p_prop_test_trend}
\title{Chi-squared test for trend in proportions}
\usage{
p_prop_test_trend(n_positive, n_total)
}
\arguments{
\item{n_positive, n_total}{Integer vectors containing number of positive observations and total number of observations at each time point respectively}
}
\value{
P value as length 1 numeric vector
}
\description{
Functionality from stats package
}
