% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{p_surv}
\alias{p_surv}
\title{P value for log rank test}
\usage{
p_surv(x)
}
\arguments{
\item{x}{coph or survdiff object}
}
\value{
P value for log-rank test as length 1 numeric vector
}
\description{
P value for log rank test
}
