% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_means.R
\name{p_ttest}
\alias{p_ttest}
\alias{p_mann_whitney}
\alias{p_wilcox}
\alias{p_by_normality}
\title{P value for t-test}
\usage{
p_ttest(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  id = NULL,
  welch = NULL,
  variance_test = p_F_test,
  paired = FALSE,
  hypothesis_type = "two.sided",
  otherwise = NA_real_,
  ...
)

p_mann_whitney(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  id = NULL,
  paired = FALSE,
  otherwise = NA_real_,
  hypothesis_type = "two.sided",
  ...
)

p_wilcox(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  id = NULL,
  paired = FALSE,
  otherwise = NA_real_,
  hypothesis_type = "two.sided",
  ...
)

p_by_normality(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  id = NULL,
  paired = FALSE,
  welch = NULL,
  normality_test = p_shapiro,
  variance_test = p_F_test,
  hypothesis_type = "two.sided",
  otherwise = NA_real_,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Entered in format of continuous variable ~ grouping variable}

\item{y}{Continuous variable. Enter as quoted variable name}

\item{x}{Categorical grouping variable. Enter as quoted variable name}

\item{id}{Identification variable to link paired observations. Enter as quoted variable name. Only relevant when paired is \code{TRUE}}

\item{welch}{If \code{NULL} (default), use of Welch's correction is determined by variance test (If P < 0.05, variance is not equal between groups and Welch's correction is applied). If \code{TRUE}, Welch's correction is applied, otherwise student's t-test is performed}

\item{variance_test}{Function used to perform variance test. Must take numeric vector and returns P value. Default is \code{p_F_test}}

\item{paired}{If \code{FALSE} (default), unpaired test is performed}

\item{hypothesis_type}{Type of hypothesis test to perform. Options: \code{two.sided} (default), \code{less}, \code{greater.} Enter as quoted hypothesis test type}

\item{otherwise}{Output if t-test fails. Default is \code{NA}}

\item{...}{Not used}

\item{normality_test}{Function used to determine normality. Must take numeric vector and returns P value. Default is \code{p_shapiro}}
}
\description{
P value for t-test

Mann-Whitney U test (Wilcoxon rank sum test)

Alias for p_mann_whitney

P value for comparison using test as determined by application of normality and variance tests
}
