% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{pairs}
\alias{pairs}
\title{Pairwise combinations of elements from 1 vector}
\usage{
pairs(x, col_names = c("V1", "V2"), na.rm = TRUE)
}
\arguments{
\item{x}{Vector with elements to be combined pairwise}

\item{col_names}{Character vector containing names of columns to be generated}

\item{na.rm}{If \code{TRUE} (default), missing values are removed from \code{x} prior to generating combinations}
}
\value{
Data frame with 2 columns that contains unique pairs of elements in each row
}
\description{
Wrapper around combos with n = 2
}
