% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{pairs_list}
\alias{pairs_list}
\title{Extract from vector a list of length 2 vectors containing pairwise combinations}
\usage{
pairs_list(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector with elements to be combined pairwise}

\item{na.rm}{If \code{TRUE} (default), missing values are removed from x prior to generating combinations}
}
\value{
List of length 2 vectors
}
\description{
Extract from vector a list of length 2 vectors containing pairwise combinations
}
