% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{paste_path}
\alias{paste_path}
\title{Build a file path by successive folder names}
\usage{
paste_path(..., start = desktop_path())
}
\arguments{
\item{...}{Either path to directory or comma separated list of quoted or unquoted folder names following desktop not wrapped in \code{c()}}

\item{start}{Beginning of file path up to desktop}
}
\value{
File path without terminal "/"
}
\description{
Build a file path by successive folder names
}
