% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_quoted}
\alias{paste_quoted}
\title{Convert unquoted terms not wrapped in c() to comma separated list of quoted terms wrapped in c()}
\usage{
paste_quoted(..., with_c = TRUE, copy = TRUE)
}
\arguments{
\item{...}{Input to be quoted. If empty, clipboard contents will be used}

\item{with_c}{If \code{TRUE} (default), quoted terms wrapped in \code{c()}. If \code{FALSE}, output limited to contents inside of \code{c()}. Only relevant when terms entered in \code{...}}

\item{copy}{If \code{TRUE} (default), output is copied to clipboard}
}
\value{
Comma separated list of quoted terms. If \code{with_c} is \code{TRUE}, terms wrapped in \code{c()}
}
\description{
If terms not entered, clipboard will be used. In that case, terms can be unquoted text separated by spaces or commas or unquoted terms with 1 term per line
}
