% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_unquoted}
\alias{paste_unquoted}
\title{Convert character vector into a comma separated list of unquoted terms}
\usage{
paste_unquoted(..., with_c = FALSE, copy = TRUE)
}
\arguments{
\item{...}{Input to be transformed}

\item{with_c}{If \code{FALSE} (default), output not wrapped in \code{c()}. If \code{TRUE}, output wrapped in \code{c()}}

\item{copy}{If \code{TRUE} (default), output is copied to clipboard}
}
\value{
Comma separated list of unquoted terms. If \code{with_c} is \code{TRUE}, terms wrapped in \code{c()}
}
\description{
Convert character vector into a comma separated list of unquoted terms
}
