% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_vertical}
\alias{paste_vertical}
\title{Convert comma separated quoted terms to vertical unquoted terms}
\usage{
paste_vertical(..., suffix = NULL, copy = TRUE)
}
\arguments{
\item{...}{Enter as comma separated list of quoted terms without \code{c()}}

\item{suffix}{Suffix to include on each element in output except last. Can enter \code{","} or \code{" = ,"}}

\item{copy}{If \code{TRUE} (default), output is copied to clipboard}
}
\value{
Unquoted terms listed vertically
}
\description{
Convert comma separated quoted terms to vertical unquoted terms
}
