% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{path_to_desktop}
\alias{path_to_desktop}
\alias{desktop_path}
\title{Path to desktop}
\usage{
path_to_desktop(include_slash = TRUE)

desktop_path(include_slash = TRUE)
}
\arguments{
\item{include_slash}{If \code{TRUE} (default), path ends with "Desktop/". If \code{FALSE}, path ends with Desktop}
}
\value{
Path to desktop location
}
\description{
Path to desktop

Alias for path_to_desktop
}
