% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_pdf_png_jpg.R
\name{pdf_plot}
\alias{pdf_plot}
\title{Export plot as df (1 plot per page)}
\usage{
pdf_plot(
  plot,
  file_name = NULL,
  directory = getwd(),
  height = 5,
  width = 5,
  device = grDevices::pdf,
  ...
)
}
\arguments{
\item{plot}{Plot object or list of plot objects}

\item{file_name}{File name for plot. Enter as quoted file name}

\item{directory}{Path to directory where file should be saved. Default is working directory}

\item{height, width}{Plot height and width (in inches) respectively. Default is \code{5} for both}

\item{device}{Plot generating function. Options: \code{grDevices::pdf} (default) or \code{grDevices::cairo_pdf}}

\item{...}{Arguments passed to \code{device}}
}
\description{
Export plot as df (1 plot per page)
}
