% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_misc.R
\name{perc_change}
\alias{perc_change}
\title{Percent change}
\usage{
perc_change(x0, x1, digits = 1)
}
\arguments{
\item{x0, x1}{Baseline/reference (\code{x0}) and subsequent (\code{x1}) values respectively}

\item{digits}{Number of digits to include after decimal. Default is \code{1}}
}
\value{
Percentage change in value expressed as a percentage of starting value (\code{x0})
}
\description{
Percent change
}
