% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_info.R
\name{pkg_fns}
\alias{pkg_fns}
\title{Extract names of functions in package containing a specified pattern}
\usage{
pkg_fns(pkg, pattern = "", internal = TRUE, external = TRUE)
}
\arguments{
\item{pkg}{Package. Enter as quoted or unquoted package name}

\item{pattern}{Pattern in name of object}

\item{internal}{If \code{TRUE} (default), internal objects are included in output}

\item{external}{If \code{TRUE} (default), external objects are included in output}
}
\value{
Character vector of function names
}
\description{
Extract names of functions in package containing a specified pattern
}
