% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\alias{pb}
\title{Bar plot}
\usage{
plot_bar(
  df,
  formula = NULL,
  grouping_var = NULL,
  rev_grouping_var_order = FALSE,
  x_order = NULL,
  x = NULL,
  y = NULL,
  bar_color_var = NULL,
  colors = c("#2A2D34", "#BC3C29", "#00A1D5", "#009872", "#6761A8"),
  bar_colors = colors,
  bar_alpha = 1,
  bar_border_thickness = 0.75,
  bar_border_colors = "black",
  width = NULL,
  bar_width = width,
  dodge = 0.7,
  show_errorbar = TRUE,
  summary_fn = Mean,
  error_fn = SE,
  errorbar_width_multiplier = 0.5,
  show_legend = FALSE,
  legend_title = "",
  y_scale = "regular",
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  x_axis_breaks = NULL,
  x_axis_labels = NULL,
  y_title = waiver(),
  y_axis_title = y_title,
  x_title = waiver(),
  x_axis_title = x_title,
  plot_title = NULL,
  show_sig = TRUE,
  stars = TRUE,
  sig_font_color = "black",
  sig_bar_color = "black",
  sig_bar_thickness = 0.75,
  show_ns = TRUE,
  ns_symbol = "ns",
  sig_text_size = 12,
  sig_star_size = 22,
  sig_bar_nudge = 0.06,
  sig_star_nudge = 0.05,
  sig_text_nudge = 0.09,
  step_increase = if (y_scale == "regular") sig_bar_nudge + 0.05 else 0,
  sig_method = "p_by_normality",
  p_case = "upper",
  p_spaces = TRUE,
  n_breaks = 3,
  breaks_fn = pretty,
  y_max = NULL,
  expand_x = waiver(),
  censor_fn = rescale_none,
  show_all = TRUE,
  ...
)

pb(
  df,
  formula = NULL,
  grouping_var = NULL,
  rev_grouping_var_order = FALSE,
  x_order = NULL,
  x = NULL,
  y = NULL,
  bar_color_var = NULL,
  colors = c("#2A2D34", "#BC3C29", "#00A1D5", "#009872", "#6761A8"),
  bar_colors = colors,
  bar_alpha = 1,
  bar_border_thickness = 0.75,
  bar_border_colors = "black",
  width = NULL,
  bar_width = width,
  dodge = 0.7,
  show_errorbar = TRUE,
  summary_fn = Mean,
  error_fn = SE,
  errorbar_width_multiplier = 0.5,
  show_legend = FALSE,
  legend_title = "",
  y_scale = "regular",
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  x_axis_breaks = NULL,
  x_axis_labels = NULL,
  y_title = waiver(),
  y_axis_title = y_title,
  x_title = waiver(),
  x_axis_title = x_title,
  plot_title = NULL,
  show_sig = TRUE,
  stars = TRUE,
  sig_font_color = "black",
  sig_bar_color = "black",
  sig_bar_thickness = 0.75,
  show_ns = TRUE,
  ns_symbol = "ns",
  sig_text_size = 12,
  sig_star_size = 22,
  sig_bar_nudge = 0.06,
  sig_star_nudge = 0.05,
  sig_text_nudge = 0.09,
  step_increase = if (y_scale == "regular") sig_bar_nudge + 0.05 else 0,
  sig_method = "p_by_normality",
  p_case = "upper",
  p_spaces = TRUE,
  n_breaks = 3,
  breaks_fn = pretty,
  y_max = NULL,
  expand_x = waiver(),
  censor_fn = rescale_none,
  show_all = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{rev_grouping_var_order}{If \code{TRUE}, levels of \code{grouping_var} are reversed}

\item{x_order}{Order of x values. Enter as character vector. Default is \code{NULL}}

\item{x, y}{Variables for x  and y axis, respectively. Enter as quoted or unquoted variable names}

\item{bar_color_var}{Variable used to determine bar color. Enter as quoted or unquoted variable name}

\item{colors}{Alias for \code{point_colors}}

\item{bar_colors}{Bar color. Enter as character vector of color names or hexadecimal codes}

\item{bar_alpha}{Alpha value for bar fill. Default is \code{0}}

\item{bar_border_thickness}{Thickness of line surrounding bars in pts. Default is \code{0.75}}

\item{bar_border_colors}{Bar border color. Default is \code{"black"}}

\item{width, bar_width}{Width of bars. Default is 0.2*number of x values}

\item{dodge}{Entry for position_dodge. Default is \code{0.7}}

\item{show_errorbar}{If \code{TRUE}, error bars are displayed}

\item{summary_fn}{Function used to determine midpoint of error bars. Default is \code{Mean}}

\item{error_fn}{Function used to determine ends of error bars. Default is \code{SE}}

\item{errorbar_width_multiplier}{Multiplied by \code{bar_width} to determine width of error bars. Default is \code{0.5}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{legend_title}{Legend title. Default is blank}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{y_title}{Alias for \code{y_axis_title}}

\item{y_axis_title}{Enter as quoted string}

\item{x_title}{Alias for \code{x_axis_title}}

\item{x_axis_title}{Enter as quoted string}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{show_sig}{If \code{TRUE} (default), significance annotation displayed. If \code{FALSE}, no significance annotation displayed}

\item{stars}{If TRUE (default), significance stars are used instead of actual P values}

\item{sig_font_color}{Color for text in significance annotations. Default is \code{"black"}}

\item{sig_bar_color}{Color for significance bars and annotations. Default is \code{"black"}}

\item{sig_bar_thickness}{Line thickness for significance bars in pts. Default is \code{0.75}}

\item{show_ns}{If \code{FALSE} (default), no annotation for P >= 0.05. If \code{TRUE}, non-significant comparisons displayed}

\item{ns_symbol}{Symbol to use when P >= 0.05. Only used if show_ns is \code{TRUE.} default is \code{"ns"}}

\item{sig_text_size}{Font size for P values in pts. Only used when \code{stars = FALSE}}

\item{sig_star_size}{Size of significance stars in pts. Only relevant when \code{stars = TRUE}}

\item{sig_bar_nudge}{Nudge factor along y axis for significance bars. Distance between point or bar and significance bar is sig_bar_nudge x maximum y value on plot. Default is \code{0.06}}

\item{sig_star_nudge}{Nudge factor along y axis for significance stars. Distance between point or bar and significance stars is sig_star_nudge*maximum y value on plot. Default is \code{0.05}}

\item{sig_text_nudge}{Nudge factor along y axis for significance annotation text. Distance between point or bar and significance annotation text is sig_text_nudge x maximum y value on plot. Default is \code{0.09}}

\item{step_increase}{Step increase as proportion of view port. Default is \code{0.11}}

\item{sig_method}{Options: \code{"p_by_normality"} (default), \code{"mann_whitney"}, \code{"t_test_welch"}, \code{"t_test_no_welch"}. Enter as quoted test type}

\item{p_case}{Case of p used in significance annotations when \code{stats = FALSE.} Options: \code{"upper"} (default) or \code{"lower"}. Enter as quoted text}

\item{p_spaces}{If \code{TRUE} (default), spaces placed between p, equality symbol, and p-value in significance annotation. If \code{FALSE}, no spaces placed between p, equality symbol, and p-value in significance annotation. Only relevant when \code{stars = FALSE}}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{breaks_fn}{Function used to generate y axis breaks. Passed to \code{.create_axis_breaks}. Default is \code{pretty}}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{expand_x}{Argument entered for expand argument of \code{scale_x_continuous}. Default is \code{waiver()}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{show_all}{If \code{TRUE} (default), all significance bars are displayed}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\description{
Bar plot

Alias for \code{plot_bar()}
}
