% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot correlation coefficient, p value, linear regression line}
\usage{
plot_cor(
  df,
  formula,
  grouping_var = NULL,
  line_colors = "#0072B5",
  line_thickness = 1,
  point_colors = "#3B3B3B",
  point_color_var = NULL,
  point_size = 4,
  point_alpha = 1,
  y_scale = "regular",
  y_axis_breaks = NULL,
  y_max = NULL,
  y_min = NULL,
  expand_y = 0.1,
  y_axis_labels = NULL,
  y_axis_title = NULL,
  x_scale = "regular",
  x_axis_breaks = NULL,
  x_max = NULL,
  x_min = NULL,
  expand_x = 0.1,
  x_axis_title = NULL,
  plot_title = "",
  show_error = FALSE,
  error_color = "grey60",
  show_legend = FALSE,
  show_corr_coef = TRUE,
  anno_position = "best",
  method = "spearman",
  corr_prefix = NULL,
  p_prefix = "P",
  show_n = TRUE,
  n_case = "lower",
  show_rsq = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{line_colors}{Color for lines. Hexadecimal or quoted color names. Default is \code{"black"}}

\item{line_thickness}{Thickness of line in mm. Default is \code{0.75}}

\item{point_colors}{Color of points. Enter as character vector of color names or hexadecimal codes}

\item{point_color_var}{Variable used to determine point colors. Enter as quoted or unquoted variable name. Default uses \code{grouping_var}}

\item{point_size}{Point size. Default is \code{5}}

\item{point_alpha}{Transparency of point colors. Default is \code{0.9}}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{expand_y, expand_x}{Expansion to add to y and x axis, respectively. Default is \code{0.1}}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{y_axis_title}{Enter as quoted string}

\item{x_scale}{Scale used for x axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_max}{Maximum value for x axis. Default is \code{NULL}}

\item{x_min}{Minimum value for x axis. Default is \code{NULL}}

\item{x_axis_title}{Enter as quoted string}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""}}

\item{show_error}{If \code{FALSE} (default), error region not displayed If \code{TRUE}, error region displayed with color as determined by error_color}

\item{error_color}{Color used to fill error region around line. Default is grey}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{show_corr_coef}{If \code{TRUE} (default), correlation coefficient and p value are displayed. If \code{FALSE}, correlation coefficient and p value are not displayed}

\item{anno_position}{Position of correlation coefficient and p value. Can enter as numeric vector of length 2 containing x and y coordinates or a combination of t or b and l or r in quotes}

\item{method}{Method for performing correlation test. Options include \code{"spearman"} (default), \code{"pearson"}, \code{"kendall"}}

\item{corr_prefix}{Prefix for correlation coefficient annotation. If \code{NULL}, rho used for spearman, R for Pearson, tau for kendall}

\item{p_prefix}{Prefix for P values. Default is \code{"P"}. Enter as quoted text}

\item{show_n}{If \code{TRUE} (default), number of observations displayed as component of correlation annotation}

\item{n_case}{Case for number of observations annotation. Default is \code{"lower"}}

\item{show_rsq}{If \code{TRUE} (default), R squared and p value are displayed. If \code{FALSE}, R squared and p value are not displayed}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\description{
Plot correlation coefficient, p value, linear regression line
}
