% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_cor_heatmap}
\alias{plot_cor_heatmap}
\title{Correlation heatmap}
\usage{
plot_cor_heatmap(
  df,
  ...,
  colors = clr_continuous(c("#0072B5", "white", "#BC3C29"), n_colors = 100),
  method = c("spearman", "pearson", "kendall"),
  min_variance = 1
)
}
\arguments{
\item{df}{Data frame in wide form (columns will be correlated with one another)}

\item{...}{Variables to be included in heatmap}

\item{colors}{Color for heatmap. Default from blue (low) to red (high)}

\item{method}{Options: \code{"spearman"} (default), \code{"pearson"}, \code{"kendall"}}

\item{min_variance}{Minimum variance allowed for variable above which variable will be excluded from heatmap. Default is \code{1} (includes all columns)}
}
\description{
Correlation heatmap
}
