% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Forest plot}
\usage{
plot_forest(
  df,
  risk,
  risk_lower,
  risk_upper,
  label_var = "variable",
  ordering_var = "HR",
  reorder = FALSE,
  x_axis_title = NULL,
  hr_label = "regular",
  color_var = NULL,
  base_size = 12,
  label_size = 8,
  hjust = 1,
  ratio = 2.5,
  add_n_bands = NULL,
  n_bands = NULL,
  x_min = NULL,
  x_max = NULL,
  colors = c("#0072B5", "#BC3C29", "black"),
  point_shape = "square",
  point_size = 1.5,
  errorbar_width = 0.15,
  line_thickness = 0.9,
  point_border_thickness = 0.7,
  vert_line_thickness = 0.5,
  vert_linetype = "dashed",
  x_label_fn = NULL,
  scale = "log",
  log_scale = "log10",
  n_rows = 1,
  scales = "fixed",
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{risk, risk_lower, risk_upper}{Columns in df containing risk ratio, lower, and upper bound of confidence interval, respectively}

\item{label_var}{Variable used to create row labels in plot. Enter as quoted or unquoted variable names}

\item{ordering_var}{Variable used to reorder rows in forest plot. Enter as quoted or unquoted variable names. Only relevant when reorder is \code{TRUE}}

\item{reorder}{If \code{TRUE} (default), rows in forest plot will be reordered using}

\item{x_axis_title}{Title for x axis}

\item{hr_label}{HR label}

\item{color_var}{Variable used to determine color or HR/CI bars. Enter as quoted or unquoted variable name}

\item{base_size}{Base size for plot font (passed to \code{theme_custom()}). Default is \code{12}}

\item{label_size}{Font size for labels}

\item{hjust}{Horizontal adjustment}

\item{ratio}{Aspect ratio of plot}

\item{add_n_bands}{Number of extra grey/white bands to add above}

\item{n_bands}{Total number of grey/white bands}

\item{x_min, x_max}{Lower and upper limit for x axis values, respectively}

\item{colors}{Colors for HR/CI bars}

\item{point_shape}{Shapes used to display HR. Options: \code{circle} (default), \code{diamond}, \code{square}, \code{triangle}}

\item{point_size}{Size of HR points}

\item{errorbar_width}{Width of error bars}

\item{line_thickness}{Line thickness for HR/CI bars}

\item{point_border_thickness}{Border thickness for HR points}

\item{vert_line_thickness}{Thickness of reference vertical line}

\item{vert_linetype}{Linetype for reference vertical line}

\item{x_label_fn}{Function used to generate x axis labels}

\item{scale}{Scale for HR}

\item{log_scale}{Type of log scale}

\item{n_rows}{Number of rows}

\item{scales}{Scales}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\description{
Forest plot
}
