% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq.R
\name{plot_freq}
\alias{plot_freq}
\title{Frequency bar plot}
\usage{
plot_freq(
  df,
  col = NULL,
  grouping_var = NULL,
  formula = NULL,
  order_by_freq = c("both", "x", "grouping_var", "none"),
  x_order = NULL,
  rev_x_order = FALSE,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  bar_color_var = NULL,
  colors = "#2A2D34",
  bar_border_thickness = 0.75,
  bar_width = NULL,
  width = bar_width,
  dodge = 0.7,
  show_legend = FALSE,
  legend_title = "",
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  x_axis_breaks = waiver(),
  x_axis_labels = waiver(),
  x_axis_label_angle = NULL,
  y_axis_title_angle = 0,
  y_axis_title = "N",
  x_axis_title = NULL,
  plot_title = NULL,
  n_breaks = 3,
  breaks_fn = pretty,
  y_max = NULL,
  censor_fn = rescale_none,
  aspect_ratio = NULL,
  show_anno = TRUE,
  anno_type = c("counts", "percent"),
  anno_font_size = NULL,
  anno_font_color = "black",
  anno_y_nudge = 0.03,
  show_n = FALSE,
  n_anno_font_size = 12,
  n_anno_font_color = "black",
  n_anno_x = "right",
  n_anno_x_nudge = 0.2,
  n_prefix = "N = ",
  ...
)
}
\arguments{
\item{df}{Data frame containing column \code{col} or vector of values}

\item{col}{Categorical variable. Enter as quoted or unquoted variable name}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{formula}{y ~ x format}

\item{order_by_freq}{Options: \code{"both"} (default, both x axis variable and \code{grouping_var} ordered by frequency), \code{"x"} (x axis variable is ordered by frequency), \code{"grouping_var"} (\code{"grouping_var"} is ordered by frequency), \code{"none"} (no variables are ordered by frequency). Default is to order groups by decreasing frequency. If \code{rev_x_order} or \code{rev_grouping_var_order} is \code{TRUE}, groups are ordered by increasing frequency. Only relevant when \code{x_order = NULL} and \code{grouping_var_order = NULL}}

\item{x_order}{Order of x values. Enter as character vector. Default is \code{NULL}}

\item{rev_x_order}{If \code{FALSE} (default), order of x axis variable is not reversed}

\item{grouping_var_order}{Order for levels of \code{grouping_var} along x axis. Enter as character vector}

\item{rev_grouping_var_order}{If \code{TRUE}, levels of \code{grouping_var} are reversed}

\item{bar_color_var}{Variable used to determine bar color. Enter as quoted or unquoted variable name}

\item{colors}{Alias for \code{point_colors}}

\item{bar_border_thickness}{Thickness of line surrounding bars in pts. Default is \code{0.75}}

\item{width, bar_width}{Width of bars. Default is 0.2*number of x values}

\item{dodge}{Entry for position_dodge. Default is \code{0.7}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{legend_title}{Legend title. Default is blank}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{x_axis_label_angle}{Angle of x axis labels}

\item{y_axis_title_angle}{Angle for y axis title. Default is \code{0}}

\item{y_axis_title}{Enter as quoted string}

\item{x_axis_title}{Enter as quoted string}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{breaks_fn}{Function used to generate y axis breaks. Passed to \code{.create_axis_breaks}. Default is \code{pretty}}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{aspect_ratio}{Aspect ratio for plot}

\item{show_anno}{If \code{TRUE} (default), group size is displayed above each bar}

\item{anno_type}{Options: \code{"counts"} (default), \code{"percent"}}

\item{anno_font_size}{Font size for counts in pts}

\item{anno_font_color}{Font color for counts. Default is \code{"black"}}

\item{anno_y_nudge}{Amount to nudge counts above bar as percent of y axis range. Default is \code{0.03}}

\item{show_n}{If \code{TRUE}, total sample size is displayed on plot}

\item{n_anno_font_size}{Font size for sample size annotation in pts. Default is \code{12}}

\item{n_anno_font_color}{Font color for sample size annotation. Default is \code{"black"}}

\item{n_anno_x}{Position of sample size annotation along x axis. Options: \code{"right"} (default),\code{"left"}}

\item{n_anno_x_nudge}{Amount to nudge sample size annotation along x axis. Default is \code{0.2}}

\item{n_prefix}{Prefix for sample size annotation. Default is \code{"N = "}}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\description{
Frequency bar plot
}
