% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_incidence.R
\name{plot_incidence}
\alias{plot_incidence}
\title{Frequency plot showing incidence over time}
\usage{
plot_incidence(
  df,
  col = NULL,
  colors = "#328EC3",
  bar_colors = colors,
  month_format = c("number", "name", "abbreviation"),
  leading_zero = FALSE,
  full_year = FALSE,
  sep = "-",
  n_bins = 20,
  x_axis_title = NULL,
  expand_x = if (n_bins > 10) 0.1 else 0.15,
  x_angle = 45,
  y_axis_title = "n",
  ...
)
}
\arguments{
\item{df}{Data frame or vector of dates}

\item{col}{Column containing dates if df is a data frame. Enter as quoted or unquoted variable name}

\item{colors}{Color of bars. Default is \code{"#328EC3"}}

\item{bar_colors}{Alias for colors}

\item{month_format}{Options: \code{"number"} (default), \code{"name"}, \code{"abbreviation"}}

\item{leading_zero}{If \code{FALSE} (default), leading 0s are removed from month and day if < 10}

\item{full_year}{If \code{FALSE} (default), last 2 digits are used. If \code{TRUE}, year is displayed using 4 digits}

\item{sep}{Separator between month, day, and year}

\item{n_bins}{Number of bins for x axis. Default is 20}

\item{x_axis_title}{Title for x axis}

\item{expand_x}{Expansion multiplier for x axis. Default is \code{0.1} if n_bins > 10 and \code{0.15} if n_bins <= 10. Only relevant if x is treated as numeric}

\item{x_angle}{Angle for x axis labels. Default is \code{45}}

\item{y_axis_title}{Title for y axis. Default is "n"}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\description{
Frequency plot showing incidence over time
}
