% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_line}
\alias{plot_line}
\title{Plot line}
\usage{
plot_line(
  df,
  formula = NULL,
  grouping_var = NULL,
  rev_grouping_var_order = FALSE,
  x = NULL,
  y = NULL,
  show_errorbar = TRUE,
  summary_fn = Mean,
  error_fn = SE,
  colors = c("#0072B5", "#BC3C29", "#868686", "#2A2D34"),
  show_points = TRUE,
  point_colors = colors,
  point_shape = "circle",
  point_size = 3.5,
  point_border_thickness = 1,
  point_border_colors = "black",
  alpha = 1,
  point_alpha = alpha,
  line_colors = colors,
  line_thickness = 0.75,
  line_type = "solid",
  errorbar_color = "black",
  errorbar_thickness = 0.75,
  errorbar_width = 0.15,
  show_legend = FALSE,
  y_scale = "regular",
  x_axis_title = waiver(),
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  y_axis_title = waiver(),
  x_axis_breaks = waiver(),
  expand_y = 0,
  expand_x = waiver(),
  plot_title = NULL,
  n_breaks = 3,
  show_sig = TRUE,
  stars = TRUE,
  show_ns = TRUE,
  ns_symbol = "ns",
  sig_text_size = 12,
  sig_star_size = 22,
  sig_star_nudge = 0.05,
  sig_text_nudge = 0.09,
  sig_method = "p_by_normality",
  p_case = "upper",
  p_spaces = TRUE,
  y_min = NULL,
  y_max = NULL,
  censor_fn = rescale_none,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable (y), categorical or integer variable (x), and grouping variable (grouping_var)}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{rev_grouping_var_order}{If \code{FALSE} (default), levels of grouping_var ordered alphabetically. If \code{TRUE}, order of grouping_var levels is reversed}

\item{x, y}{Variables for x and y axis respectively. Enter as quoted or unquoted variable names}

\item{show_errorbar}{If \code{TRUE} (default), error bars are displayed. If \code{FALSE}, error bars are not displayed}

\item{summary_fn}{Statistical function used to as summary estimate. Default is \code{Mean}}

\item{error_fn}{Statistical function used to calculate error. Default is \code{SE}}

\item{colors}{Color used for points and lines}

\item{show_points}{If \code{TRUE} (default), points are displayed}

\item{point_colors}{Fill color for points. Hexadecimal or quoted color names. Enter using \code{c()}}

\item{point_shape}{Shape of points. Options include \code{"square"}, \code{"circle"}, \code{"triangle"}, \code{"diamond"}, or list of shapes for each group. Enter as quoted or unquoted shape or list of shapes using \code{c()}}

\item{point_size}{Point size. Default is \code{3}}

\item{point_border_thickness}{Thickness of line surrounding points in pt units. Default is \code{1}}

\item{point_border_colors}{Color for line surrounding points. Default is \code{"black"}}

\item{alpha}{Alias for point_alpha}

\item{point_alpha}{Alpha value for point fill. Default is \code{1}}

\item{line_colors}{Color for lines. Hexadecimal or quoted color names. Default is \code{"black"}}

\item{line_thickness}{Thickness of line in mm. Default is \code{0.75}}

\item{line_type}{Type of line. Options include \code{"solid"} (default), \code{"dashed"}, \code{"longdash"}, \code{"twodash"}, \code{"dotdash"}, \code{"dotted"}. Enter as quoted linetype}

\item{errorbar_color}{Color of error bars. Default is \code{"black"}}

\item{errorbar_thickness}{Thickness of error bars. Default is \code{0.75}}

\item{errorbar_width}{Width of error bars. Default is \code{point_size/6}}

\item{show_legend}{If \code{TRUE}, legend displayed. If \code{FALSE}, legend not displayed}

\item{y_scale}{Scaling for y-axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{x_axis_title}{Enter as quoted string}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{y_axis_title}{Enter as quoted string}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0}}

\item{expand_x}{Argument entered for \code{expand} argument of \code{scale_x_continuous}}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{show_sig}{If \code{TRUE} (default), significance annotation displayed. If \code{FALSE}, no significance annotation displayed}

\item{stars}{If \code{TRUE} (default), significance stars are displayed rather than actual P values}

\item{show_ns}{If \code{FALSE} (default), no annotation for P > 0.05. If \code{TRUE}, non-significant comparisons displayed}

\item{ns_symbol}{Symbol to use when P > 0.05. Only used if show_ns is TRUE. default is \code{"ns"}}

\item{sig_text_size}{Font size for P values in pts. Only used when stars is \code{FALSE}}

\item{sig_star_size}{Size of significance stars. Only used when stars is \code{TRUE.} Default is \code{8}}

\item{sig_star_nudge}{Nudge factor along y axis for significance stars. Distance between point or bar and significance stars is \code{sig_star_nudge} x maximum y value on plot. Default is \code{0.05}}

\item{sig_text_nudge}{Nudge factor along y axis for significance annotation text. Distance between point or bar and significance annotation text is sig_text_nudge*maximum y value on plot. Default is \code{0.09}}

\item{sig_method}{Options: \code{"p_by_normality"} (default), \code{"mann_whitney"}, \code{"t_test_welch"}, \code{"t_test_no_welch"}. Enter as quoted test type}

\item{p_case}{Case of p used in significance annotations when stars is FALSE. Options: \code{"upper"} (default) or \code{"lower"}. Enter as quoted text}

\item{p_spaces}{If \code{TRUE} (default), spaces placed between p, equality symbol, and p-value in significance annotation. If \code{FALSE}, no spaces placed between p, equality symbol, and p-value in significance annotation. Only relevant when stars is \code{FALSE}}

\item{y_min}{Minimum value to use for y axis Default is \code{0}}

\item{y_max}{Maximum y axis value. Default is \code{max(df$y)}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\description{
Plot line
}
