% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{plot_point}
\alias{plot_point}
\alias{pp}
\title{Point plot}
\usage{
plot_point(
  df,
  formula = NULL,
  grouping_var = NULL,
  ...,
  x = NULL,
  y = NULL,
  beeswarm = FALSE,
  beeswarm_method = "smiley",
  width = 0.2,
  beeswarm_width = width,
  n_bins = 25,
  band_width = 1 + n_bins/5,
  dodge = if (beeswarm) 0.8 else 0.7,
  jitter_width = width,
  x_order = NULL,
  rev_x_order = FALSE,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  point_shapes = c("circle", "circle"),
  point_shape_var = NULL,
  point_color_var = NULL,
  colors = c("#0072B5", "#BC3C29", "#8C8C8C", "#2A2D34", "#009872", "#6761A8"),
  point_colors = colors,
  alpha = 0.9,
  point_alpha = alpha,
  point_size = 5,
  point_border_thickness = 1,
  point_border_color = "black",
  show_errorbar = FALSE,
  summary_fn = Mean,
  error_fn = SE,
  errorbar_width = NULL,
  errorbar_thickness = 0.75,
  show_legend = FALSE,
  legend_title = "",
  y_title = waiver(),
  y_axis_title = y_title,
  y_scale = "regular",
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  breaks_fn = pretty,
  n_breaks = 4,
  y_min = NULL,
  y_max = NULL,
  expand_y = 0.1,
  censor_fn = rescale_none,
  x_title = waiver(),
  x_axis_title = x_title,
  x_axis_breaks = NULL,
  x_axis_labels = NULL,
  expand_x = waiver(),
  plot_title = NULL,
  show_sig = TRUE,
  sig_method = "p_by_normality",
  stars = TRUE,
  show_ns = TRUE,
  ns_symbol = "ns",
  sig_bar_nudge = if (y_scale == "log") 0.9 else 0.11,
  sig_star_nudge = if (y_scale == "log") -0.45 else sig_bar_nudge - 0.03,
  sig_text_nudge = if (y_scale == "log") 0.5 else sig_bar_nudge + 0.03,
  step_increase = if (y_scale == "log") 0 else sig_bar_nudge + 0.05,
  sig_bar_thickness = 0.75,
  sig_bar_color = "black",
  sig_font_color = "black",
  sig_star_size = 22,
  sig_text_size = 12,
  p_case = "upper",
  p_spaces = TRUE,
  seed = 1234,
  theme_fn = theme_custom
)

pp(
  df,
  formula = NULL,
  grouping_var = NULL,
  ...,
  x = NULL,
  y = NULL,
  beeswarm = FALSE,
  beeswarm_method = "smiley",
  width = 0.2,
  beeswarm_width = width,
  n_bins = 25,
  band_width = 1 + n_bins/5,
  dodge = if (beeswarm) 0.8 else 0.7,
  jitter_width = width,
  x_order = NULL,
  rev_x_order = FALSE,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  point_shapes = c("circle", "circle"),
  point_shape_var = NULL,
  point_color_var = NULL,
  colors = c("#0072B5", "#BC3C29", "#8C8C8C", "#2A2D34", "#009872", "#6761A8"),
  point_colors = colors,
  alpha = 0.9,
  point_alpha = alpha,
  point_size = 5,
  point_border_thickness = 1,
  point_border_color = "black",
  show_errorbar = FALSE,
  summary_fn = Mean,
  error_fn = SE,
  errorbar_width = NULL,
  errorbar_thickness = 0.75,
  show_legend = FALSE,
  legend_title = "",
  y_title = waiver(),
  y_axis_title = y_title,
  y_scale = "regular",
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  breaks_fn = pretty,
  n_breaks = 4,
  y_min = NULL,
  y_max = NULL,
  expand_y = 0.1,
  censor_fn = rescale_none,
  x_title = waiver(),
  x_axis_title = x_title,
  x_axis_breaks = NULL,
  x_axis_labels = NULL,
  expand_x = waiver(),
  plot_title = NULL,
  show_sig = TRUE,
  sig_method = "p_by_normality",
  stars = TRUE,
  show_ns = TRUE,
  ns_symbol = "ns",
  sig_bar_nudge = if (y_scale == "log") 0.9 else 0.11,
  sig_star_nudge = if (y_scale == "log") -0.45 else sig_bar_nudge - 0.03,
  sig_text_nudge = if (y_scale == "log") 0.5 else sig_bar_nudge + 0.03,
  step_increase = if (y_scale == "log") 0 else sig_bar_nudge + 0.05,
  sig_bar_thickness = 0.75,
  sig_bar_color = "black",
  sig_font_color = "black",
  sig_star_size = 22,
  sig_text_size = 12,
  p_case = "upper",
  p_spaces = TRUE,
  seed = 1234,
  theme_fn = theme_custom
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{...}{Arguments passed to \code{theme_fn()}}

\item{x, y}{Variables for x  and y axis, respectively. Enter as quoted or unquoted variable names}

\item{beeswarm}{If \code{FALSE} (default), jittered points. If \code{TRUE}, beeswarm plot}

\item{beeswarm_method}{Method for \code{geom_quasirandom} function. Default is \code{"smiley"}}

\item{width}{Alias for either \code{jitter_width} or \code{beeswarm_width}}

\item{beeswarm_width}{Width argument for beeswarm points}

\item{n_bins}{Number of bins. Used in \code{ggbeeswarm::geom_quasirandom} function. Default uses 30/number of points}

\item{band_width}{Bandwidth. Used in \code{ggbeeswarm::geom_quasirandom} function. Default is \code{1 + n_bins/5}}

\item{dodge}{Entry for position_dodge. Default is \code{0.7}}

\item{jitter_width}{Width between groups of points. Default is \code{0.2}}

\item{x_order}{Order of x values. Enter as character vector. Default is \code{NULL}}

\item{rev_x_order}{If \code{FALSE} (default), order of x axis variable is not reversed}

\item{grouping_var_order}{Order for levels of \code{grouping_var} along x axis. Enter as character vector}

\item{rev_grouping_var_order}{If \code{TRUE}, levels of \code{grouping_var} are reversed}

\item{point_shapes}{Point shapes. Options: \code{"circle"} (default), \code{"square"}, \code{"triangle"}, \code{"diamond"}. Enter as quoted shape}

\item{point_shape_var}{Variable used to determine point shapes. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{point_color_var}{Variable used to determine point colors. Enter as quoted or unquoted variable name. Default uses \code{grouping_var}}

\item{colors}{Alias for \code{point_colors}}

\item{point_colors}{Color of points. Enter as character vector of color names or hexadecimal codes}

\item{alpha}{Alias for \code{point_alpha}}

\item{point_alpha}{Transparency of point colors. Default is \code{0.9}}

\item{point_size}{Point size. Default is \code{5}}

\item{point_border_thickness}{Thickness of line surrounding points in pts. Default is \code{1}}

\item{point_border_color}{Color of line surrounding points. Default is \code{"black"}}

\item{show_errorbar}{If \code{TRUE}, error bars are displayed}

\item{summary_fn}{Function used to determine midpoint of error bars. Default is \code{Mean}}

\item{error_fn}{Function used to determine ends of error bars. Default is \code{SE}}

\item{errorbar_width}{Width of error bars. Default is \code{n_unique(df$x)/10}}

\item{errorbar_thickness}{Line thickness of error bar. Default is \code{0.75}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{legend_title}{Legend title. Default is blank}

\item{y_title}{Alias for \code{y_axis_title}}

\item{y_axis_title}{Enter as quoted string}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{breaks_fn}{Function used to generate y axis breaks. Passed to \code{.create_axis_breaks}. Default is \code{pretty}}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{x_title}{Alias for \code{x_axis_title}}

\item{x_axis_title}{Enter as quoted string}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{expand_x}{Argument entered for expand argument of \code{scale_x_continuous}. Default is \code{waiver()}}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{show_sig}{If \code{TRUE} (default), significance annotation displayed. If \code{FALSE}, no significance annotation displayed}

\item{sig_method}{Options: \code{"p_by_normality"} (default), \code{"mann_whitney"}, \code{"t_test_welch"}, \code{"t_test_no_welch"}. Enter as quoted test type}

\item{stars}{If TRUE (default), significance stars are used instead of actual P values}

\item{show_ns}{If \code{FALSE} (default), no annotation for P >= 0.05. If \code{TRUE}, non-significant comparisons displayed}

\item{ns_symbol}{Symbol to use when P >= 0.05. Only used if show_ns is \code{TRUE.} default is \code{"ns"}}

\item{sig_bar_nudge}{Nudge factor along y axis for significance bars. Distance between point or bar and significance bar is sig_bar_nudge x maximum y value on plot. Default is \code{0.06}}

\item{sig_star_nudge}{Nudge factor along y axis for significance stars. Distance between point or bar and significance stars is sig_star_nudge*maximum y value on plot. Default is \code{0.05}}

\item{sig_text_nudge}{Nudge factor along y axis for significance annotation text. Distance between point or bar and significance annotation text is sig_text_nudge x maximum y value on plot. Default is \code{0.09}}

\item{step_increase}{Step increase as proportion of view port. Default is \code{0.11}}

\item{sig_bar_thickness}{Line thickness for significance bars in pts. Default is \code{0.75}}

\item{sig_bar_color}{Color for significance bars and annotations. Default is \code{"black"}}

\item{sig_font_color}{Color for text in significance annotations. Default is \code{"black"}}

\item{sig_star_size}{Size of significance stars in pts. Only relevant when \code{stars = TRUE}}

\item{sig_text_size}{Font size for P values in pts. Only used when \code{stars = FALSE}}

\item{p_case}{Case of p used in significance annotations when \code{stats = FALSE.} Options: \code{"upper"} (default) or \code{"lower"}. Enter as quoted text}

\item{p_spaces}{If \code{TRUE} (default), spaces placed between p, equality symbol, and p-value in significance annotation. If \code{FALSE}, no spaces placed between p, equality symbol, and p-value in significance annotation. Only relevant when \code{stars = FALSE}}

\item{seed}{Length 1 integer vector to set seed for reproducibility when using jittering. Default is \code{1234}}

\item{theme_fn}{Theme function. Default is \code{theme_custom}}
}
\description{
Point plot

Alias for plot_point
}
