% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{plot_resize}
\alias{plot_resize}
\title{Resize plot aspect ratio}
\usage{
plot_resize(x = last_plot(), ratio = 1.618034)
}
\arguments{
\item{x}{ggplot object to resize. Default is last plot}

\item{ratio}{New aspect ratio. Defaults to golden ratio}
}
\description{
Resize plot aspect ratio
}
