% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{ROC curve}
\usage{
plot_roc(
  df,
  outcome_var,
  predictor_var,
  type = ">=",
  line_color = "#0072B5",
  line_thickness = 1,
  point_color = "black",
  point_size = 1.5,
  show_points = FALSE,
  show_error = FALSE,
  show_auc = TRUE,
  show_n = TRUE,
  auc_position = "left",
  auc_prefix = "AUC ",
  n_prefix = "n = ",
  annotation_x = 0.45,
  annotation_y = 0.15,
  error_color = "grey70",
  error_alpha = 0.4,
  x_title = "FPR",
  x_axis_title = x_title,
  y_title = "TPR",
  y_axis_title = y_title,
  expand = 0.1,
  expand_x = expand,
  expand_y = expand,
  breaks = seq(0, 1, 0.25),
  percent = FALSE,
  ci = 0.95,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{outcome_var}{Binary variable. Enter as quoted or unquoted variable name}

\item{predictor_var}{Continuous variable. Enter as quoted or unquoted variable name}

\item{type}{Type of comparison. Options: \code{">="} (default), \code{">"}, \code{"<="}, \code{"<"}}

\item{line_color}{Color of ROC curve. Default is red}

\item{line_thickness}{Thickness of ROC curve. Default is \code{1}}

\item{point_color}{Color of points. Default is \code{"black"}}

\item{point_size}{Size of points. Default is \code{1.5}}

\item{show_points}{If \code{FALSE} (default), individual points are not displayed}

\item{show_error}{If \code{FALSE} (default), error in ROC curve estimation is not displayed}

\item{show_auc}{If \code{TRUE} (default), AUC is displayed}

\item{show_n}{If \code{TRUE} (default), number of observations is displayed}

\item{auc_position}{Position of AUC annotation. Options: \code{"left"} (default), \code{"right"}}

\item{auc_prefix}{Text used to prefix AUC annotation. Default is \code{"AUC "}}

\item{n_prefix}{Text used to prefix n annotation. Default is \code{"n = "}}

\item{annotation_x, annotation_y}{Position of annotation along x and y axis respectively}

\item{error_color}{Color of error region around ROC curve. Default is \code{"grey"}}

\item{error_alpha}{Alpha for colored error region. Default is \code{0.4}}

\item{x_title}{Alias for \code{x_axis_title}}

\item{x_axis_title}{Title for x axis. Default is \code{"FPR"}}

\item{y_title}{Alias for \code{x_axis_title}}

\item{y_axis_title}{Title for x axis. Default is \code{"TPR"}}

\item{expand}{Expansion added around x and y axis. Default is \code{0.1}}

\item{expand_x, expand_y}{Expansion added around x and y axis respectively. Default is \code{0.1}}

\item{breaks}{Axis breaks. Enter as numeric vector. Default is \code{c(0, 0.25, 0.5, 0.75, 1)}}

\item{percent}{If \code{FALSE} (default), axes reflect proportions. If \code{TRUE}, axes reflect percentages}

\item{ci}{Confidence interval. Default is \code{0.95}}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\description{
ROC curve
}
