% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot}
\usage{
plot_scatter(
  df,
  formula = NULL,
  x = NULL,
  y = NULL,
  size = 3,
  point_size = size,
  point_size_var = NULL,
  shape = c("circle", "square", "triangle", "diamond"),
  point_shape = shape,
  point_shape_var = NULL,
  colors = c("#0072B5", "#BC3C29", "#868686", "#3B3B3B"),
  point_colors = colors,
  point_color_var = NULL,
  border_color = "black",
  point_border_color = border_color,
  point_border_color_var = NULL,
  stroke = 0.75,
  point_border_thickness = stroke,
  alpha = 0.75,
  point_alpha = alpha,
  n_breaks = 3,
  y_max = NULL,
  y_min = NULL,
  x_max = NULL,
  x_min = NULL,
  y_axis_title = waiver(),
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  x_axis_title = waiver(),
  x_axis_breaks = NULL,
  x_axis_labels = NULL,
  expand = 0.1,
  expand_y = expand,
  expand_x = expand,
  scale = "regular",
  x_scale = scale,
  y_scale = scale,
  show_legend = FALSE,
  censor_fn = rescale_none,
  plot_title = NULL,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{x, y}{Continuous variables variables. Enter as quoted or unquoted variable names}

\item{point_size}{Point size. Default is \code{5}}

\item{point_size_var}{Variable used to determine point size. Enter as quoted or unquoted variable name}

\item{shape, size, stroke}{Alias for \code{point_shape}, \code{point_size}, \code{point_border_thickness} respectively}

\item{point_shape}{Point shape Options: \code{"circle"} (default), \code{"square"}, \code{"triangle"}, \code{"diamond"}}

\item{point_shape_var}{Variable used to determine point shapes. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{colors}{Alias for \code{point_colors}}

\item{point_colors}{Color of points. Enter as character vector of color names or hexadecimal codes}

\item{point_color_var}{Variable used to determine point colors. Enter as quoted or unquoted variable name. Default uses \code{grouping_var}}

\item{border_color}{Alias for \code{point_border_color}}

\item{point_border_color}{Color of line surrounding points. Default is \code{"black"}}

\item{point_border_color_var}{Variable used to determine border color of points. Enter as quoted or unquoted variable name}

\item{point_border_thickness}{Thickness of line surrounding points in pts. Default is \code{1}}

\item{alpha}{Alias for \code{point_alpha}}

\item{point_alpha}{Transparency of point colors. Default is \code{0.9}}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{x_max}{Maximum value for x axis. Default is \code{NULL}}

\item{x_min}{Minimum value for x axis. Default is \code{NULL}}

\item{y_axis_title}{Enter as quoted string}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{x_axis_title}{Enter as quoted string}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{expand}{Alias for \code{expand_y} and \code{expand_x.} Sets both simultaneously}

\item{expand_y, expand_x}{Expansion to add to y and x axis, respectively. Default is \code{0.1}}

\item{scale}{Alias for \code{y_scale} and \code{x_scale} that sets both simultaneously}

\item{x_scale}{Scale used for x axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""}}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\description{
Scatter plot
}
