% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smooth.R
\name{plot_smooth}
\alias{plot_smooth}
\title{Plot regression line or loess smoothed mean per group}
\usage{
plot_smooth(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  grouping_var = NULL,
  smoothing_fn = stats::loess,
  x_range_data = c(-Inf, Inf),
  show_error = TRUE,
  ci = 0.95,
  line_type = "solid",
  colors = c("#2A2D34", "#BC3C29", "#00A1D5", "#6761A8", "#009872"),
  line_colors = colors,
  line_thickness = 1,
  plot_formula = y ~ x,
  show_points = FALSE,
  point_colors = colors,
  point_shape = "circle",
  point_size = 2,
  point_border_thickness = 1,
  point_border_colors = "black",
  point_alpha = 1,
  y_scale = "regular",
  expand_y = 0.1,
  y_axis_label_fn = axis_label_numeric,
  y_axis_title = NULL,
  x_axis_title = NULL,
  x_axis_breaks = NULL,
  breaks_fn = pretty,
  x_min = NULL,
  x_max = NULL,
  plot_title = NULL,
  n_breaks = 4,
  show_legend = FALSE,
  censor_fn = rescale_none,
  error_fill_color = "grey60",
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable (y), categorical or integer variable (x), and grouping variable (\code{grouping_var})}

\item{formula}{y ~ x format}

\item{y, x}{Variables used for plotting. Enter as quoted or unquoted variable names}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{smoothing_fn}{Function used for smoothing. Options: \code{"loess"} (default) \code{"lm"}, \code{"glm"}}

\item{x_range_data}{Range of x values for input data. Default is -Inf to Inf. Enter as length 2 numeric vector}

\item{show_error}{If \code{TRUE} (default), error estimate for curve displayed as shaded region}

\item{ci}{Confidence interval used for error region. Default is \code{0.95}}

\item{line_type}{Type of line. Options include \code{"solid"} (default), \code{"dashed"}, \code{"longdash"}, \code{"twodash"}, \code{"dotdash"}, \code{"dotted"}. Enter as quoted linetype}

\item{colors}{Color used for points and lines}

\item{line_colors}{Color for lines. Hexadecimal or quoted color names. Default is \code{"black"}}

\item{line_thickness}{Thickness of line in mm Default is \code{1}}

\item{plot_formula}{Formula for calculation. Default is \code{y ~ x}}

\item{show_points}{If \code{FALSE} (default), points not displayed. If \code{TRUE}, raw data points displayed}

\item{point_colors}{Fill color for points. Hexadecimal or quoted color names. Enter using \code{c()}}

\item{point_shape}{Shape of points. Options include \code{"square"}, \code{"circle"}, \code{"triangle"}, \code{"diamond"}, or list of shapes for each group. Enter as quoted or unquoted shape or list of shapes using \code{c()}}

\item{point_size}{Point size. Default is \code{2}}

\item{point_border_thickness}{Thickness of line surrounding points in pt units. Default is \code{1}}

\item{point_border_colors}{Color for line surrounding points. Default is \code{"black"}}

\item{point_alpha}{Alpha value for point fill. Default is \code{1}}

\item{y_scale}{Scaling for y-axis. Options: \code{"regular"} (default) or \code{"log"}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0.1}}

\item{y_axis_label_fn}{Vector or function specifying y axis tick labels. Default is \code{axis_label_numeric.} Use axis_label_x10 for 10^x labels}

\item{y_axis_title}{Enter as quoted string. Default is \code{""}}

\item{x_axis_title}{Enter as quoted string. Default is \code{"Days post-infection"}}

\item{x_axis_breaks}{Enter as numeric vector of x axis breaks}

\item{breaks_fn}{Function used to generate y axis breaks. Default is \code{pretty}}

\item{x_min}{Minimum value to use for x axis}

\item{x_max}{Maximum value to use for x axis}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{4}}

\item{show_legend}{If \code{TRUE}, legend displayed. If \code{FALSE}, legend not displayed}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{error_fill_color}{Color of shaded error region. Default is \code{"grey60"}}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\value{
Plot with loess smoothed means for each group
}
\description{
Plot regression line or loess smoothed mean per group
}
