% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spaghetti.R
\name{plot_spaghetti}
\alias{plot_spaghetti}
\title{Spaghetti plot}
\usage{
plot_spaghetti(
  df,
  formula,
  id,
  line_color_var = NULL,
  line_color_categorical = TRUE,
  line_color = "#3B3B3B",
  line_thickness = 1,
  y_scale = "regular",
  n_breaks = 3,
  y_axis_breaks = NULL,
  y_min = NULL,
  y_max = NULL,
  expand_y = 0.1,
  y_axis_labels = NULL,
  y_axis_title = NULL,
  expand_x = 0.1,
  x_axis_title = NULL,
  plot_title = "",
  show_legend = FALSE,
  censor_fn = rescale_none,
  ...
)
}
\arguments{
\item{df}{Data frame in long format with columns for id, y (continuous variable), x (continuous variable)}

\item{formula}{y ~ x format}

\item{id}{Variable to identify subject. Enter as quoted variable name}

\item{line_color_var}{Variable used to determine line color. Enter as quoted variable name}

\item{line_color_categorical}{If \code{TRUE} (default), line color variable determined by \code{line_color_var} is set to factor (which uses a categorical palette). If \code{FALSE}, \code{line_color_var} will be converted to a continuous variable}

\item{line_color}{Color for lines. Hexadecimal or quoted color names. Default is \code{"black"}}

\item{line_thickness}{Thickness of line in mm Default is \code{1}}

\item{y_scale}{Scaling for y-axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_min}{Minimum value to use for y axis Default is \code{NULL}}

\item{y_max}{Maximum value to use for y axis Default is \code{NULL}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0.1}}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{y_axis_title}{Enter as quoted string. Default is \code{NULL}}

\item{expand_x}{Expansion of x axis at ends. Default is \code{0.1}}

\item{x_axis_title}{Enter as quoted string. Default is \code{NULL}}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed. If \code{TRUE}, legend for lines is shown}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\description{
Spaghetti plot
}
