% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin plot}
\usage{
plot_violin(
  df,
  formula,
  grouping_var = NULL,
  rev_grouping_var_order = FALSE,
  x_order = NULL,
  fn_x_order = mean,
  rev_x_order = FALSE,
  x = NULL,
  y = NULL,
  colors = c("#2A2D34", "#00A1D5", "#D75725", "#6761A8", "#009872"),
  color_border = "black",
  border_color = color_border,
  alpha = 1,
  width = 0.7,
  border_thickness = 0.75,
  show_legend = FALSE,
  y_scale = "regular",
  y_title = NULL,
  y_axis_title = y_title,
  x_title = NULL,
  x_axis_title = x_title,
  plot_title = NULL,
  n_breaks = 3,
  x_breaks = NULL,
  x_axis_breaks = x_breaks,
  x_labels = NULL,
  x_axis_labels = x_labels,
  y_breaks = NULL,
  y_axis_breaks = y_breaks,
  y_labels = NULL,
  y_axis_labels = y_labels,
  y_max = NULL,
  y_min = NULL,
  expand_y = 0.1,
  trim = TRUE,
  lines = c(0.25, 0.5, 0.75),
  linetype = "solid",
  linetype_quantile = "dotted",
  aspect_ratio = 1,
  censor_fn = rescale_none,
  show_points = FALSE,
  point_method = c("jitter", "beeswarm"),
  beeswarm_method = "smiley",
  point_size = 1,
  point_colors = "black",
  point_shapes = 19,
  point_width = width * 0.15,
  dodge = width * 1.1,
  n_bins = 30,
  band_width = n_bins/10,
  scaling = "width",
  ...
)
}
\arguments{
\item{df}{Data frame in long format}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Must have only 2 levels. Enter as quoted or unquoted variable name}

\item{rev_grouping_var_order}{If \code{FALSE} (default), levels of grouping_var ordered alphabetically. If \code{TRUE}, order of grouping_var levels is reversed}

\item{x_order}{Order of x values along x axis. Options: \code{NULL} (default), \code{"ascending"}, \code{"descending"}, or list of quoted values default is \code{NULL}}

\item{fn_x_order}{Function applied to y to determine order of x. Default is \code{mean}}

\item{rev_x_order}{Whether to reverse order of x values along x axis. Default is \code{FALSE}}

\item{x, y}{Variables for x and y axis, respectively. Enter as quoted or unquoted variable names}

\item{colors}{Fill color for violin plots. Enter as hexadecimal or quoted color name(s)}

\item{color_border}{Color of lines surrounding violin Default is \code{"black"}}

\item{border_color}{Alias for \code{color_border}}

\item{alpha}{Alpha value for violin fill. Default is \code{1}}

\item{width}{Width of violins. Default is \code{0.8}}

\item{border_thickness}{Thickness of line surrounding violin in pt units. Default is \code{0.75}}

\item{show_legend}{If \code{FALSE} (default), no legend displayed. If \code{TRUE}, legend is displayed}

\item{y_scale}{Scaling for y-axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_title}{Enter as quoted string. Default is \code{""}}

\item{y_axis_title}{Alias for \code{y_title}}

\item{x_title}{Enter as quoted string. Default is \code{""}}

\item{x_axis_title}{Alias for \code{x_title}}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{x_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_axis_breaks}{Alias for \code{x_breaks}}

\item{x_labels}{Vector or function specifying x axis tick labels}

\item{x_axis_labels}{Alias for \code{x_labels}}

\item{y_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_breaks}{Alias for \code{y_breaks}}

\item{y_labels}{Vector or function specifying y axis tick labels}

\item{y_axis_labels}{Alias for \code{y_labels}}

\item{y_max}{Maximum y axis value. Default is \code{max(df_subgrouped$ms)}}

\item{y_min}{Minimum value to use for y axis Default is \code{0}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0.1}}

\item{trim}{If \code{TRUE} (default), violin tails are trimmed. If \code{FALSE}, violin tails not trimmed.}

\item{lines}{Quantile line (horizontal) in violin. Default displays lines at 25th, 50th, and 75th percentile. Enter \code{NULL} to remove lines}

\item{linetype}{Line type for border of violins. Options include \code{"solid"} (default), \code{"dashed"}, \code{"dotted"}, \code{"longdash"}. Enter as quoted or unquoted linetype}

\item{linetype_quantile}{Line type for quantile lines. Options include \code{"solid"} (default), \code{"dashed"}, \code{"dotted"}, \code{"longdash"}. Enter as quoted or unquoted linetype}

\item{aspect_ratio}{Plot aspect ratio. Default is \code{1}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{show_points}{If \code{FALSE} (default), points not displayed. If \code{TRUE}, points displayed}

\item{point_method}{Type of points. Options include \code{"jitter"} (default) or \code{"beeswarm"}. Enter as quoted type}

\item{beeswarm_method}{Method for \code{geom_quasirandom} function. Default is \code{"smiley"}}

\item{point_size}{Size of points. Default is \code{1}}

\item{point_colors}{Color of points. Default is \code{"black"}}

\item{point_shapes}{Shape of points. Default is solid circle (no fill color)}

\item{point_width}{Width of point spread. Default is width of violin*0.15}

\item{dodge}{Argument for position dodge. Default is width*1.1}

\item{n_bins}{Number of bins. Used in \code{ggbeeswarm::geom_quasirandom}. Default is \code{30}}

\item{band_width}{Bandwidth. Used in \code{ggbeeswarm::geom_quasirandom}. Default is \code{n_bins/10}}

\item{scaling}{Method used to make violin size equal across groups. Options: \code{"width"} (default), \code{"area"}, \code{"count"}}

\item{...}{Arguments passed to theme_custom}
}
\description{
Violin plot
}
