% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(
  df,
  var = "gene",
  fc = "fc",
  pval = "p",
  cond_color = NULL,
  cond_label = NULL,
  point_size = 2.5,
  point_alpha = 0.7,
  point_border_thickness = 0.5,
  colors = c("#BC3C29", "#3B3B3B"),
  fc_threshold = 2,
  pval_threshold = 0.01,
  line_type = "dashed",
  line_color = "#909090",
  line_thickness = 0.5,
  x_axis_title = expression("log"[2] * " fold change"),
  y_axis_title = expression("-log"[10] * " " * italic("P") * " value"),
  x_axis_breaks = waiver(),
  y_axis_breaks = waiver(),
  expand_x = c(0.3, 0.3),
  expand_y = c(0.05, 0.05),
  str_clean_fn = function(x) str_capitalize(x, other_letters_lower = TRUE),
  italic = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame containing columns for fc, p value, and variable name}

\item{var}{Variable in df containing rownames (gene names if expression data). Enter as quoted variable name. Default is \code{"gene"}}

\item{fc}{Variable in df containing fold change values. Enter as quoted variable name. Default is \code{"fc"}}

\item{pval}{Variable in df containing p values. Enter as quoted variable name. Default is \code{"p"}}

\item{cond_color}{Conditional statement used to color data points. Can also enter comma separated list of quoted values for var wrapped in \code{c()}}

\item{cond_label}{Conditional statement used to select which data points to label. Can also enter comma separated list of quoted values for var wrapped in \code{c()}}

\item{point_size}{Size of points in mm. Default is \code{2.5}}

\item{point_alpha}{Transparency of points. Default is \code{0.7}}

\item{point_border_thickness}{Thickness of point border. Default is \code{0.5}}

\item{colors}{Colors used to fill data points}

\item{fc_threshold}{Threshold fc value. Used to generated vertical line. If \code{NULL}, no line drawn}

\item{pval_threshold}{Threshold p value. Used to generated horizontal line. If \code{NULL}, no line drawn}

\item{line_type}{Line type for threshold lines. Default is \code{"dashed"}}

\item{line_color}{Color for threshold lines. Default is \code{"#909090"}}

\item{line_thickness}{Thickness for threshold lines in lines units. Default is \code{0.5}}

\item{x_axis_title}{Title for x axis. Default is \code{"Fold change"}}

\item{y_axis_title}{Title for y axis. Default is \code{"-log10 P value"}}

\item{x_axis_breaks, y_axis_breaks}{Breaks for x and y axis respectively}

\item{expand_x, expand_y}{Expansion for x and y axis respectively. Enter as length 2 numeric vector for \% expansion of lower and upper end of axis respectively. Default uses 5\% for y axis and 30\% for x axis}

\item{str_clean_fn}{Function applied to values of var. Default capitalizes entries}

\item{italic}{If \code{TRUE} (default), var labels converted to italic}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\description{
Volcano plot
}
