% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_ppt.R
\name{ppt}
\alias{ppt}
\title{Export 1 or more plots to powerpoint}
\usage{
ppt(
  plot_object,
  file_name = NULL,
  directory = getwd(),
  height = NULL,
  width = NULL,
  y = NULL,
  top = y,
  x = NULL,
  left = x,
  editable = TRUE,
  font_family = "Helvetica",
  landscape = "horizontal",
  slide_layout = "Plot_center",
  theme = "default_helvetica",
  fullsize = FALSE,
  fit_within_slide = TRUE
)
}
\arguments{
\item{plot_object}{Plot or list of plots}

\item{file_name}{File name for plot. Enter as quoted file name}

\item{directory}{Path to directory where file should be saved. Default is working directory}

\item{height, width}{Plot height and width (in inches) respectively. Default is \code{5} for both}

\item{y}{Vertical position (in inches) of top of plot relative to top of slide. Default is centered}

\item{top}{Alias for \code{y}}

\item{x}{Position (in inches) of leftmost edge of plot relative to leftmost edge of slide. Default is centered}

\item{left}{Alias for \code{x}}

\item{editable}{If \code{TRUE} (default), plot will be stored as svg. If \code{FALSE}, plot will be stored as png}

\item{font_family}{Font family. Default is \code{"Helvetica"}. Enter as quoted or unquoted font name}

\item{landscape}{Layout of plot in powerpoint. Options: \code{"horizontal"} (default), \code{"vertical"}}

\item{slide_layout}{Options: \code{"Plot_center"} (default, centered 5" x 5" plot), \code{"Plot_center_small"} (centered 3.5" x 3.5" plot) \code{"Plot_5_5"} (5" x 5" plot on left edge of slide), \code{"Plot_7_7"} (7" x 7" plot on left edge of slide), \code{"Plot_5_7"} (5" tall, 7" wide plot on left edge of slide), \code{"Plot"} (5" tall, 11.5" wide plot on left edge of slide). Enter as quoted slide layout}

\item{theme}{Name of theme (slide master). Must be a theme present in pptx file \code{.pptx_template_path}}

\item{fullsize}{If \code{FALSE} (default), plot size according to entered or default parameters. If \code{TRUE}, plot printed to entire size of slide}

\item{fit_within_slide}{If \code{TRUE} (default), attempt is made to fit plot within slide boundaries. If FALSE, plot can extend beyond slide limits}
}
\description{
Functionality from David Gohel's excellent package officer
}
