% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_ppt.R
\name{ppt_by_code}
\alias{ppt_by_code}
\title{Export plot to powerpoint using plot code}
\usage{
ppt_by_code(
  ...,
  file_name = "plot",
  directory = getwd(),
  height = 5,
  width = 5,
  top = 1.67,
  left = 1,
  font_family = "Helvetica",
  slide_layout = "Plot_center",
  fullsize = FALSE
)
}
\arguments{
\item{...}{Code to generate plot}

\item{file_name}{File name for plot. Enter as quoted file name}

\item{directory}{Directory where plot will be saved. Default is working directory}

\item{height, width}{Plot height and width (in inches) respectively. Default is \code{5} for both}

\item{top}{Alias for \code{y}}

\item{left}{Alias for \code{x}}

\item{font_family}{Font family. Default is \code{"Helvetica"}. Enter as quoted or unquoted font name}

\item{slide_layout}{Options: \code{"Plot_center"} (default, centered 5" x 5" plot), \code{"Plot_center_small"} (centered 3.5" x 3.5" plot) \code{"Plot_5_5"} (5" x 5" plot on left edge of slide), \code{"Plot_7_7"} (7" x 7" plot on left edge of slide), \code{"Plot_5_7"} (5" tall, 7" wide plot on left edge of slide), \code{"Plot"} (5" tall, 11.5" wide plot on left edge of slide). Enter as quoted slide layout}

\item{fullsize}{If \code{FALSE} (default), plot size according to entered or default parameters. If \code{TRUE}, plot printed to entire size of slide}
}
\description{
Functionality from David Gohel's excellent package officer
}
