% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_pdf_png_jpg.R
\name{preview}
\alias{preview}
\title{Open plot in temporary file to preview output}
\usage{
preview(
  plot_object,
  height = 5,
  width = 5,
  y = 1.67,
  top = y,
  x = 1,
  left = x,
  font_family = "Helvetica",
  ext = "pptx",
  slide_layout = "Plot_center",
  theme = "default_helvetica",
  fullsize = FALSE,
  cairo = FALSE
)
}
\arguments{
\item{plot_object}{Plot or list of plots}

\item{height, width}{Plot height and width (in inches) respectively. Default is \code{5} for both}

\item{y}{Vertical position (in inches) of top of plot relative to top of slide. Default is centered}

\item{top}{Alias for \code{y}}

\item{x}{Position (in inches) of leftmost edge of plot relative to leftmost edge of slide. Default is centered}

\item{left}{Alias for \code{x}}

\item{font_family}{Font family. Default is \code{"Helvetica"}. Enter as quoted or unquoted font name}

\item{ext}{Extension for output file. Default is \code{"pptx"}}

\item{slide_layout}{Options: \code{"Plot_center"} (default, centered 5" x 5" plot), \code{"Plot_center_small"} (centered 3.5" x 3.5" plot) \code{"Plot_5_5"} (5" x 5" plot on left edge of slide), \code{"Plot_7_7"} (7" x 7" plot on left edge of slide), \code{"Plot_5_7"} (5" tall, 7" wide plot on left edge of slide), \code{"Plot"} (5" tall, 11.5" wide plot on left edge of slide). Enter as quoted slide layout}

\item{theme}{Name of theme (slide master). Must be a theme present in pptx file \code{.pptx_template_path}}

\item{fullsize}{If \code{FALSE} (default), plot size according to entered or default parameters. If \code{TRUE}, plot printed to entire size of slide}

\item{cairo}{If \code{FALSE} (default), cairo device not used. If \code{TRUE}, cairo device is used}
}
\description{
Open plot in temporary file to preview output
}
