% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_prism.R
\name{prism_sheets}
\alias{prism_sheets}
\title{Get names of sheets in prism file}
\usage{
prism_sheets(file_path)
}
\arguments{
\item{file_path}{Path to file}
}
\description{
Functionality from Yue Jiang's excellent package pzfx
}
