% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{q_value}
\alias{q_value}
\alias{q_value.default}
\alias{q_value.data.frame}
\title{Generic q_value function}
\usage{
q_value(x, p_column = "p", q_col_name = "q_val", ...)

\method{q_value}{default}(x, p_column = "p", q_col_name = "q_val", ...)

\method{q_value}{data.frame}(x, p_column = "p", q_col_name = "q_val", ...)
}
\arguments{
\item{x}{Data frame or numeric vector}

\item{p_column}{Column containing p values. Enter as quoted column name. Default is \code{p}}

\item{q_col_name}{Column name for q values. Enter as quoted column name. Default is \code{q_val}}

\item{...}{Arguments passed to class-specific \code{q_value} function}
}
\value{
Numeric vector of q values
}
\description{
Generic q_value function

q values - default method

q values - data frame
}
