% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{Quantile}
\alias{Quantile}
\title{Quantile}
\usage{
Quantile(x, probs = c(0, 0.25, 0.5, 0.75, 1), type = 7)
}
\arguments{
\item{x}{Numeric vector}

\item{probs}{Probabilities. Enter numeric 0-1. Default uses min, Q1, median, Q3, max}

\item{type}{Type of quantile algorithm. Options: \code{7} (default), \code{6} (used by Prism)}
}
\value{
Numeric vector of length equal to the length of probs
}
\description{
Functionality from \code{stats::quantile}
}
