% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_misc.R
\name{random_number}
\alias{random_number}
\title{Random number}
\usage{
random_number(min = 1, max = 100)
}
\arguments{
\item{min, max}{Minimum and maximum values allowed for random number}
}
\value{
Length 1 integer
}
\description{
Random number
}
