% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_prism.R
\name{read_prism}
\alias{read_prism}
\title{Import prism file}
\usage{
read_prism(file_name, ..., table = NULL, desktop = TRUE)
}
\arguments{
\item{file_name}{Name of Prism file. Enter as quoted or unquoted name (later only if no spaces in name)}

\item{...}{Arguments passed to \code{paste_path()}}

\item{table}{Table in Prism file to import. Enter quoted table name or integer indicating position of table. Default imports all tables}

\item{desktop}{If \code{TRUE} (default), files can be uploaded from desktop. If \code{FALSE}, files can't be uploaded from desktop. Use \code{FALSE} to upload file from working directory when a file with the same name also exists on the desktop}
}
\description{
Functionality from Yue Jiang's excellent package pzfx
}
