% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{rel_risk}
\alias{rel_risk}
\alias{rr_ci}
\title{Calculate RR, 95\% CI, and P value}
\usage{
rel_risk(
  ...,
  ci = 0.95,
  method = c("both", "best", "chi", "fisher"),
  n_min_chisq = 5
)

rr_ci(
  ...,
  ci = 0.95,
  method = c("both", "best", "chi", "fisher"),
  n_min_chisq = 5
)
}
\arguments{
\item{...}{Enter raw counts with or without c(), 2 x 2 matrix or table. If x is a vector of counts, enter as exposure-/outcome-, exposure-/outcome+, exposure+/outcome-, exposure+/outcome+ (control/outcome-, control/outcome+, case/outcome-, case/outcome+). Entered into matrix by column. For matrix input, columns = predictor/exposure/risk factor/case vs. control (- or control in col 1, + or case in col 2), rows = truth/outcome/disease (outcome- in row 1, outcome+ in row 2)}

\item{ci}{Confidence interval. Default is \code{0.95}}

\item{method}{Function for testing. Options: \code{"best"} (default), \code{"chi"}, \code{"fisher"}}

\item{n_min_chisq}{Minimum group size that will be allowed for chi-squared test. Default is \code{5}. Only relevant when \code{method = "best"}}
}
\value{
Data frame containing columns for rr, rr_lower, rr_upper, and either p, label, method or p_fisher and p_chi
}
\description{
Calculate RR, 95\% CI, and P value

Alias for rel_risk
}
