% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{remove_axis}
\alias{remove_axis}
\title{Remove axis}
\usage{
remove_axis(axis = c("x", "y"))
}
\arguments{
\item{axis}{Options: \code{"x"}, \code{"y"}, or both (default)}
}
\value{
theme object with axis lines removed. Enter as plot + remove_axis("x")
}
\description{
Remove axis
}
