% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{remove_axis_lines}
\alias{remove_axis_lines}
\title{Remove axis lines}
\usage{
remove_axis_lines(axis = c("x", "y"))
}
\arguments{
\item{axis}{Options: \code{"x"}, \code{"y"}, or both (default)}
}
\value{
theme object with axis lines removed. Enter as plot + remove_axis_lines("x")
}
\description{
Remove axis lines
}
