% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Remove rows which contain missing values in every column}
\usage{
remove_empty_rows(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Columns to be searched for missing values. Enter using tidyselect syntax. If empty, all columns searched}
}
\value{
\code{df} without rows containing exclusively NA in the columns specified by \code{...}
}
\description{
Remove rows which contain missing values in every column
}
