% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{remove_geoms}
\alias{remove_geoms}
\title{Remove main geoms from plot}
\usage{
remove_geoms(x, geom_pattern = NULL, nth_instance = NULL)
}
\arguments{
\item{x}{ggplot object}

\item{geom_pattern}{Pattern in geom class to search in layers. If \code{NULL} (default), all geoms are removed}

\item{nth_instance}{Integer vector indicating indices of layers with \code{geom_pattern} that should be removed. If \code{NULL} (default), all instances are removed}
}
\description{
Remove main geoms from plot
}
