% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{remove_na}
\alias{remove_na}
\title{Remove rows with missing values}
\usage{
remove_na(df, cols = names(df))
}
\arguments{
\item{df}{Data frame}

\item{cols}{Columns to search for missing values. Enter as character vector. Default uses all columns}
}
\value{
\code{df} without rows containing missing values in the columns specified by cols argument
}
\description{
Similar to \code{tidyr::drop_na}
}
