% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{remove_null}
\alias{remove_null}
\title{Remove NULL values from a list}
\usage{
remove_null(x)
}
\arguments{
\item{x}{List}
}
\value{
List without \code{NULL} or other length 0 elements
}
\description{
Remove NULL values from a list
}
