% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{remove_plot_element}
\alias{remove_plot_element}
\title{Remove theme element from plot}
\usage{
remove_plot_element(x)
}
\arguments{
\item{x}{Character vector of \code{ggplot2::theme()} arguments to remove from plot}
}
\value{
theme object with \code{x} removed. Enter as plot + remove_plot_element("axis.line.x")
}
\description{
Remove theme element from plot
}
