% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{remove_plot_margins}
\alias{remove_plot_margins}
\title{Remove plot margins}
\usage{
remove_plot_margins(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
ggplot object with no margins
}
\description{
Remove plot margins
}
