% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{replace_na}
\alias{replace_na}
\title{Replace missing values}
\usage{
replace_na(df, ..., new_value = 0)
}
\arguments{
\item{df}{Data frame or vector}

\item{...}{Columns to replace \code{NA} with \code{new_value}. Enter using tidyselect syntax. Default uses all columns. Only relevant when \code{df} is a data frame}

\item{new_value}{Value used to replace missing values. Default is \code{0}}
}
\description{
Replace missing values
}
