% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{replace_values}
\alias{replace_values}
\title{Replace values in character or factor vector using lookup}
\usage{
replace_values(x, replacements)
}
\arguments{
\item{x}{Character or factor vector}

\item{replacements}{Named character vector. Names match elements in \code{x} that will be replaced. Values are replacements}
}
\value{
Vector with same class as input
}
\description{
Functionality from ggplot2
}
