% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{rescale_none}
\alias{rescale_none}
\title{Censor function}
\usage{
rescale_none(x, ...)
}
\arguments{
\item{x}{Vector}

\item{...}{Not used}
}
\value{
Input
}
\description{
Functionality from scales package
}
