% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R, R/plot_manipulation.R
\name{rotate_x_labels}
\alias{rotate_x_labels}
\title{Rotate x axis labels}
\usage{
rotate_x_labels(angle = 45)

rotate_x_labels(angle = 45)
}
\arguments{
\item{angle}{Angle to rotate x axis labels. Default is \code{45}}
}
\value{
Add to ggplot object just as \code{theme()} is used

ggplot object with x axis labels rotated \code{angle} degrees. Enter as plot + rot_x_labs()
}
\description{
Rotate x axis labels

Rotate x axis labels
}
