% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{round_up}
\alias{round_up}
\title{Round up for 5 or higher}
\usage{
round_up(x, digits = 2)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of digits after decimal place to include in output}
}
\description{
Functionality from Tyler Rinker's excellent package numform
}
