% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{row_apply}
\alias{row_apply}
\title{Row-wise summary statistic}
\usage{
row_apply(df, fn, ..., new_col_name = "row_value")
}
\arguments{
\item{df}{Data frame}

\item{fn}{Summary function. Must take a vector of values and return a single number}

\item{...}{Columns used in calculation. Enter using tidyselect syntax. Default uses all columns}

\item{new_col_name}{Name for column containing row-wise statistic. Default is \code{"row_value"}}
}
\value{
Data frame with column for row-wise summary statistic
}
\description{
Row-wise summary statistic
}
