% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{row_prop}
\alias{row_prop}
\title{Convert counts in contingency table to row proportions}
\usage{
row_prop(x, ..., incl_na = FALSE)
}
\arguments{
\item{x}{Contingency table or data frame}

\item{...}{If \code{x} is a data frame, enter column names using tidyselect syntax}

\item{incl_na}{If \code{TRUE} (default), missing values are included in output}
}
\value{
Table containing row proportions (i.e., row proportions sum to 1)
}
\description{
Convert counts in contingency table to row proportions
}
