% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{sample_size_binary}
\alias{sample_size_binary}
\title{Sample size calculation, binary outcome}
\usage{
sample_size_binary(p1, p2, beta = 0.2)
}
\arguments{
\item{p1, p2}{Proportion with outcome for groups 1 and 2, respectively}

\item{beta}{power = 1 - beta. Default is \code{0.2}}
}
\value{
Number of subjects required to reach signifance
}
\description{
Sample size calculation, binary outcome
}
