% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{sample_size_continuous}
\alias{sample_size_continuous}
\title{Sample size calculation, continuous outcome}
\usage{
sample_size_continuous(mean1, sd1, mean2, sd2, beta = 0.2)
}
\arguments{
\item{mean1, mean2}{Mean value for groups 1 and 2, respectively}

\item{sd1, sd2}{sd for groups 1 and 2, respectively}

\item{beta}{power = 1 - beta. Default is \code{0.2}}
}
\value{
Number of subjects required to reach signifance
}
\description{
Sample size calculation, continuous outcome
}
