% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{scale_axis_clean}
\alias{scale_axis_clean}
\title{Input to guide in scale functions to generate clean axis}
\usage{
scale_axis_clean(
  axis = "y",
  scale = "regular",
  plot_limits,
  axis_limits,
  title = NULL,
  breaks = NULL,
  breaks_fn = pretty,
  n_breaks = 4,
  labels = NULL,
  expand_lower = 0,
  expand_upper = 0,
  position = NULL,
  censor_fn = rescale_none,
  ...
)
}
\arguments{
\item{axis}{Options include \code{"y"} (default) or \code{"x"}}

\item{scale}{Options include \code{"regular"} (default), \code{"scientific"}, \code{"log10"}, \code{"log2"}}

\item{plot_limits}{Limits of data for plotting all objects in plotting space}

\item{axis_limits}{Limits used to generate axis breaks}

\item{title}{Axis title}

\item{breaks}{Axis breaks}

\item{breaks_fn}{Function used to create breaks. Enter as call using (). Default is \code{pretty}. Alternative is \code{scales::breaks_extended(n = n_breaks, only.loose = TRUE)}}

\item{n_breaks}{Desired number of axis breaks. Default is \code{4}}

\item{labels}{Vector or function specifying axis tick labels}

\item{expand_lower, expand_upper}{Expansion to add to ends of axis}

\item{position}{Location of axis. Options: \code{"left"} (default for y axis), \code{"right"}, "bottom" (default for x axis), "top"}

\item{censor_fn}{Function used to transform data outside axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{...}{Arguments passed to scale function}
}
\description{
Needed to place significance annotation outside of plot limits
}
