% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{scale_fill_new}
\alias{scale_fill_new}
\alias{scale_color_new}
\title{New fill scale}
\usage{
scale_fill_new()

scale_color_new()
}
\value{
Called directly by user
}
\description{
Functionality from ggnewscale package

Functionality from ggnewscale package
}
