% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{select_binary}
\alias{select_binary}
\alias{select_binary_01}
\alias{select_fct}
\alias{select_chr}
\alias{select_int}
\alias{select_can_be_numeric}
\alias{select_lgl}
\alias{select_date}
\title{Data frame with binary columns only}
\usage{
select_binary(df, invert = FALSE)

select_binary_01(df, invert = FALSE)

select_fct(df, invert = FALSE)

select_chr(df, invert = FALSE)

select_int(df, invert = FALSE)

select_can_be_numeric(df, invert = FALSE)

select_lgl(df, invert = FALSE)

select_date(df, invert = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\description{
Data frame with binary columns only

Data frame with binary columns coded as 0 or 1 only

Data frame with factor columns only

Data frame with character columns only

Data frame with integer columns only

Data frame with columns coercible to a numeric only

Data frame with logical columns only

Data frame with date columns only
}
