% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{select_id_vars}
\alias{select_id_vars}
\title{Data frame with id columns only}
\usage{
select_id_vars(df, na.rm = FALSE, invert = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{na.rm}{If \code{TRUE}, missing values are not considered}

\item{invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\description{
Data frame with id columns only
}
