% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{select_names_which}
\alias{select_names_which}
\title{Select data frame columns with names matching a specified pattern}
\usage{
select_names_which(df, pattern, ..., search_fn = NULL)
}
\arguments{
\item{df}{Data frame, named vector, or named list}

\item{pattern}{Regex pattern to search in column names}

\item{...}{Arguments passed to \code{search_fn}}

\item{search_fn}{Function used to search column names for \code{pattern}. Default is \code{str_contains} if 1 term is entered as \code{pattern} and \code{str_contains_all} if > 1 term is entered. Function must have arguments \code{x} and \code{pattern}}
}
\value{
Data frame with columns in which names contain \code{pattern}
}
\description{
Select data frame columns with names matching a specified pattern
}
