% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{select_not_constant}
\alias{select_not_constant}
\title{Remove constant columns}
\usage{
select_not_constant(df, na.rm = TRUE)
}
\arguments{
\item{df}{Data frame}

\item{na.rm}{If \code{TRUE} (default), NA are not considered in determining whether a column is constant}
}
\value{
Data frame with constant columns removed
}
\description{
Remove constant columns
}
