% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{select_which}
\alias{select_which}
\title{Get a subset of data frame columns according to a predicate}
\usage{
select_which(.df, .predicate_fn, ..., .invert = FALSE)
}
\arguments{
\item{.df}{Data frame}

\item{.predicate_fn}{Predicate function to be applied to each column in \code{.df}. Must evaluate to a length 1 logical (not \code{NULL}), thus non-vectorized functions only}

\item{...}{Arguments passed to \code{.predicate_fn}}

\item{.invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\value{
Data frame with columns meeting predicate criterion
}
\description{
Get a subset of data frame columns according to a predicate
}
