% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{set_max_time_follow_up}
\alias{set_max_time_follow_up}
\alias{set_max_follow_up_time}
\title{Set maximum duration of follow up time}
\usage{
set_max_time_follow_up(df, time_var = "time", max_time, outcome_var = "death")

set_max_follow_up_time(df, time_var = "time", max_time, outcome_var = "death")
}
\arguments{
\item{df}{Data frames}

\item{time_var}{Variable containing follow up time. Enter as quoted variable name. Default is \code{"time"}}

\item{max_time}{Maximum duration of follow up. Enter as numeric}

\item{outcome_var}{Variable containing outcome variable coded as \code{1} (event) or \code{0} (no event/censored). Enter as quoted variable name. Default is \code{"death"}}
}
\description{
Set maximum duration of follow up time

Alias for set_max_time_follow_up
}
