% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{set_names}
\alias{set_names}
\title{Set names of a vector}
\usage{
set_names(x, names = x)
}
\arguments{
\item{x}{Vector}

\item{names}{Character vector of names (or coercible to a character vector). If length of names is shorter than length of \code{x}, names are recycled. Default entry will name each component of \code{x} with it's value}
}
\value{
Named version of input
}
\description{
Set names of a vector
}
