% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{set_plot_labels}
\alias{set_plot_labels}
\title{Edit labels for 1 or more plots}
\usage{
set_plot_labels(
  plot_list,
  ...,
  indices = seq_along(plot_list),
  action = "replace",
  sep = " "
)
}
\arguments{
\item{plot_list}{List of ggplot objects}

\item{...}{Enter name-value pair where name refer to plot component to relabel ("x", "y", "title") and values are new labels to be added entered as quoted string(s). If number of new labels is less than number of plots, terms will be recycled}

\item{indices}{Plots to apply new new labels to entered as number index. Default includes all plots}

\item{action}{Method for creating new plot label. Options: \code{"replace"} (default), \code{"paste_after"}, \code{"paste_before"}. Enter as quoted or unquoted action}

\item{sep}{Text used to separate prior label from new label. Default is \code{" "}}
}
\value{
List of ggplot objects with updated labels
}
\description{
Functionality from package kwb.plot
}
