% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{Setdiff}
\alias{Setdiff}
\title{Elements in x but not y}
\usage{
Setdiff(x, y)
}
\arguments{
\item{x, y}{Vectors}
}
\description{
Similar to \code{base::setdiff} but doesn't convert \code{x} and \code{y} to vectors
}
