% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{setdiff_all}
\alias{setdiff_all}
\title{Setdiff}
\usage{
setdiff_all(...)
}
\arguments{
\item{...}{Either a comma separated list of vectors not wrapped in \code{c()} or a list containing vectors}
}
\description{
Equivalent to sequentially running \code{Setdiff} on > 2 vectors
}
