% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{sig_stars}
\alias{sig_stars}
\title{Convert p values into significance stars}
\usage{
sig_stars(
  p,
  cutpoints = c(0, 1e-04, 0.001, 0.01, 0.05, 1),
  symbols = c("****", "***", "**", "*", "ns")
)
}
\arguments{
\item{p}{Numeric vector of p values}

\item{cutpoints}{Cutpoints for significance stars. Enter as numeric vector including 0 and 1}

\item{symbols}{Symbols representing significance levels from lowest to highest P value. Enter as character vector}
}
\value{
Character vector containing significance stars
}
\description{
Convert p values into significance stars
}
