% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{Skew}
\alias{Skew}
\alias{Kurtosis}
\title{Skewness}
\usage{
Skew(x, type = 1)

Kurtosis(x, type = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{type}{Method used to calculate skewness. Options: \code{1} (default), \code{2}, \code{3}}
}
\description{
Skewness

Kurtosis
}
