% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{sort_by_length}
\alias{sort_by_length}
\title{Sort strings by number of characters}
\usage{
sort_by_length(x, desc = FALSE)
}
\arguments{
\item{x}{Vector}

\item{desc}{If \code{FALSE} (default), \code{x} is order from shortest to longest string. If \code{TRUE}, \code{x} is order from longest to shortest string}
}
\description{
Sort strings by number of characters
}
