% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{sort_by_name}
\alias{sort_by_name}
\title{Sort vector or columns in a data frame alphabetically by names of elements}
\usage{
sort_by_name(x, blank_last = TRUE, reverse = FALSE)
}
\arguments{
\item{x}{Named vector, named list, or data frame}

\item{blank_last}{If \code{TRUE} (default), elements without names will be placed at the end}

\item{reverse}{If \code{TRUE}, names ordered reverse alphabetically}
}
\description{
Sort vector or columns in a data frame alphabetically by names of elements
}
