% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{squish}
\alias{squish}
\title{Squish a vector into a given range}
\usage{
squish(x, range = c(0, 1))
}
\arguments{
\item{x}{Numeric vector}

\item{range}{Desired range. Enter as length 2 numeric vector. Default is \code{c(0, 1)}}
}
\value{
Transformed numeric vector with range defined by range argument
}
\description{
Functionality from \code{squish} function in scales package
}
