% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sina.R
\docType{data}
\name{StatSina}
\alias{StatSina}
\title{ggproto for sina plot}
\format{
An object of class \code{StatSina} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
StatSina
}
\description{
ggproto for sina plot
}
\keyword{datasets}
