% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_capitalize}
\alias{str_capitalize}
\title{Capitalize each word}
\usage{
str_capitalize(x, other_letters_lower = FALSE)
}
\arguments{
\item{x}{Character vector}

\item{other_letters_lower}{If \code{FALSE} (default), subsequent letters left as is. If \code{TRUE}, subsequent letters converted to lower case}
}
\value{
String with first letter capitalized (remaining letters not capitalized)
}
\description{
Capitalize each word
}
