% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_common_prefix}
\alias{str_common_prefix}
\alias{str_common_suffix}
\alias{str_remove_common}
\title{Identify the longest prefix shared by a character vector}
\usage{
str_common_prefix(x, ignore_case = FALSE)

str_common_suffix(x, ignore_case = FALSE)

str_remove_common(x, ignore_case = FALSE)
}
\arguments{
\item{x}{Character vector}

\item{ignore_case}{If \code{FALSE} (default), case of both \code{x} and \code{pattern} are ignored in search. If \code{TRUE}, case used in \code{x} and \code{pattern} must match}
}
\value{
Character vector of length 1 (when shared prefix exists) or 0 (no shared prefix)

Character vector of length 1 (when shared suffix exists) or 0 (no shared suffix)

Character vector with same length as input
}
\description{
Identify the longest prefix shared by a character vector

Identify the longest suffix shared by a character vector

Remove prefix/suffix shared by a character vector from each strings
}
