% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_contains}
\alias{str_contains}
\alias{str_contains_all}
\alias{str_contains_any}
\alias{str_contains_none}
\alias{str_empty}
\title{Detect pattern in string}
\usage{
str_contains(x, pattern, invert = FALSE, ignore_case = FALSE, ...)

str_contains_all(x, pattern, ignore_case = FALSE, ...)

str_contains_any(x, pattern, ignore_case = FALSE, ...)

str_contains_none(x, pattern, ignore_case = FALSE, ...)

str_empty(x)
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Substring to detect in \code{x.} Enter as character vector}

\item{invert}{If \code{TRUE}, search will be performed for values in \code{x} that do not contain \code{pattern}. Equivalent of \code{!str_contains(x)}}

\item{ignore_case}{If \code{FALSE} (default), case of both \code{x} and \code{pattern} are ignored in search. If \code{TRUE}, case used in \code{x} and \code{pattern} must match}

\item{...}{Arguments passed to \code{grepl}}
}
\value{
Logical vector with length same as input. \code{NA} values return \code{FALSE}
}
\description{
Equivalent of \code{!str_contains_any(x, pattern)}
}
