% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_count}
\alias{str_count}
\title{Count occurrence of substring}
\usage{
str_count(x, pattern, ignore_case = FALSE, ...)
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Pattern to count per element of \code{x}}

\item{ignore_case}{If \code{FALSE} (default), case of both \code{x} and \code{pattern} are ignored in search. If \code{TRUE}, case used in \code{x} and \code{pattern} must match}

\item{...}{Arguments passed to \code{gregexpr}}
}
\value{
Integer
}
\description{
Count occurrence of substring
}
