% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_date}
\alias{str_date}
\title{Extract dates}
\usage{
str_date(x, sep = "/", all = TRUE)
}
\arguments{
\item{x}{Character vector}

\item{sep}{Separator between month, day, year. Default is \code{"/"}}

\item{all}{If \code{TRUE}, all dates are extracted. If \code{FALSE} (default), only first date is extracted}
}
\value{
Date as character vector or list of character vectors
}
\description{
Extract dates
}
