% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_extract}
\alias{str_extract}
\title{Extract text following a particular pattern (1st occurrence only)}
\usage{
str_extract(x, pattern, all = TRUE, fixed = FALSE)
}
\arguments{
\item{x}{String}

\item{pattern}{Pattern to search for in \code{x}}

\item{all}{If \code{TRUE} (default), all matches for \code{pattern} will be extracted. If \code{FALSE}, only the first match for \code{pattern} will be extracted}

\item{fixed}{If \code{TRUE}, \code{pattern} is matched "as is"}
}
\value{
Substring
}
\description{
Extract text following a particular pattern (1st occurrence only)
}
