% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_extract_after}
\alias{str_extract_after}
\alias{str_remove_before}
\title{Extract text after a pattern}
\usage{
str_extract_after(x, pattern)

str_remove_before(x, pattern, all = FALSE)

str_extract_after(x, pattern)
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Pattern before text to be extracted. Enter as length 1 character vector}

\item{all}{If \code{FALSE} (default), only the first match for \code{pattern} will be removed. If \code{TRUE}, all matches for \code{pattern} will be removed}
}
\value{
Text in \code{x} after \code{pattern}

Character vector with length equal to input. If no match is found, \code{""} is returned. If multiple matches present, only the first is returned
}
\description{
If a string has more than one match, this will only return the first one
}
