% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_extract_before}
\alias{str_extract_before}
\alias{str_remove_after}
\title{Extract text before a pattern}
\usage{
str_extract_before(x, pattern)

str_remove_after(x, pattern, all = FALSE)

str_extract_before(x, pattern)
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Pattern that follows text to be extracted. Enter as length 1 character vector}

\item{all}{If \code{FALSE} (default), only the first match for \code{pattern} will be removed. If \code{TRUE}, all matches for \code{pattern} will be removed}
}
\value{
Text in \code{x} before \code{pattern}

Character vector with length equal to input. If no match is found, \code{""} is returned. If multiple matches present, only the first is returned
}
\description{
Extract text before a pattern

Alias for str_extract_before

Extract text that occurs before a specified pattern
}
