% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_extract_between}
\alias{str_extract_between}
\title{Extract text that occurs between two specified patterns}
\usage{
str_extract_between(x, before, after)
}
\arguments{
\item{x}{Character vector}

\item{before, after}{Pattern before and after text to be extracted. Enter as length 1 character vectors}
}
\value{
Character vector with length equal to input. If no match is found, \code{""} is returned. If multiple matches present, only the first is returned
}
\description{
Extract text that occurs between two specified patterns
}
