% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{str_file_ext}
\alias{str_file_ext}
\alias{file_ext}
\title{Extract file extension}
\usage{
str_file_ext(x)

file_ext(x)
}
\arguments{
\item{x}{Character vector of file paths}
}
\value{
Character vector of file extension not including "."
}
\description{
Functionality from \code{file_ext} function in tools package
}
