% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_first_n}
\alias{str_first_n}
\alias{str_last_n}
\title{Extract first n characters from string}
\usage{
str_first_n(x, n)

str_last_n(x, n)
}
\arguments{
\item{x}{String}

\item{n}{Number of characters to extract}
}
\description{
Extract first n characters from string

Extract last n characters from string
}
