% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{str_force_ext}
\alias{str_force_ext}
\alias{str_add_ext}
\alias{str_replace_ext}
\title{Ensure that string has specific file extension}
\usage{
str_force_ext(x, ext)

str_add_ext(x, ext)

str_replace_ext(x, ext)
}
\arguments{
\item{x}{Character vector of file paths with or without extension included}

\item{ext}{File extension with or without "."}
}
\value{
Character vector of file paths with same length as \code{x} containing \code{ext} as extension
}
\description{
Ensure that string has specific file extension

Alias for str_force_ext

Alias for str_force_ext
}
