% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_info}
\alias{str_info}
\title{Wrapper for grepl, grep, agrep}
\usage{
str_info(
  x,
  pattern,
  output = c("logical", "value", "index", "all"),
  invert = FALSE,
  ignore_case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  use_bytes = FALSE,
  fuzzy = FALSE
)
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Substring to detect in \code{x.} Enter as character vector}

\item{output}{Options: \code{"logical"} (default), \code{"value"}, \code{"index"}, \code{"all"} (list of length 3 with components "logical", "value", "index")}

\item{invert}{If \code{TRUE}, search will be performed for values in \code{x} that do not contain \code{pattern}. Equivalent of \code{!str_contains(x)}}

\item{ignore_case}{If \code{FALSE} (default), case of both \code{x} and \code{pattern} are ignored in search. If \code{TRUE}, case used in \code{x} and \code{pattern} must match}

\item{fixed}{If \code{TRUE}, exact search is performed for \code{pattern} in \code{x}}

\item{perl}{If \code{TRUE}, perl-style regular expressions can be entered as \code{pattern}}

\item{use_bytes}{If \code{TRUE}, search is performed byte-by-byte. If \code{FALSE} (default), search is performed character-by-character}

\item{fuzzy}{If \code{TRUE}, fuzzy matching is performed}

\item{value}{If \code{TRUE}, value of \code{x}}
}
\value{
Output is a vector with length equal to input and class determined by \code{output} (same class as input if \code{output = "value"}, logical if \code{output = "logical"}, integer vector of indices in \code{x} if \code{output = "index"}) of output determined by \code{output} argument. Length identical to input
}
\description{
Wrapper for grepl, grep, agrep
}
