% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_nth_number}
\alias{str_nth_number}
\title{Extract nth number from string}
\usage{
str_nth_number(
  x,
  n = 1,
  incl_decimal = TRUE,
  incl_comma = FALSE,
  incl_neg = FALSE,
  as_numeric = FALSE
)
}
\arguments{
\item{x}{String}

\item{n}{nth number in \code{x}. Can be numeric, \code{"first"} (\code{n = 1}), or \code{"last"} (\code{n = -1}). For second to last occurrence enter \code{n = -2}}

\item{incl_decimal}{If \code{TRUE} (default), decimals can be included in numbers. If \code{FALSE}, values will be split at decimals. Only relevant if \code{single_number} is \code{FALSE}}

\item{incl_comma}{If \code{FALSE} (default), values will be split at commas. If \code{TRUE}, commas can be included in numbers. Only relevant if \code{single_number} is \code{FALSE}}

\item{incl_neg}{If \code{FALSE} (default), negative sign not included in output. Only relevant if single_number is \code{FALSE}}

\item{as_numeric}{If \code{FALSE} (default), output is a character. If \code{TRUE}, output is numeric}
}
\description{
Extract nth number from string
}
