% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_numerics}
\alias{str_numerics}
\title{Extract all numbers from string}
\usage{
str_numerics(
  x,
  single_number = FALSE,
  incl_decimal = TRUE,
  incl_comma = FALSE,
  incl_neg = FALSE,
  as_numeric = FALSE
)
}
\arguments{
\item{x}{String}

\item{single_number}{If \code{FALSE} (default), output is a list of numbers. If \code{TRUE}, output is concatenated numbers that appear throughout string}

\item{incl_decimal}{If \code{TRUE} (default), decimals can be included in numbers. If \code{FALSE}, values will be split at decimals. Only relevant if \code{single_number} is \code{FALSE}}

\item{incl_comma}{If \code{FALSE} (default), values will be split at commas. If \code{TRUE}, commas can be included in numbers. Only relevant if \code{single_number} is \code{FALSE}}

\item{incl_neg}{If \code{FALSE} (default), negative sign not included in output. Only relevant if single_number is \code{FALSE}}

\item{as_numeric}{If \code{FALSE} (default), output is a character. If \code{TRUE}, output is numeric}
}
\value{
1 or more numbers. If \code{single_number} is \code{FALSE}, output is a list of numbers (as a character if \code{as_numeric} is \code{FALSE})
}
\description{
Extract all numbers from string
}
