% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_pattern}
\alias{str_pattern}
\title{Identify general pattern in strings}
\usage{
str_pattern(x, shorten_letters = TRUE, shorten_numbers = TRUE, summary = TRUE)
}
\arguments{
\item{x}{Character vector or data frame}

\item{shorten_letters, shorten_numbers}{If \code{TRUE} (default), sequences of letter cases or numbers are shortened to a single symbol}

\item{summary}{If \code{TRUE} (default), number of occurrences of each pattern in \code{x} is included in output}
}
\value{
Patterns identified in \code{x}. "A" represents upper case letters, "a" represents lower case letters, "#" represents numeric digit 0-9
}
\description{
Functionality from Brandon Greenwell's excellent package bpa
}
