% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_position}
\alias{str_position}
\title{Determine the position of a pattern within a string}
\usage{
str_position(x, pattern, fixed = FALSE, ignore_case = FALSE, ...)
}
\arguments{
\item{x}{String}

\item{pattern}{Pattern to identify in string}

\item{fixed}{If \code{FALSE} (default), pattern is not used as is}

\item{ignore_case}{If \code{FALSE} (default), case of both \code{x} and \code{pattern} are ignored in search. If \code{TRUE}, case used in \code{x} and \code{pattern} must match}

\item{...}{Arguments passed to \code{gregexpr}}
}
\value{
List of matrices with columns for start and end position of pattern
}
\description{
Functionality from Hadley Wickham's excellent package stringb
}
