% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_remove}
\alias{str_remove}
\alias{str_remove_numerics}
\alias{str_alphanum}
\alias{str_remove_punc}
\alias{str_remove_space}
\title{Remove a pattern from a string}
\usage{
str_remove(x, pattern, all = TRUE, ...)

str_remove_numerics(x, all = TRUE)

str_alphanum(x)

str_remove_punc(x)

str_remove_space(x)
}
\arguments{
\item{x}{String}

\item{pattern}{Pattern to remove from string. Enter as quoted regular expression or exact text}

\item{all}{If \code{TRUE} (default), all matches for \code{pattern} will be removed. If \code{FALSE}, only the first match for \code{pattern} will be removed}

\item{...}{Arguments passed to \code{sub} or \code{gsub}}
}
\description{
Remove a pattern from a string

Remove numbers from a string

Remove all non-number/non-text characters from string

Alias for str_alphanum

Remove all spaces/whitespace (not just ends of string)
}
