% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{str_remove_ext}
\alias{str_remove_ext}
\alias{remove_ext}
\title{Remove file extension from string}
\usage{
str_remove_ext(x)

remove_ext(x)
}
\arguments{
\item{x}{File name or file path}
}
\value{
File name or file path without extension
}
\description{
Remove file extension from string

Alias for str_remove_ext
}
