% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_remove_letters}
\alias{str_remove_letters}
\title{Remove letters from string}
\usage{
str_remove_letters(x, case = "both", all = TRUE)
}
\arguments{
\item{x}{String}

\item{case}{Case of letters to be removed. Options: \code{"both"} (default), \code{"lower"}, \code{"upper"}}

\item{all}{If \code{TRUE} (default), all matches for \code{pattern} will be removed. If \code{FALSE}, only the first match for \code{pattern} will be removed}
}
\description{
Remove letters from string
}
