% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_replace_between_parens}
\alias{str_replace_between_parens}
\title{Replace text inside parentheses}
\usage{
str_replace_between_parens(x, replacement = "", sub_fn = sub, ...)
}
\arguments{
\item{x}{String}

\item{replacement}{Text to be inserted between parenthesis}

\item{sub_fn}{Function used to substitute strings. Default is \code{sub}}

\item{...}{Arguments passed to \code{sub_fn}}
}
\description{
Replace text inside parentheses
}
