% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_replace_numerics}
\alias{str_replace_numerics}
\title{Replace or remove numbers in string}
\usage{
str_replace_numerics(x, replacement = "")
}
\arguments{
\item{x}{String}

\item{replacement}{Replacement for numbers in \code{x}. Enter as numeric or string. Default removes all numbers from \code{x}}
}
\value{
New string containing replacement wherever numbers occur in \code{x}
}
\description{
Replace or remove numbers in string
}
