% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_snake_case}
\alias{str_snake_case}
\alias{str_camel_case}
\alias{str_split_camel}
\title{Convert from CamelCase to snake_case}
\usage{
str_snake_case(x, lower = TRUE)

str_camel_case(x)

str_split_camel(x)
}
\arguments{
\item{x}{Character vector}

\item{lower}{If \code{TRUE} (default), output is converted to lower case}
}
\value{
Character vector with length same as input
}
\description{
Convert from CamelCase to snake_case

Convert snake_case to CamelCase

Split CamelCase to Camel Case
}
