% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_wrap}
\alias{str_wrap}
\title{Wrap strings with prefix and/or suffix}
\usage{
str_wrap(x, prefix = "", suffix = "")
}
\arguments{
\item{x}{Character vector}

\item{prefix, suffix}{String to attach to beginning (\code{prefix}) or end (\code{suffix}) of each value in \code{x}. Each must be length 1}
}
\value{
Character vector with length same as input
}
\description{
Wrap strings with prefix and/or suffix
}
