% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{subscript}
\alias{subscript}
\title{Subscript notation}
\usage{
subscript(
  base,
  subscript,
  font_face_base = c("plain", "italic", "bold", "bolditalic"),
  font_face_subscript = c("plain", "italic", "bold", "bolditalic")
)
}
\arguments{
\item{base, subscript}{Text for base and subscript respectively. Enter as quoted or unquoted text}

\item{font_face_base, font_face_subscript}{Font face for base and subscript text respectively. Options: \code{"plain"} (default), \code{"italic"}, \code{"bold"}, \code{"bolditalic"}. Enter as quoted type}
}
\value{
Expression that can be entered as ggplot2 title argument
}
\description{
Subscript notation
}
