% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{superscript}
\alias{superscript}
\title{Superscript notation}
\usage{
superscript(
  base,
  superscript,
  font_face_base = c("plain", "italic", "bold", "bolditalic"),
  font_face_superscript = c("plain", "italic", "bold", "bolditalic")
)
}
\arguments{
\item{base, superscript}{Text for base and superscript respectively. Enter as quoted or unquoted text}

\item{font_face_base, font_face_superscript}{Font face for base and superscript text respectively. Options: \code{"plain"} (default), \code{"italic"}, \code{"bold"}, \code{"bolditalic"}. Enter as quoted type}
}
\value{
Expression that can be entered as ggplot2 title argument
}
\description{
Superscript notation
}
