% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_glance}
\alias{surv_glance}
\title{Summarize coxph model (similar to broom::glance output)}
\usage{
surv_glance(x)
}
\arguments{
\item{x}{coxph object}
}
\value{
Data frame with 1 row and columns for outcome_var, time_var, covariates, n, n_events, p_wald_model (P value for Wald test for global model), p_log_rank_coxph_model (P value for log-rank test for global model), p_log_likelihood_model (P value for log likelihood ratio test for global model), concordance, concordance_se, r_sq, r_sq_max, aic, bic, univariate
}
\description{
Summarize coxph model (similar to broom::glance output)
}
