% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_pairwise}
\alias{surv_pairwise}
\title{Pairwise HR, CI, P values}
\usage{
surv_pairwise(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  adjust_method = "holm"
)
}
\arguments{
\item{df}{Data frame}

\item{predictor_var}{Variable to be compared pairwise. Must be categorical with > 1 level. Enter as character vector}

\item{time_var}{Variable containing follow up time. Enter as quoted variable name. Default is \code{time}}

\item{outcome_var}{Variable containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as quoted variable name. Default is \code{death}}

\item{adjust_method}{Method of adjustment for multiple comparisons}
}
\value{
Data frame containing columns for Group1, Group2, hr, hr_lower, hr_upper, p_lrt, p_lrt_adj, label, p_wald, p_wald_adj, n, n_events
}
\description{
Pairwise HR, CI, P values
}
