% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survdiff_summary}
\alias{survdiff_summary}
\title{Extract sample sizes/number of events and p value from survdiff object}
\usage{
survdiff_summary(x)
}
\arguments{
\item{x}{survdiff object}
}
\value{
List containing p_lrt and a data frame with information about grouping variables, sample size (n), number of events (n_events)
}
\description{
Extract sample sizes/number of events and p value from survdiff object
}
