% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{swap_names_vals}
\alias{swap_names_vals}
\title{Swap values with names or vice versa}
\usage{
swap_names_vals(x)
}
\arguments{
\item{x}{Named vector}
}
\value{
Vector with names of \code{x} now forming values of output vector and values of \code{x} now forming names of output vector
}
\description{
Swap values with names or vice versa
}
