% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{swatch}
\alias{swatch}
\alias{pal}
\title{Display colors as swatch}
\usage{
swatch(
  ...,
  direction = c("vertical", "horizontal"),
  border_thickness = 2.5,
  reverse = FALSE
)

pal(..., hex = TRUE, ncol = NULL)
}
\arguments{
\item{...}{List, character vector, or comma separated list of colors as quoted color names or hexadecimal codes}

\item{direction}{Orientation of color squares. Options: \code{"vertical"} (default) or \code{"horizontal"}. Enter as quoted direction}

\item{border_thickness}{Square border thickness}

\item{reverse}{Reverse order of colors}

\item{hex}{If \code{TRUE} (default), hexadecimal codes of input colors displayed}

\item{ncol}{Number of columns in output}
}
\description{
Inspired by \code{scales::show_col}
}
