% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_2_by_2.R
\name{tab_2_by_2.character}
\alias{tab_2_by_2.character}
\alias{tab_2_by_2.factor}
\title{tab_2_by_2 - character}
\usage{
\method{tab_2_by_2}{character}(x, ...)

\method{tab_2_by_2}{factor}(x, ...)
}
\arguments{
\item{x}{Character vector}

\item{...}{Arguments passed to table}
}
\description{
tab_2_by_2 - character

tab_2_by_2 - factor vector
}
