% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_2_by_2.R
\name{tab_2_by_2.data.frame}
\alias{tab_2_by_2.data.frame}
\title{tab_2_by_2 - data.frame}
\usage{
\method{tab_2_by_2}{data.frame}(x, ...)
}
\arguments{
\item{x}{Data frame of counts}

\item{...}{Columns to use in \code{x}. Enter as formula (y ~ x or ~ y + x. outcome ~ predictor), comma separated list of quoted or unquoted column names not wrapped in \code{c()}, or tidyselect syntax}
}
\description{
tab_2_by_2 - data.frame
}
