% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_2_by_2.R
\name{tab_2_by_2.logical}
\alias{tab_2_by_2.logical}
\title{tab_2_by_2 - logical}
\usage{
\method{tab_2_by_2}{logical}(x, ...)
}
\arguments{
\item{x}{Logical vector}

\item{...}{Arguments passed to table}
}
\description{
tab_2_by_2 - logical
}
