% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_2_by_2.R
\name{tab_2_by_2.matrix}
\alias{tab_2_by_2.matrix}
\title{tab_2_by_2 - matrix}
\usage{
\method{tab_2_by_2}{matrix}(x, ...)
}
\arguments{
\item{x}{2 x 2 matrix}

\item{...}{Not used}
}
\description{
tab_2_by_2 - matrix
}
