% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_2_by_2.R
\name{tab_2_by_2.numeric}
\alias{tab_2_by_2.numeric}
\title{tab_2_by_2 - numeric}
\usage{
\method{tab_2_by_2}{numeric}(x, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{...}{Counts or arguments passed to table}
}
\description{
tab_2_by_2 - numeric
}
