% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_2_by_2.R
\name{tab_2_by_2}
\alias{tab_2_by_2}
\alias{tab_2_by_2.default}
\title{Create 2 x 2 table for a broad range of input types}
\usage{
tab_2_by_2(x, ...)

\method{tab_2_by_2}{default}(x, ...)
}
\arguments{
\item{x}{Data frame of counts, matrix, table, or length 4 numeric vector of counts. Enter by column (i.e., row 1/col 1, then row 2/col 1, then row 1/col 2, then row 2/col 2)}

\item{...}{Arguments passed to class-specific \code{tab_2_by_2} function}
}
\value{
2 x 2 table
}
\description{
Create 2 x 2 table for a broad range of input types

tab_2_by_2 - default method
}
