% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_2_by_2.R
\name{tab_2_by_2.table}
\alias{tab_2_by_2.table}
\title{tab_2_by_2 - table}
\usage{
\method{tab_2_by_2}{table}(x, ...)
}
\arguments{
\item{x}{2 x 2 table}

\item{...}{Not used}
}
\description{
tab_2_by_2 - table
}
