% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_to_counts}
\alias{tab_to_counts}
\alias{counts}
\title{Create data frame with counts column}
\usage{
tab_to_counts(x, ..., colname_counts = "Freq")

counts(x, ..., colname_counts = "Freq")
}
\arguments{
\item{x}{Table or matrix with dimnames or data frame}

\item{...}{Columns used to generate table. Only relevant if \code{x} is a data frame}

\item{colname_counts}{Column name for counts. Enter as quoted column name. Default is \code{"Freq"}}
}
\value{
Data frame containing column with counts
}
\description{
Output similar to \code{dplyr::count()} output when applied to df but with different row order
}
