% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tab_to_data}
\alias{tab_to_data}
\title{Convert contingency table into raw data}
\usage{
tab_to_data(x, colname_counts = "Freq")
}
\arguments{
\item{x}{Contingency table or matrix containing dimnames or a data frame with 1 column for counts and other columns containing grouping variables (1 row per group). Data frame input is should be formatted as the output of \code{dplyr::count}}

\item{colname_counts}{Column in \code{x} containing counts. Only relevant when \code{x} is a data frame containing counts. Default is \code{"Freq"}}
}
\value{
Raw data with 1 row per individual and 1 column for each variable
}
\description{
Convert contingency table into raw data
}
