% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{theme_blank}
\alias{theme_blank}
\title{Blank/empty theme}
\usage{
theme_blank(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::theme}}
}
\description{
No grid lines, axis lines/ticks, axis labels/titles, background color, border, or legend
}
