% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{theme_custom}
\alias{theme_custom}
\alias{theme_clean}
\title{Custom ggplot2 theme}
\usage{
theme_custom(
  base_size = 16,
  font_family = "",
  aspect_ratio = 1,
  ratio = aspect_ratio,
  axis_tick_length = base_size/2.5,
  axis_line_thickness = 0.7,
  x_axis_line_thickness = axis_line_thickness,
  y_axis_line_thickness = axis_line_thickness,
  x_axis_tick_length = axis_tick_length,
  y_axis_tick_length = axis_tick_length,
  plot_margin_top = (20/16) * base_size,
  plot_margin_bottom = base_size/2,
  plot_margin_left = base_size/2,
  plot_margin_right = base_size/2,
  plot_title_font_face = "plain",
  plot_title_font_color = "black",
  plot_title_font_size = base_size + 2,
  x_axis_title_font_size = base_size + 2,
  x_axis_label_font_size = base_size,
  y_axis_title_font_size = base_size + 2,
  y_axis_label_font_size = base_size,
  x_axis_title_margin_top = base_size/4,
  x_axis_label_margin_top = 0.3 * base_size,
  x_axis_label_margin_right = 0,
  y_axis_title_angle = 90,
  y_axis_title_margin_right = base_size/2,
  y_axis_label_margin_right = 0.3 * base_size,
  x_axis_label_font_face = "plain",
  x_axis_label_angle = 0,
  x_angle = x_axis_label_angle,
  x_axis_label_vjust = 0.5,
  x_axis_label_hjust = 0.5,
  y_axis_label_font_face = "plain",
  y_axis_label_angle = 0,
  y_angle = y_axis_label_angle,
  legend_position = "right",
  legend_title_font_face = "plain",
  legend_title_font_color = "black",
  legend_title_font_size = base_size,
  legend_label_font_face = "plain",
  legend_label_font_color = "black",
  legend_label_font_size = base_size,
  legend_direction = "vertical",
  legend_margin = base_size/2,
  legend_symbol_size = base_size * 0.75,
  legend_symbol_border_thickness = 1,
  legend_spacing = 1,
  facet_font_face = "plain",
  facet_font_color = "white",
  facet_fill_color = "#4D4D4D",
  facet_spacing_x = base_size/2,
  facet_spacing_y = base_size/2
)

theme_clean(base_size = 14, font_family = "")
}
\arguments{
\item{base_size}{Size of text in pts. Default is \code{16}}

\item{font_family}{Font used in plots. Default uses helvetica}

\item{aspect_ratio}{Aspect ratio, y/x. Default is \code{1}}

\item{ratio}{Alias for \code{aspect_ratio}}

\item{axis_tick_length}{Axis tick length in pts. Default is \code{base_size/2.5}}

\item{axis_line_thickness}{Axis and tick thickness in mm. Default is \code{0.7}}

\item{x_axis_line_thickness, y_axis_line_thickness}{Default is \code{0.7}}

\item{x_axis_tick_length, y_axis_tick_length}{Length of axis ticks in pts. Default is \code{base_size/2.5}}

\item{plot_margin_top}{Extra space added above plot in pts. Default is \code{base_size/2}}

\item{plot_margin_bottom}{Extra space added below plot in pts. Default is \code{base_size/2}}

\item{plot_margin_left}{Extra space added to left of plot in pts. Default is \code{base_size/2}}

\item{plot_margin_right}{Extra space added to right of plot in pts. Default is \code{base_size/2}}

\item{plot_title_font_face}{Title font face. Options: \code{"plain"} (default), \code{"italic"}, \code{"bold"}, \code{"bold.italic"}}

\item{plot_title_font_color}{Title font color. Default is \code{"black"}}

\item{plot_title_font_size}{Title font size in pts. Default is \code{base_size + 2}}

\item{x_axis_title_font_size, y_axis_title_font_size}{Default is \code{base_size + 2}}

\item{x_axis_label_font_size, y_axis_label_font_size}{Axis tick label font size in pts. Default is \code{18}}

\item{x_axis_title_margin_top}{Distance from x axis to x axis title. Units in pts. Default is \code{5}}

\item{x_axis_label_margin_top}{Margin between x axis tick labels to tick mark in pts. Default is \code{0.3*base_size}}

\item{x_axis_label_margin_right}{Margin to the right of x axis tick labels. Default is \code{0}}

\item{y_axis_title_angle}{Angle of text used for y axis title. Default is 90}

\item{y_axis_title_margin_right}{Distance from y axis to x axis title in pts. Default is \code{0.5*base_size}}

\item{y_axis_label_margin_right}{Margin between y axis tick labels to tick mark. Units in pts. Default is \code{0.3*base_size}}

\item{x_axis_label_font_face, y_axis_label_font_face}{Options: \code{"plain"} (default), \code{"italic"}, \code{"bold"}, \code{"bold.italic"}}

\item{x_axis_label_angle, y_axis_label_angle}{Angle of x and y axis tick labels. Default is \code{0}}

\item{x_angle, y_angle}{Alias for \code{x_axis_label_angle} and \code{y_axis_label_angle}, respectively}

\item{x_axis_label_hjust, x_axis_label_vjust}{Horizontal and vertical justification of x axis text. Default is \code{0.5}}

\item{legend_position}{Legend position. Default is \code{"right"}}

\item{legend_title_font_face}{Options: \code{"plain"} (default), \code{"italic"}, \code{"bold"}, \code{"bold.italic"}}

\item{legend_title_font_color}{Font color of legend title. Default is \code{"black"}}

\item{legend_title_font_size}{Units in pts. Default is \code{base_size}}

\item{legend_label_font_face}{Options: \code{"plain"} (default), \code{"italic"}, \code{"bold"}, \code{"bold.italic"}}

\item{legend_label_font_color}{Font color of legend labels Default is \code{"black"}}

\item{legend_label_font_size}{Units in pts. Default is \code{base_size}}

\item{legend_direction}{Legend direction. Options: \code{"vertical"} (default), \code{"horizontal"}}

\item{legend_margin}{Space between text labels in legend. Units in pts. Default is \code{base_size/2}}

\item{legend_symbol_size}{Size of symbols in legend. Units in pts. Default is \code{base_size*0.75}}

\item{legend_symbol_border_thickness}{Thickness of symbols in legend. Units in pt. Default is \code{1}}

\item{legend_spacing}{Spacing between legend labels. Units in pts. Default is \code{1}}

\item{facet_font_face}{Options: \code{"plain"} (default), \code{"italic"}, \code{"bold"}, \code{"bold.italic"}}

\item{facet_font_color}{Default is \code{"white"}}

\item{facet_fill_color}{Facet box fill color. Default is \code{"#4D4D4D"}}

\item{facet_spacing_x, facet_spacing_y}{Horizontal and vertical spacing between facets. Units in pts. Default is \code{base_size/2}}
}
\description{
Functionality from Claus Wilke's excellent package cowplot
}
