% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse_exports.R
\name{tidyeval}
\alias{tidyeval}
\title{rlang}
\description{
sym: converts string -> symbol
syms: converts character vector -> list of symbols
}
\details{
enquo: returns an unevaluated expression that can be run later
enquos: returns a list of unevaluated expressions that can be run later

expr: run expression that was previously wrapped in !!enquo
exprs: run expression that was previously wrapped in !!!enquos

as_name: converts variable name -> string
as_label: same as as_name but input can be quoted function call or vector
}
