% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{time_interval}
\alias{time_interval}
\title{Time interval between 2 dates}
\usage{
time_interval(
  t0,
  t1,
  units = "days",
  date_fn = function(x) as.Date(x, tryFormats = c("\%Y-\%m-\%d", "\%Y/\%m/\%d"))
)
}
\arguments{
\item{t0, t1}{Dates for beginning and end of time period, respectively. Can be character or date vector}

\item{units}{Units for time interval. Default is \code{"days"}}

\item{date_fn}{Function used to convert \code{t0} and \code{t1} to dates. Default is \code{as.Date.} Other options: \code{lubridate::ymd}}
}
\value{
Numeric vector with units of time determined by \code{units}
}
\description{
Time interval between 2 dates
}
