% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\docType{data}
\name{trans_log10_1}
\alias{trans_log10_1}
\title{log10 + 1 axis transformation}
\format{
An object of class \code{trans} of length 7.
}
\usage{
trans_log10_1
}
\description{
Function for \code{coord_trans()} to prevent log(0).  ggplot(...) + coord_trans(y = trans_log10_1)
}
\keyword{datasets}
