% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\docType{data}
\name{trans_log10_nonzero}
\alias{trans_log10_nonzero}
\title{log10 axis transformation that allows for 0s}
\format{
An object of class \code{trans} of length 7.
}
\usage{
trans_log10_nonzero
}
\description{
Function for \code{coord_trans()} to prevent log(0). ggplot(...) + coord_trans(y = trans_log10_nonzero)
}
\keyword{datasets}
