% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{transform_asinh}
\alias{transform_asinh}
\alias{transform_logit}
\title{Hyperbolic arc-sine transformation}
\usage{
transform_asinh(x, inverse = FALSE, constant = 5)

transform_logit(x, inverse = FALSE)
}
\arguments{
\item{x}{Numeric vector not containing any 0s}

\item{inverse}{If \code{TRUE}, inverse transformation is performed}

\item{constant}{Constant factor for transformation. Default is \code{5}}
}
\description{
Hyperbolic arc-sine transformation

Logit transformation
}
