% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{transform_boxcox}
\alias{transform_boxcox}
\alias{transform_yj}
\title{Box–Cox transformation}
\usage{
transform_boxcox(x, inverse = FALSE, lambda = NULL, tol = 0.001)

transform_yj(x, inverse = FALSE, lambda = NULL, tol = 0.001)
}
\arguments{
\item{x}{Numeric vector not containing any 0s}

\item{inverse}{If \code{TRUE}, inverse transformation is performed}

\item{lambda}{Lambda value}

\item{tol}{Numeric value close to zero used to determine whether lambda > 0}
}
\value{
Numeric vector with attribute "lambda"
}
\description{
Functionality from bestNormalize package

Functionality from bestNormalize package
}
