% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{transform_by_row}
\alias{transform_by_row}
\title{Transform subset of columns by row}
\usage{
transform_by_row(df, fn, ...)
}
\arguments{
\item{df}{Data frame}

\item{fn}{Function used to transform values}

\item{...}{Columns to transform. Enter using tidyselect syntax}
}
\value{
Data frame with row-wise transformed values in columns specified by \code{...}
}
\description{
Transform subset of columns by row
}
