% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{transpose_df}
\alias{transpose_df}
\title{Transpose a data frame}
\usage{
transpose_df(df, col_names = "variable")
}
\arguments{
\item{df}{Data frame}

\item{col_names}{Character vector of column names in transposed data frame}
}
\value{
Transposed data frame. If any input variable is a character, all output columns will be characters
}
\description{
Transpose a data frame
}
