% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{Try}
\alias{Try}
\alias{tryNULL}
\title{Try to perform evaluation silently}
\usage{
Try(x, otherwise = NULL)

tryNULL(x)
}
\arguments{
\item{x}{Expression}

\item{otherwise}{Return value if \code{x} can't be evaluated. Default is \code{NULL}}
}
\value{
Value of \code{x} or otherwise
}
\description{
Try to perform evaluation silently

Alias for \code{Try(x, otherwise = NULL)}
}
