% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{try_map}
\alias{try_map}
\title{Silent version of lapply/map}
\usage{
try_map(.x, .f, ..., .otherwise = NULL, .remove_null = FALSE)
}
\arguments{
\item{.x}{Input to mapping function}

\item{.f}{Function to apply to each element of \code{.x}}

\item{...}{Arguments passed to \code{.f}}

\item{.otherwise}{Value to be returned if output of \code{.f} is an error. Default is \code{NULL}}

\item{.remove_null}{If \code{TRUE}, \code{NULL} values are removed from output}
}
\description{
Silent version of lapply/map
}
