% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{try_map2_dbl}
\alias{try_map2_dbl}
\title{Silent version of map2_dbl}
\usage{
try_map2_dbl(.x, .y, .f, ..., .otherwise = NA_real_)
}
\arguments{
\item{.x, .y}{Input to mapping function}

\item{.f}{Function to apply to each element of \code{.x}}

\item{...}{Arguments passed to \code{.f}}

\item{.otherwise}{Value to be returned if output of \code{.f} is an error. Default is \code{NULL}}
}
\description{
Silent version of map2_dbl
}
