% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{try_map2_dfr}
\alias{try_map2_dfr}
\title{Silent version of map2_dfr}
\usage{
try_map2_dfr(.x, .y, .f, ..., .otherwise = NULL, .prefix = "V")
}
\arguments{
\item{.x, .y}{Input to mapping function}

\item{.f}{Function to apply to each element of \code{.x}}

\item{...}{Arguments passed to \code{.f}}

\item{.otherwise}{Value to be returned if output of \code{.f} is an error. Default is \code{NULL}}

\item{.prefix}{Prefix for column names. Enter as quoted prefix. Default is \code{"V"}}
}
\description{
Silent version of map2_dfr
}
