% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{try_map_dbl}
\alias{try_map_dbl}
\title{Silent version of vapply for numeric output (or map_dbl)}
\usage{
try_map_dbl(.x, .f, ..., .otherwise = NA_real_, .use_names = FALSE)
}
\arguments{
\item{.x}{Input to mapping function}

\item{.f}{Function to apply to each element of \code{.x}}

\item{...}{Arguments passed to \code{.f}}

\item{.otherwise}{Value to be returned if output of \code{.f} is an error. Default is \code{NULL}}

\item{.use_names}{If \code{TRUE}, names of \code{.x} are applied to names of output}
}
\description{
Silent version of vapply for numeric output (or map_dbl)
}
