% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{try_walk}
\alias{try_walk}
\title{Silent version of walk}
\usage{
try_walk(.x, .f, ..., .otherwise = NULL)
}
\arguments{
\item{.x}{Input to mapping function}

\item{.f}{Function to apply to each element of \code{.x}}

\item{...}{Arguments passed to \code{.f}}

\item{.otherwise}{Value to be returned if output of \code{.f} is an error. Default is \code{NULL}}
}
\description{
Silent version of walk
}
