% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{Unlist}
\alias{Unlist}
\title{Unlist without keeping names}
\usage{
Unlist(x, recursive = TRUE)
}
\arguments{
\item{x}{List, data frame}

\item{recursive}{If \code{TRUE} (default), sublists are unlisted}
}
\value{
Unlisted and unnamed contents of x
}
\description{
Unlist without keeping names
}
