% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{update_list}
\alias{update_list}
\title{Update list}
\usage{
update_list(old, new)
}
\arguments{
\item{old, new}{Old and new lists respectively}
}
\value{
Updated list in which named elements shared by both lists will be updated by replacing the named element in \code{old} with the matching named element in \code{new}. If element named is present in both lists and is \code{NULL} in \code{new}, it will be removed from updated list. Named elements in \code{new} that do not have a matching named element in \code{old} will be added to \code{new}
}
\description{
Functionality from \code{modifyList} function in utils package
}
