% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{update_values}
\alias{update_values}
\alias{update_values.default}
\alias{update_values.data.frame}
\title{Generic update_values function}
\usage{
update_values(x, lookup, old = NULL, new = NULL, ...)

\method{update_values}{default}(
  x,
  lookup,
  old = NULL,
  new = NULL,
  col = NULL,
  new_col_name = NULL,
  ...
)

\method{update_values}{data.frame}(
  x,
  lookup,
  old = NULL,
  new = NULL,
  col = NULL,
  new_col_name = NULL,
  ...
)
}
\arguments{
\item{x}{Data frame or vector to transform}

\item{lookup}{Lookup table or named vector. If \code{lookup} is a vector, names of \code{lookup} correspond to values of \code{x} (i.e. old values) and values of \code{lookup} represent replacement values (i.e. new values). For example, to replace lower case values with upper case values, \code{lookup} is c(a = "A", b = "B", etc.)}

\item{old, new}{Column names of \code{lookup} containing values that match \code{x} and replacement values, respectively. Only relevant if \code{lookup} is a data frame. Enter as quoted or unquoted column names}

\item{...}{Not used}

\item{col}{Var in \code{x} that matches names of \code{lookup} vector. Enter as quoted or unquoted column name}

\item{new_col_name}{Name of new variable to add to \code{x} If \code{NULL} (default), values in \code{x$col} are replaced. If column name is specified, a new column will be added to \code{x} using this name. Enter as quoted or unquoted column name}
}
\value{
Updated version of data frame or vector
}
\description{
Values in x are only replaced if included in names of lookup table
}
