% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{var_class}
\alias{var_class}
\title{Class of each variable}
\usage{
var_class(df, col = NULL)
}
\arguments{
\item{df}{Data frame}

\item{col}{Column(s) to evaluate. If \code{NULL} (default), all columns included. Enter as character or integer vector}
}
\value{
Data frame containing columns for var_name, col_number, class
}
\description{
Class of each variable
}
