% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{vars_all_na}
\alias{vars_all_na}
\title{Get names of variables that contain only missing values}
\usage{
vars_all_na(df, invert = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\description{
Get names of variables that contain only missing values
}
