% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{vars_binary_01}
\alias{vars_binary_01}
\title{Get names of binary variables coded as 0 or 1}
\usage{
vars_binary_01(df, invert = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\description{
Get names of binary variables coded as 0 or 1
}
