% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{vars_can_be_numeric}
\alias{vars_can_be_numeric}
\title{Get names of variables which are coercible to a numeric}
\usage{
vars_can_be_numeric(df, invert = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\description{
Get names of variables which are coercible to a numeric
}
