% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{vars_constant}
\alias{vars_constant}
\title{Get names of constant (zero-variance) variables}
\usage{
vars_constant(df, invert = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\description{
Get names of constant (zero-variance) variables
}
