% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{vars_dbl}
\alias{vars_dbl}
\alias{vars_numeric}
\alias{vars_continuous}
\title{Get names of numeric variables}
\usage{
vars_dbl(df, incl_integer = TRUE, invert = FALSE)

vars_numeric(df, incl_integer = TRUE, invert = FALSE)

vars_continuous(df, invert = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{incl_integer}{If \code{TRUE} (default), integer variables are included}

\item{invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\description{
Get names of numeric variables

Alias for vars_dbl

Get names of continuous variables
}
