% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{vars_id}
\alias{vars_id}
\title{Get names of id variables}
\usage{
vars_id(df, na.rm = FALSE, invert = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{na.rm}{If \code{TRUE}, missing values are not considered}

\item{invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\description{
Get names of id variables
}
