% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{vars_int}
\alias{vars_int}
\title{Get names of integer variables}
\usage{
vars_int(df, invert = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\description{
Get names of integer variables
}
