% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{vars_which}
\alias{vars_which}
\title{Get names of variables based on predicate function applied to column values}
\usage{
vars_which(.df, .predicate_fn, ..., .invert = FALSE)
}
\arguments{
\item{.df}{Data frame}

\item{.predicate_fn}{Predicate function that will be applied to column values to select variable names. Must evaluate to logical (not \code{NULL}) and should not be vectorized (i.e. for vector input, single value should be returned)}

\item{...}{Arguments passed to \code{.predicate_fn}}

\item{.invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}
}
\value{
Character vector containing variable names that result in \code{TRUE} when \code{.predicate_fn} is applied to columns in \code{df}
}
\description{
Get names of variables based on predicate function applied to column values
}
