% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_coerce.R
\name{vec_to_df}
\alias{vec_to_df}
\title{Convert vectors to data.frame}
\usage{
vec_to_df(..., prefix = "V")
}
\arguments{
\item{...}{Vectors or list of vectors. Can have different lengths (lengths do not need to share a common multiple which is in contrast to vector to data frame conversion by data.frame() function)}

\item{prefix}{Prefix for newly created column names. Default is \code{"V"}}
}
\value{
Data frame. Each input vector forms a column in output data frame. If vectors do not share the same length, shorter vectors are recycled to match the length of the longest input vector (see examples)
}
\description{
Convert vectors to data.frame
}
