% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{vlookup}
\alias{vlookup}
\alias{vlookup.default}
\alias{vlookup.factor}
\alias{vlookup.data.frame}
\title{Generic vlookup function}
\usage{
vlookup(x, lookup, old = NULL, new = NULL, silent = FALSE, ...)

\method{vlookup}{default}(x, lookup, old = NULL, new = NULL, silent = FALSE, ...)

\method{vlookup}{factor}(x, lookup, old = NULL, new = NULL, silent = FALSE, ...)

\method{vlookup}{data.frame}(
  x,
  lookup,
  old = NULL,
  new = NULL,
  silent = FALSE,
  col = NULL,
  new_col_name = NULL,
  ...
)
}
\arguments{
\item{x}{Data frame or vector to transform}

\item{lookup}{Mapping between old and new values. Enter 1 of the following
\itemize{
\item Data frame. Must contain columns specified by \code{"old"} and \code{"new"} arguments
\item Named vector. Names are current ("old") values in \code{x}. Values are replacement ("new") values
For example, to replace lower case values with upper case values, \code{lookup} use c(a = "A", b = "B", etc.)
}}

\item{old, new}{Columns in \code{lookup} containing values that match \code{x} ("old") and replacement ("new") values, respectively. Only relevant when \code{lookup} is a data frame. Enter as quoted or unquoted column names}

\item{silent}{If \code{FALSE} (default), messages regarding irregularities in matching are reported}

\item{...}{Not used}

\item{col}{Variable in \code{x} to match to names of lookup vector. Enter as quoted or unquoted column name}

\item{new_col_name}{Name of new variable to add to \code{z}. If \code{NULL} (default), values in \code{x$col} are replaced. If column name is specified, a new column will be added to \code{x} using this name. Enter as quoted or unquoted column name}
}
\value{
Updated version of data frame or vector
}
\description{
Generic vlookup function

vlookup - default method

vlookup - factor

vlookup - data frame
}
