% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_prism.R
\name{write_prism}
\alias{write_prism}
\title{Export prism file}
\usage{
write_prism(df, directory = desktop_path(), row_names = TRUE)
}
\arguments{
\item{df}{Data frame. Can enter multiple data frames using list("name df1" = df1, etc.)}

\item{directory}{Path to directory where file should be saved. Default is \code{desktop_path()}}

\item{row_names}{If \code{TRUE} (default), row names included in 1st column Prism table}
}
\description{
Functionality from Yue Jiang's excellent package pzfx
}
