% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{x_table}
\alias{x_table}
\title{Cross table with or without row/column totals}
\usage{
x_table(
  df,
  ...,
  total = FALSE,
  incl_na = TRUE,
  title_row_totals = "Total",
  title_col_totals = "Total"
)
}
\arguments{
\item{df}{Data frame}

\item{...}{Columns in \code{df} used to generate contingency table. Enter using tidyselect syntax}

\item{total}{If \code{FALSE} (Default), no columns or rows for totals in output table. If \code{TRUE}, row and column totals are included in output table}

\item{incl_na}{If \code{TRUE} (default), missing values are included in output}

\item{title_row_totals}{Title of column containing row totals. Default is \code{"Total"}}

\item{title_col_totals}{Title of row containing column totals. Default is \code{"Total"}}
}
\value{
Cross table +/- rows and columns for totals
}
\description{
Cross table with or without row/column totals
}
